### 2022-04-28 #532 @bcjaeger
aorsf; accelerated oblique random survival forests
Submitting Author Name: Byron C Jaeger
Submitting Author Github Handle: @bcjaeger
Other Package Authors Github handles: (comma separated, delete if none) @nmpieyeskey, @sawyerWeld
Repository:  https://github.com/bcjaeger/aorsf
Version submitted: 0.0.0.9000
Submission type: Stats
Badge grade: gold
Reviewers: @chjackson, @mnwright, @jemus42
Version accepted: TBD
---

```
Package: aorsf
Title: Accelerated Oblique Random Survival Forests
Version: 0.0.0.9000
    person(given = "Byron",
           family = "Jaeger",
           email = "bjaeger@wakehealth.edu",
    person(given = "Nicholas",  family = "Pajewski", role = "ctb"),
    person(given = "Sawyer", family = "Welden", role = "ctb", email = "swelden@wakehealth.edu")
    )
Description: Fit, interpret, and make predictions with oblique random
    survival forests. Oblique decision trees are notoriously slow compared
    to their axis based counterparts, but 'aorsf' runs as fast or faster than
    axis-based decision tree algorithms for right-censored time-to-event
    outcomes.
Roxygen: list(markdown = TRUE, roclets = c ("namespace", "rd", "srr::srr_stats_roclet"))
    Rcpp,
    RcppArmadillo
    table.glue,
    Rcpp,
    data.table
    https://bcjaeger.github.io/aorsf
    survival,
    survivalROC,
    glmnet,
    covr,
    units
```

- [X] A pre-submission inquiry has been approved in [issue#525](#<issue_num>)

Target audience: people who want to fit and interpret a risk prediction model, i.e., a prediction model for right-censored time-to-event outcomes.

Applications: fit an oblique random survival forest, compute predicted risk at a given time, estimate the importance of individual variables, and compute partial dependence to depict relationships between specific predictors and predicted risk.

    - *An improvement on other implementations of similar algorithms in* **R**.

    Please include hyperlinked references to all other relevant software. The [obliqueRSF package](https://github.com/bcjaeger/obliqueRSF) was the original R package for oblique random forests. I wrote it and it is very slow. I wrote `aorsf` because I had multiple ideas about how to make `obliqueRSF` faster, specifically using a partial Newton Raphson algorithm instead of using penalized regression to derive linear combinations of variables in decision nodes. It would have been possible to rewrite `obliqueRSF`, but it would have been difficult to make the re-write backwards compatible with the version of `obliqueRSF` on CRAN.

Gold

1. _Compliance with a good number of standards beyond those identified as minimally necessary._ `aorsf` complies with over 100 combined standards in the general and ML categories.
2. _Demonstrating excellence in compliance with multiple standards from at least two broad sub-categories._ See 1. above
3. _Internal aspects of package structure and design._ `aorsf` uses an optimized routine to partially complete Newton Raphson scoring for the Cox proportional hazards model and also an optimized routine to compute likelihood ratio tests. Both of these routines are heavily used when fitting oblique random survival forests, and both demonstrate the exact same answers as corresponding functions in the `survival` package (see tests in `aorsf`) while running at least twice as fast (thanks to Rcpparmadillo).

I _think_ `aorsf` is passing `autotest` and `srr_stats_pre_submit()`. I am having some issues running these on R 4.2. Currently, autotest is returning NULL, which I understand to be a good thing, and srr_stats_pre_submit is not able to run (not sure why; but it was fine before I updated to R 4.2).
