### 2022-04-28 #532 @bcjaeger
> 🚀
>
> The following problem was found in your submission template:
>
> * 'statsgrade' variable must be one of [bronze, silver, gold]
>   Editors: Please ensure these problems with the submission template are rectified. Package checks have been started regardless.
>
> 👋

Just updated to fix this. I'm not sure if reviewers will think that 'gold' is the right statsgrade, but might as well aim high.
