+++ README.md
@@ -12,6 +12,8 @@ public.](https://www.repostatus.org/badges/latest/wip.svg)](https://www.repostat
 coverage](https://codecov.io/gh/bcjaeger/aorsf/branch/master/graph/badge.svg)](https://app.codecov.io/gh/bcjaeger/aorsf?branch=master)
 [![R-CMD-check](https://github.com/bcjaeger/aorsf/workflows/R-CMD-check/badge.svg)](https://github.com/bcjaeger/aorsf/actions)
 [![pkgcheck](https://github.com/bcjaeger/aorsf/workflows/pkgcheck/badge.svg)](https://github.com/bcjaeger/aorsf/actions?query=workflow%3Apkgcheck)
+[![Status at rOpenSci Software Peer
+Review](https://badges.ropensci.org/532_status.svg)](https://github.com/ropensci/software-review/issues/532)
 <!-- badges: end -->
 
 `aorsf` provides optimized software to fit, interpret, and make
@@ -47,6 +49,7 @@ The `orsf()` function is used to fit ORSFs. Printing the output from
 
 ``` r
 library(aorsf)
+#> Warning: package 'aorsf' was built under R version 4.1.3
 
 fit <- orsf(data_train = pbc_orsf,
             formula = Surv(time, status) ~ . - id)
@@ -74,14 +77,14 @@ How about interpreting the fit?
 
     ``` r
     orsf_vi_negate(fit)
-    #>          bili           age       spiders       ascites        copper 
-    #>  0.0156282559  0.0127109815  0.0065638675  0.0059387372  0.0046363826 
-    #>       protime         edema         stage        hepato          trig 
-    #>  0.0045842884  0.0035386837  0.0022921442  0.0013023547  0.0010418837 
-    #>           sex           ast           trt      alk.phos      platelet 
-    #>  0.0007293186  0.0006251302 -0.0011981663 -0.0018753907 -0.0019795791 
-    #>          chol       albumin 
-    #> -0.0023963326 -0.0047405710
+    #>          bili           age           sex        copper       protime 
+    #>  0.0169827047  0.0141175245  0.0087518233  0.0075536570  0.0065638675 
+    #>       ascites         stage       spiders       albumin        hepato 
+    #>  0.0055740779  0.0049489477  0.0042196291  0.0036465930  0.0032819337 
+    #>         edema           ast          trig          chol      alk.phos 
+    #>  0.0029582056  0.0022921442  0.0002083767 -0.0002604709 -0.0020837675 
+    #>           trt      platelet 
+    #> -0.0021879558 -0.0024484268
     ```
 
 -   use `orsf_pd_ice()` or `orsf_pd_summary()` for individual or
@@ -89,12 +92,13 @@ How about interpreting the fit?
 
     ``` r
     orsf_pd_summary(fit, pd_spec = list(bili = c(1:5)))
-    #>    bili      mean        lwr      medn       upr
-    #> 1:    1 0.2352732 0.01833523 0.1196051 0.8675898
-    #> 2:    2 0.2821136 0.04112874 0.1714480 0.8919186
-    #> 3:    3 0.3371132 0.06911118 0.2545069 0.9177142
-    #> 4:    4 0.3922269 0.10551068 0.3141783 0.9270324
-    #> 5:    5 0.4376877 0.13935792 0.3658866 0.9337528
+    #>     bili      mean        lwr      medn       upr
+    #>    <int>     <num>      <num>     <num>     <num>
+    #> 1:     1 0.2372611 0.01574818 0.1294677 0.8765468
+    #> 2:     2 0.2823997 0.03808830 0.1751359 0.9008787
+    #> 3:     3 0.3342751 0.06253403 0.2385166 0.9125669
[…]
