+++ NEWS.md
@@ -5,3 +5,7 @@
 * Vignettes + Readme covering usage of core features.
 
 * Website hosted through GitHub pages, managed with `pkgdown`.
+
+# aorsf 0.0.1
+
+* To be filled in as ROpenSci review continues
