### 2022-05-10 #532 @bcjaeger
> Please add this badge to the README of your package repository:
>
> ```lisp
> ```
>
> Furthermore, if your package does not have a NEWS.md file yet, please create one to capture the changes made during the review process. See https://devguide.ropensci.org/releasing.html#news

Done! Looking forward to the review.
