+++ joss-paper/paper.bib
@@ -1,0 +1,71 @@
+
+@Manual{survival-package,
+    title = {A Package for Survival Analysis in R},
+    author = {Terry M Therneau},
+    year = {2022},
+    note = {R package version 3.3-1},
+    url = {https://CRAN.R-project.org/package=survival},
+  }
+
+@misc{therneau_survival_2022,
+	title = {Survival Package Source Code Documentation},
+	url = {https://github.com/therneau/survival/blob/5440691d44abea537b08aeb60153a31654d66a9b/noweb},
+	abstract = {Survival package for R},
+	urldate = {2022-05-12},
+	author = {Therneau, Terry},
+	month = apr,
+	year = {2022},
+	note = {original-date: 2016-04-28T18:05:02Z},
+}
+
+@article{breiman2001random,
+  title={Random forests},
+  author={Breiman, Leo},
+  journal={Machine learning},
+  volume={45},
+  number={1},
+  pages={5--32},
+  year={2001},
+  publisher={Springer}
+}
+
+@inproceedings{menze2011oblique,
+  title={On oblique random forests},
+  author={Menze, Bjoern H and Kelm, B Michael and Splitthoff, Daniel N and Koethe, Ullrich and Hamprecht, Fred A},
+  booktitle={Joint European Conference on Machine Learning and Knowledge Discovery in Databases},
+  pages={453--469},
+  year={2011},
+  organization={Springer}
+}
+
+@article{jaeger2019oblique,
+  title={Oblique random survival forests},
+  author={Jaeger, Byron C and Long, D Leann and Long, Dustin M and Sims, Mario and Szychowski, Jeff M and Min, Yuan-I and Mcclure, Leslie A and Howard, George and Simon, Noah},
+  journal={The Annals of Applied Statistics},
+  volume={13},
+  number={3},
+  pages={1847--1883},
+  year={2019},
+  publisher={Institute of Mathematical Statistics}
+}
+
+@article{segar2021development,
+  title={Development and Validation of Machine Learning--Based Race-Specific Models to Predict 10-Year Risk of Heart Failure: A Multicohort Analysis},
+  author={Segar, Matthew W and Jaeger, Byron C and Patel, Kershaw V and Nambi, Vijay and Ndumele, Chiadi E and Correa, Adolfo and Butler, Javed and Chandra, Alvin and Ayers, Colby and Rao, Shreya and others},
+  journal={Circulation},
+  volume={143},
+  number={24},
+  pages={2370--2383},
+  year={2021},
+  publisher={Am Heart Assoc}
+}
+
+@article{katuwal2020heterogeneous,
+  title={Heterogeneous oblique random forest},
+  author={Katuwal, Rakesh and Suganthan, Ponnuthurai Nagaratnam and Zhang, Le},
+  journal={Pattern Recognition},
+  volume={99},
+  pages={107078},
+  year={2020},
+  publisher={Elsevier}
+}
