+++ joss-paper/paper.Rmd
@@ -1,0 +1,95 @@
+---
+# Example from https://joss.readthedocs.io/en/latest/submitting.html
+title: 'aorsf: An R package for supervised learning using the oblique random survival forest'
+tags:
+  - R
+  - machine learning
+  - supervised learning
+  - survival
+  - random forest
+authors:
+  - name: Byron C. Jaeger
+    orcid: 0000-0001-7399-2299
+    affiliation: 1 # (Multiple affiliations must be quoted)
+  - name: Sawyer Welden
+    orcid: 0000-0000-0000-0000
+    affiliation: 1
+  - name: Kristin Lenoir
+    orcid: 0000-0000-0000-0000
+    affiliation: 1
+  - name: Nicholas M Pajewski
+    orcid: 0000-0000-0000-0000
+    affiliation: 1
+affiliations:
+ - name: Wake Forest School of Medicine
+   index: 1
+citation_author: Jaeger et. al.
+date: 11 May 2022
+year: 2022
+bibliography: paper.bib
+output: rticles::joss_article
+csl: apa.csl
+journal: JOSS
+---
+
+# Summary
+
+The random forest (RF) is a supervised learning method that combines predictions from a large set (i.e., an ensemble) of decision trees [@breiman2001random]. To de-correlate decision trees in the ensemble, random subsets of training data are used to grow each tree and a subset of randomly selected predictor variables is considered at each non-terminal node in the trees. In @breiman2001random, both axis based and oblique RFs are described. Axis-based trees split non-terminal nodes using individual predictor variables whereas oblique trees use a linear combination of variables. Although oblique RFs outperform their axis-based counterparts in prediction tasks [@menze2011oblique], they are also more computationally intensive and less easily interpreted. Thus, software packages have focused mostly on axis based RFs, with few packages supporting oblique RFs, and even fewer supporting oblique random survival forests (ORSFs; @jaeger2019oblique). 
+
+``aorsf`` is an R package optimized for fitting, interpreting, and computing predictions with ORSFs. The target audience includes both __practitioners__ aiming to develop an accurate and interpretable risk prediction model (e.g., see @segar2021development) and __researchers__ who want to conduct experiments comparing different techniques for identifying linear combinations of predictor variables in a controlled environment (e.g., see @katuwal2020heterogeneous). Key features of ``aorsf`` include computational efficiency compared to existing software, extensive unit and integration testing that ensure cross-platform consistency and reproducibility, and user-friendly documentation paired with an application programming interface that facilitates proper usage of the core algorithms. Interpretation is facilitated by partial dependence and negation importance, a novel method for variable importance designed for compatibility with oblique decision trees.
+
+# Existing software 
+
+The `obliqueRF` R package supports classification and regression using oblique random forests but does not support survival analysis. The `ranger` and `randomForestSRC` packages support survival analysis but not oblique decision trees. The ``obliqueRSF`` R package is currently the only software available to fit ORSFs. Penalized Cox regression models are used to identify linear combinations of inputs. ``aorsf`` is a re-write of the `obliqueRSF` package, with a focus on computational efficiency and flexibility. For example, ``aorsf`` allows oblique decision trees to be grown using Newton Raphson scoring (the default), penalized Cox regression, or a user-defined function.
+
+# Newton Raphson scoring
+
[…]
