+++ joss-paper/paper.Rmd
@@ -10,7 +10,7 @@ tags:
 authors:
   - name: Byron C. Jaeger
     orcid: 0000-0001-7399-2299
-    affiliation: 1 # (Multiple affiliations must be quoted)
+    affiliation: 1 # (Note: multiple affiliations must be quoted)
   - name: Sawyer Welden
     orcid: 0000-0000-0000-0000
     affiliation: 1
@@ -34,9 +34,9 @@ journal: JOSS
 
 # Summary
 
-The random forest (RF) is a supervised learning method that combines predictions from a large set (i.e., an ensemble) of decision trees [@breiman2001random]. To de-correlate decision trees in the ensemble, random subsets of training data are used to grow each tree and a subset of randomly selected predictor variables is considered at each non-terminal node in the trees. In @breiman2001random, both axis based and oblique RFs are described. Axis-based trees split non-terminal nodes using individual predictor variables whereas oblique trees use a linear combination of variables. Although oblique RFs outperform their axis-based counterparts in prediction tasks [@menze2011oblique], they are also more computationally intensive and less easily interpreted. Thus, software packages have focused mostly on axis based RFs, with few packages supporting oblique RFs, and even fewer supporting oblique random survival forests (ORSFs; @jaeger2019oblique). 
+The random forest (RF) is a supervised learning method that combines predictions from a large set of decision trees [@breiman2001random]. To de-correlate decision trees in the RF, random subsets of training data are used to grow each tree and a subset of randomly selected predictor variables is considered at each non-terminal node in the trees. In @breiman2001random, RFs grown with axis based trees and oblique trees are described. Axis-based trees split non-terminal nodes using individual predictor variables whereas oblique trees use a linear combination of variables. Although oblique RFs outperform their axis-based counterparts in prediction tasks [@menze2011oblique], they are also more computationally intensive and less easily interpreted. Thus, software packages have focused mostly on axis based RFs, with few packages supporting oblique RFs, and even fewer supporting oblique random survival forests (ORSFs; @jaeger2019oblique). 
 
-``aorsf`` is an R package optimized for fitting, interpreting, and computing predictions with ORSFs. The target audience includes both __practitioners__ aiming to develop an accurate and interpretable risk prediction model (e.g., see @segar2021development) and __researchers__ who want to conduct experiments comparing different techniques for identifying linear combinations of predictor variables in a controlled environment (e.g., see @katuwal2020heterogeneous). Key features of ``aorsf`` include computational efficiency compared to existing software, extensive unit and integration testing that ensure cross-platform consistency and reproducibility, and user-friendly documentation paired with an application programming interface that facilitates proper usage of the core algorithms. Interpretation is facilitated by partial dependence and negation importance, a novel method for variable importance designed for compatibility with oblique decision trees.
[…]
