### 2022-05-16 #532 @bcjaeger
Hi @tdhock,

Thank you for acting as the editor of this submission. I have recently added a paper.md file to the repository in hopes that `aorsf` can be eligible for an expedited review at the [Journal of Open Source Software](https://joss.theoj.org/) after the ropensci review.  Can you let me know if any other files are needed or if other actions are needed from me to qualify for an expedited review at JOSS? Thank you!
