+++ README.md
@@ -21,7 +21,7 @@ predictions with oblique random survival forests (ORSFs).
 
 ## Why aorsf?
 
--   over 500 times faster than `obliqueRSF`.
+-   over 400 times faster than `obliqueRSF`.
 
 -   accurate predictions for time-to-event outcomes.
 
@@ -49,7 +49,6 @@ The `orsf()` function is used to fit ORSFs. Printing the output from
 
 ``` r
 library(aorsf)
-#> Warning: package 'aorsf' was built under R version 4.1.3
 
 fit <- orsf(data_train = pbc_orsf,
             formula = Surv(time, status) ~ . - id)
@@ -77,14 +76,14 @@ How about interpreting the fit?
 
     ``` r
     orsf_vi_negate(fit)
-    #>          bili           age           sex        copper       protime 
-    #>  0.0169827047  0.0141175245  0.0087518233  0.0075536570  0.0065638675 
-    #>       ascites         stage       spiders       albumin        hepato 
-    #>  0.0055740779  0.0049489477  0.0042196291  0.0036465930  0.0032819337 
-    #>         edema           ast          trig          chol      alk.phos 
-    #>  0.0029582056  0.0022921442  0.0002083767 -0.0002604709 -0.0020837675 
-    #>           trt      platelet 
-    #> -0.0021879558 -0.0024484268
+    #>          bili           age        copper       ascites       protime 
+    #>  1.390915e-02  1.213795e-02  8.387164e-03  6.511773e-03  6.199208e-03 
+    #>         stage           sex        hepato         edema       albumin 
+    #>  4.428006e-03  4.375912e-03  3.490310e-03  3.170303e-03  2.240050e-03 
+    #>       spiders           trt      alk.phos           ast          trig 
+    #>  1.041884e-03  1.562826e-04  5.209419e-05 -1.198166e-03 -1.250260e-03 
+    #>      platelet          chol 
+    #> -1.406543e-03 -5.782455e-03
     ```
 
 -   use `orsf_pd_ice()` or `orsf_pd_summary()` for individual or
@@ -94,11 +93,11 @@ How about interpreting the fit?
     orsf_pd_summary(fit, pd_spec = list(bili = c(1:5)))
     #>     bili      mean        lwr      medn       upr
     #>    <int>     <num>      <num>     <num>     <num>
-    #> 1:     1 0.2372611 0.01574818 0.1294677 0.8765468
-    #> 2:     2 0.2823997 0.03808830 0.1751359 0.9008787
-    #> 3:     3 0.3342751 0.06253403 0.2385166 0.9125669
-    #> 4:     4 0.3846965 0.08521571 0.3106505 0.9235278
-    #> 5:     5 0.4281614 0.11914911 0.3608596 0.9266794
+    #> 1:     1 0.2327332 0.01715402 0.1278394 0.8629424
+    #> 2:     2 0.2803282 0.03730412 0.1726425 0.8873881
+    #> 3:     3 0.3399960 0.06581874 0.2534968 0.9011461
+    #> 4:     4 0.3936117 0.10559414 0.3220519 0.9190984
+    #> 5:     5 0.4360391 0.13796185 0.3741413 0.9339884
     ```
 
 -   use `orsf_summarize_uni()` to show the top predictor variables in an
@@ -117,44 +116,44 @@ How about interpreting the fit?
     #>         |---------------- risk ----------------|
     #>   Value      Mean    Median     25th %    75th %
     #>  <char>     <num>     <num>      <num>     <num>
-    #>    0.80 0.2332029 0.1229387 0.04171414 0.3596673
[…]
