### 2022-05-16 #532 @noamross
@bcjaeger The `paper.md` file should be sufficient. For JOSS submission, after RO review is complete you should submit to JOSS and link to this review thread in the pre-review JOSS thread.  JOSS editors will review the paper.md and can opt to use RO's software reviews rather than finding additional reviewers.
