+++ joss/paper.bib
@@ -1,3 +1,13 @@
+@article{ishwaran2008random,
+  title={Random survival forests},
+  author={Ishwaran, Hemant and Kogalur, Udaya B and Blackstone, Eugene H and Lauer, Michael S},
+  journal={The annals of applied statistics},
+  volume={2},
+  number={3},
+  pages={841--860},
+  year={2008},
+  publisher={Institute of Mathematical Statistics}
+}
 
 @Manual{survival-package,
     title = {A Package for Survival Analysis in R},
