### 2022-05-31 314c4e4
dont scale x if a custom function is being used
