### 2022-06-02 9bd9098
:tada: new funding added!
