+++ README.md
@@ -76,14 +76,14 @@ How about interpreting the fit?
 
     ``` r
     orsf_vi_negate(fit)
-    #>          bili           age        copper       ascites       protime 
-    #>  1.390915e-02  1.213795e-02  8.387164e-03  6.511773e-03  6.199208e-03 
-    #>         stage           sex        hepato         edema       albumin 
-    #>  4.428006e-03  4.375912e-03  3.490310e-03  3.170303e-03  2.240050e-03 
-    #>       spiders           trt      alk.phos           ast          trig 
-    #>  1.041884e-03  1.562826e-04  5.209419e-05 -1.198166e-03 -1.250260e-03 
-    #>      platelet          chol 
-    #> -1.406543e-03 -5.782455e-03
+    #>          bili           age           sex        copper       ascites 
+    #>  0.0160971036  0.0120858512  0.0084392582  0.0068243384  0.0055740779 
+    #>       spiders         stage         edema        hepato           ast 
+    #>  0.0052094186  0.0043238175  0.0022673374  0.0021358616  0.0012502605 
+    #>      platelet       albumin          chol          trig       protime 
+    #>  0.0005209419  0.0003646593  0.0001562826 -0.0001562826 -0.0010939779 
+    #>      alk.phos           trt 
+    #> -0.0015107314 -0.0040112523
     ```
 
 -   use `orsf_pd_ice()` or `orsf_pd_summary()` for individual or
@@ -93,11 +93,11 @@ How about interpreting the fit?
     orsf_pd_summary(fit, pd_spec = list(bili = c(1:5)))
     #>     bili      mean        lwr      medn       upr
     #>    <int>     <num>      <num>     <num>     <num>
-    #> 1:     1 0.2327332 0.01715402 0.1278394 0.8629424
-    #> 2:     2 0.2803282 0.03730412 0.1726425 0.8873881
-    #> 3:     3 0.3399960 0.06581874 0.2534968 0.9011461
-    #> 4:     4 0.3936117 0.10559414 0.3220519 0.9190984
-    #> 5:     5 0.4360391 0.13796185 0.3741413 0.9339884
+    #> 1:     1 0.2309318 0.01708277 0.1168329 0.8542309
+    #> 2:     2 0.2801009 0.04071749 0.1739319 0.8854798
+    #> 3:     3 0.3378653 0.07326663 0.2480462 0.9002184
+    #> 4:     4 0.3923115 0.10844580 0.3137873 0.9213191
+    #> 5:     5 0.4341055 0.13426231 0.3711324 0.9299818
     ```
 
 -   use `orsf_summarize_uni()` to show the top predictor variables in an
@@ -116,44 +116,43 @@ How about interpreting the fit?
     #>         |---------------- risk ----------------|
     #>   Value      Mean    Median     25th %    75th %
     #>  <char>     <num>     <num>      <num>     <num>
-    #>    0.80 0.2276726 0.1248961 0.05111495 0.3557978
-    #>    1.40 0.2473343 0.1394002 0.06143315 0.3806993
-    #>    3.52 0.3692963 0.2953896 0.17716271 0.5528412
+    #>    0.80 0.2256092 0.1153241 0.04558336 0.3588942
+    #>    1.40 0.2461640 0.1313467 0.06323910 0.3871898
+    #>    3.52 0.3671496 0.2847901 0.16566117 0.5423875
     #> 
     #> -- age (VI Rank: 2) -----------------------------
     #> 
     #>         |---------------- risk ----------------|
     #>   Value      Mean    Median     25th %    75th %
     #>  <char>     <num>     <num>      <num>     <num>
-    #>    41.5 0.2713199 0.1407412 0.04671284 0.4618216
-    #>    49.7 0.2984502 0.1723371 0.05323084 0.5051649
[…]
