### 2022-07-08 #532 @chjackson
@tdhock Is this the right link for the reviewer guide: (https://devguide.ropensci.org/reviewerguide.html).  The link in the bot post gives a 404.
