### 2022-07-09 200685a
split statistic for tree growing
I had this hard-coded for awhile and meant to change it to an
input for orsf but it slipped off the radar. Good to avoid hard
coded numeric values plus give a possible tuning parameter.
