+++ R/orsf_attr.R
@@ -26,10 +26,10 @@ get_leaf_min_events  <- function(object) attr(object, 'leaf_min_events')
 get_leaf_min_obs     <- function(object) attr(object, 'leaf_min_obs')
 get_split_min_events <- function(object) attr(object, 'split_min_events')
 get_split_min_obs    <- function(object) attr(object, 'split_min_obs')
+get_split_min_stat   <- function(object) attr(object, 'split_min_stat')
 get_cph_method       <- function(object) attr(object, 'cph_method')
 get_cph_eps          <- function(object) attr(object, 'cph_eps')
 get_cph_iter_max     <- function(object) attr(object, 'cph_iter_max')
-get_cph_pval_max     <- function(object) attr(object, 'cph_pval_max')
 get_cph_do_scale     <- function(object) attr(object, 'cph_do_scale')
 get_net_alpha        <- function(object) attr(object, 'net_alpha')
 get_net_df_target    <- function(object) attr(object, 'net_df_target')

[…]
