### 2022-07-18 #532 @chjackson
## Package Review

    -   None

-   ☒ As the reviewer I confirm that there are no [conflicts of interest](https://devguide.ropensci.org/policies.html#coi) for me to review this work (if you are unsure whether you are in conflict, please speak to your editor *before* starting your review).

#### Documentation

-   ☒ **A statement of need:** clearly stating problems the software is designed to solve and its target audience in README

-   ☒ **Installation instructions:** for the development version of package and any non-standard dependencies in README

-   ☒ **Vignette(s):** demonstrating major functionality that runs successfully locally

-   ☒ **Function Documentation:** for all exported functions

    -   I built and installed the github package locally, and did `help(package="aorsf")`. I couldn’t see a `help("aorsf-package")` from R, is this because it it labelled with keyword “internal”? I don’t know if a package overview help page is strictly necessary given the DESCRIPTION and README files, but I sometimes find them helpful (is this the “multiple points of entry” principle?).

-   ☒ **Examples:** (that run successfully locally) for all exported functions

    -   There are currently no examples for [orsf](https://bcjaeger.github.io/aorsf/reference/orsf.html) or [orsf\_train](https://bcjaeger.github.io/aorsf/reference/orsf.html) on the website, but these appeared in the local R help. All the others ran with no problem.

-   ☒ **Community guidelines:** including contribution guidelines in the README or CONTRIBUTING, and DESCRIPTION with `URL`, `BugReports` and `Maintainer` (which may be autogenerated via `Authors@R`).

    -   I couldn’t find anything about community / contribution guidelines.

#### Functionality

-   ☒ **Installation:** Installation succeeds as documented.

-   ☒ **Functionality:** Any functional claims of the software been confirmed.

-   ☐ **Performance:** Any performance claims of the software been confirmed.

    -   Given my lack of machine learning experience I don’t think I’m qualified to judge relative peformance of different ML packages - I wouldn’t spot if they were using different assumptions that would cloud the comparison.

-   ☒ **Automated tests:** Unit tests cover essential functions of the package and a reasonable range of inputs and conditions. All tests pass on the local machine.

    -   Looks fairly comprehensive on a quick look through, and they passed.

-   ☒ **Packaging guidelines**: The package conforms to the rOpenSci packaging guidelines.

    -   I didn’t have time to go through every single point made in these guidelines, but what I saw of the package made me confident that it does what a well-organised and accessible package should do.

Estimated hours spent reviewing: 5

-   ☒ Should the author(s) deem it appropriate, I agree to be acknowledged as a package reviewer (“rev” role) in the package DESCRIPTION file.

### Review Comments

While I have a lot of experience with parametric statistical modelling of survival data, I am not knowledgeable about machine learning. So I can’t go into depth on the ML methods side of things. Though the vignettes explained clearly what the method is supposed to do - I felt like I learnt some useful things from them.

As far as I can tell, this is a polished, easily accessible package that implements a widely useful method. The differences and advances from other packages and methods are explained concisely.

I tried the package out by using it on a survival dataset that I had been using to test some of my own work, the `colon` data from the `survival` package. This is a simple randomised trial that compares time to cancer recurrence or death for three treatment groups, and I considered treatment as the only predictor. This may not be an ideal application for a random forest model, but hopefully it is helpful to see how the method does on this case.

    library(survival)
    library(dplyr)
    colonc <- colon |>
     filter(etype == 1) |>
     mutate(lev = factor(rx=="Lev"),
            lev5fu = factor(rx=="Lev+5FU"))

When using `orsf` with a single factor predictor, I got `"Error: formula must have at least 2 predictors"`. Perhaps it would be helpful to explain the restriction to 2+ predictors somewhere, though I guess that it is rare that people will want to use a machine learning method with only one predictor.

    fit <- orsf(data_train = colonc, formula = Surv(time, status) ~ rx)

When I recoded the data so the formula included two binary predictors, it worked. Though I realise that in this case the data contain only three out of the four possible combinations of the two binary variables, so it is impossible to identify whether there is an interaction.

I then calculated predicted survival at 2000 and 3000 days, under the two methods, for each of the four combinations of predictors.

    library(aorsf)
    fit <- orsf(data_train = colonc, formula = Surv(time, status) ~ lev + lev5fu)
    fit <- orsf(data_train = colonc, formula = Surv(time, status) ~ lev + lev5fu, n_tree = 100, n_split=3)
    pd_spec <- list(lev=c("FALSE","TRUE"),
                    lev5fu=c("FALSE","TRUE"))
    pd_data <- orsf_pd_summary(object = fit, pd_spec = pd_spec,
                               pred_horizon = c(2000, 3000),
                               oobag = TRUE, expand_grid = TRUE,
                               risk=FALSE)
    pd_data

    ## 8:         3000  TRUE   TRUE 0.5405302 0.5183706 0.5406791 0.5655176

I compared the `osrf` model to a flexible parametric spline-based, non-proportional hazards survival model from the `flexsurv` package.

    library(flexsurv)
    spl <- flexsurvspline(Surv(time, status) ~ lev + lev5fu + gamma1(lev) + gamma1(lev5fu), data=colonc, k=5)
    summary(spl, newdata=expand.grid(pd_spec), t=c(2000,3000), tidy=TRUE) |>
     arrange(time, lev5fu, lev)

    ## 8 3000 0.5958252 0.5095980 0.6783956  TRUE   TRUE

The predictions agreed for each combination except the one with both factor levels `TRUE`. This is understandable because these combinations do not appear in the data, and the two methods will be relying on different assumptions to extrapolate to this combination.

At this point I wondered how the intervals in the output were determined, and what do they mean? Do they reflect uncertainty about an expected value, or variability in observed values, or something else? The [help page](https://bcjaeger.github.io/aorsf/reference/orsf_pd_summary.html) (argument `prob_values`) explains them as quantiles, but of what? It can’t be quantiles of some subset of the data with that combination of predictors, as there isn’t any such data in the training set. Perhaps this is obvious to someone with more experience with random forests or related ML methods.

#### Observations from the Introduction vignette

-   The y axis for “partial dependence of bilirubin and edema” should start at 0. I’d also end it at 1, but this is perhaps only necessary if the graph is intended to be compared with other graphs of predicted risk.

-   It wasn’t clear what the package was doing when the `pred_horizon` argument was not supplied to the functions that did prediction. What horizon is being used? I found that `orsf_pd_summary` didn’t show in the output what prediction time was being used, in cases where the prediction time wasn’t supplied by the user. Though from `orsf_summarize_uni`, I deduced this was the median follow-up time in the data. I’d recommend showing good practice in the worked examples, e.g. using a meaningful prediction time that someone doing the analysis might use (e.g. 5 years). As a developer, I’ve found that it’s common for users to do silly things by copying from worked examples in manuals without thinking.
