### 2022-07-18 #532 @bcjaeger
@chjackson, thank you! All of the suggestions from your review make sense and are actionable. I plan to make changes to `aorsf`'s documentation based on your feedback.

Regarding performance, I decided to put almost all of the benchmarking results in a separate project: https://github.com/bcjaeger/aorsf-bench. If you would like to review the performance claims, I recommend checking this repo out and in particular paper/jmlr/main.pdf, which is the first draft of a paper introducing the methods (not the software) involved in this R package. The results in this paper verify claims about performance in the `aorsf` documentation.
