### 2022-07-18 #532 @nicholasjhorton
Dear @tdhock: I'm sorry but I'm feeling quite confused. What did I sign up? I seem to have been added but don't know what I was asked to do. Can you please resend the invitation?
