### 2022-07-18 #532 @tdhock
Sorry for the confusion. You did not agree to review the aorsf package? If you can write the review please see instructions, https://devguide.ropensci.org/reviewerguide.html Otherwise please tell me if you would like me to remove you from the reviewer list.
