### 2022-07-18 #532 @nicholasjhorton
I didn't agree. The first I heard of the request was https://github.com/ropensci/software-review/issues/532#issuecomment-1179231500.

Unfortunately, I'm not able to review at this time. Can you please remove me from the reviewer list?
