### 2022-07-19 #532 @chjackson
Thanks, @bcjaeger - I'm happy with the performance claims from a quick scan of that paper.
