### 2022-07-19 292768b
multiple points of entry
The package doc page is now accessible and the package logo appears in
the website, the doc page, and the readme.
