+++ R/aorsf-package.R
@@ -1,10 +1,10 @@
 #' @keywords internal
+#' @import data.table
 "_PACKAGE"
 
 # The following block is used by usethis to automatically manage
 # roxygen namespace tags. Modify with care!
 ## usethis namespace: start
-#' @import data.table
 #' @importFrom Rcpp sourceCpp
 #' @useDynLib aorsf, .registration = TRUE
 ## usethis namespace: end
