+++ README.md
@@ -1,7 +1,7 @@
 
 <!-- README.md is generated from README.Rmd. Please edit that file -->
 
-# aorsf
+# aorsf <a href="https://bcjaeger.github.io/aorsf"><img src="man/figures/logo.png" align="right" height="138" /></a>
 
 <!-- badges: start -->
 
@@ -61,7 +61,7 @@ print(fit)
 #>                  N trees: 500
 #>       N predictors total: 17
 #>    N predictors per node: 5
-#>  Average leaves per tree: 19
+#>  Average leaves per tree: 24
 #> Min observations in leaf: 5
 #>       Min events in leaf: 1
 #>           OOB stat value: 0.84
@@ -76,14 +76,14 @@ How about interpreting the fit?
 
     ``` r
     orsf_vi_negate(fit)
-    #>          bili           age           sex        copper       ascites 
-    #>  0.0160971036  0.0120858512  0.0084392582  0.0068243384  0.0055740779 
-    #>       spiders         stage         edema        hepato           ast 
-    #>  0.0052094186  0.0043238175  0.0022673374  0.0021358616  0.0012502605 
-    #>      platelet       albumin          chol          trig       protime 
-    #>  0.0005209419  0.0003646593  0.0001562826 -0.0001562826 -0.0010939779 
-    #>      alk.phos           trt 
-    #> -0.0015107314 -0.0040112523
+    #>          bili           age       protime       ascites       albumin 
+    #>  0.0129714524  0.0125026047  0.0085955407  0.0063554907  0.0061992082 
+    #>       spiders           sex        copper         edema           ast 
+    #>  0.0057303605  0.0049489477  0.0044280058  0.0027783566  0.0018753907 
+    #>        hepato          trig         stage      alk.phos      platelet 
+    #>  0.0018232965  0.0016670140  0.0005209419 -0.0006772244 -0.0008856012 
+    #>          chol           trt 
+    #> -0.0016670140 -0.0025526151
     ```
 
 -   use `orsf_pd_ice()` or `orsf_pd_summary()` for individual or
@@ -91,13 +91,12 @@ How about interpreting the fit?
 
     ``` r
     orsf_pd_summary(fit, pd_spec = list(bili = c(1:5)))
-    #>     bili      mean        lwr      medn       upr
-    #>    <int>     <num>      <num>     <num>     <num>
-    #> 1:     1 0.2309318 0.01708277 0.1168329 0.8542309
-    #> 2:     2 0.2801009 0.04071749 0.1739319 0.8854798
-    #> 3:     3 0.3378653 0.07326663 0.2480462 0.9002184
-    #> 4:     4 0.3923115 0.10844580 0.3137873 0.9213191
-    #> 5:     5 0.4341055 0.13426231 0.3711324 0.9299818
+    #>    bili      mean        lwr      medn       upr
+    #> 1:    1 0.2376652 0.01286480 0.1331328 0.8669733
+    #> 2:    2 0.2904995 0.04154464 0.1889294 0.8943053
+    #> 3:    3 0.3441193 0.06285072 0.2538335 0.9146453
+    #> 4:    4 0.3973732 0.10091026 0.3247072 0.9271855
+    #> 5:    5 0.4424865 0.14005297 0.3807676 0.9315099
     ```
 
 -   use `orsf_summarize_uni()` to show the top predictor variables in an
@@ -111,48 +110,44 @@ How about interpreting the fit?
 
     orsf_summarize_uni(object = fit, n_variables = 5)
     #> 
-    #> -- bili (VI Rank: 1) ----------------------------
+    #> -- bili (VI Rank: 1) ---------------------------
     #> 
-    #>         |---------------- risk ----------------|
[…]
