### 2022-07-19 cf10d04
example code pkgdown fix
