+++ man/aorsf-package.Rd
@@ -6,6 +6,8 @@
 \alias{aorsf-package}
 \title{aorsf: Accelerated Oblique Random Survival Forests}
 \description{
+\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}
+
 Fit, interpret, and make predictions with oblique random survival forests. Oblique decision trees are notoriously slow compared to their axis based counterparts, but 'aorsf' runs as fast or faster than axis-based decision tree algorithms for right-censored time-to-event outcomes.
 }
 \seealso{
