### 2022-07-19 febc3ba
clarifications
the error presented when <2 predictors were supplied did not indicate
why orsf needs at least 2 predictors. Also, the probs argument in
orsf_pd did not indicate what the quantiles were being computed from.
