### 2022-07-19 7ad9a94
clarify intro vignette
the y axes should be aligned and cover the entire spectrum of possible
values for PD. Also, pred_horizon should generally be specified and it
was not helpful for the examples to leave it unspecified.
