+++ man/orsf_summarize_uni.Rd
@@ -39,6 +39,13 @@ an object of class 'aorsf_summary'
 \description{
 ORSF summary of univariate information
 }
+\details{
+if \code{pred_horizon} is left unspecified, the median value of
+the time-to-event variable in \code{object}'s training data will be used.
+It is recommended to always specify your own prediction horizon,
+as the median time may not be an especially meaningful horizon to
+compute predicted risk values at.
+}
 \examples{
 
 object <- orsf(pbc_orsf, Surv(time, status) ~ . - id)
