### 2022-07-20 #532 @bcjaeger
Thank you! I pushed some changes to the main branch of `aorsf` today based on the feedback from your review. I really appreciate your time and thoughts on the package. You are more than welcome to review the added clarifications to the documentation and let me know if there is anything you'd recommend modifying. Seeing your response above, I will add you as 'rev' to the DESCRIPTION file.
