### 2022-07-20 #532 @chjackson
That 's fine to add me, thanks.  The doc changes look good.

Is it worth adding to the help pages what the functions do you omit `pred_horizon`, and add in the corresponding output data frame the specific prediction time being used?  Even if it's sensible to use an explicit prediction time, perhaps someone would have a genuine reason for using a default based on a summary of the data, or they may leave `pred_horizon` out by accident then be curious about what the output means.

For my own interest, I'm still not sure what the intervals mean.  When you say "The quantiles are calculated based on predictions from object at each set of values indicated by pd_spec.", does that mean that the predicted risk for a given set of covariate values is stochastic?  Then how is the point prediction defined, and does this stochastic variation describe uncertainty about the expected risk (which can be reduced by collecting more data), or variation in observed outcomes (which is a characteristic of the population that the data are drawn from)?
