### 2022-07-20 #532 @bcjaeger
Thanks! I think it would be worth noting that `pred_horizon` plays a role in all prediction functions. Perhaps I could write that statement into the details or directly into the description of the `pred_horizon` input?

It is also worth mentioning that `pred_horizon` can be pre-determined by the object returned by `orsf()`. When you run `orsf()`, you can specify an input called `oobag_time` that will set `pred_horizon` for your out-of-bag predictions.

__possible API change__ I am wondering if it would make sense to change the name of the input `oobag_time` in `orsf()` to `oobag_pred_horizon` for consistency within the package. Do you think that would be helpful?

As far as partial dependence goes, suppose we have an outcome Y modeled as a function of X and we want to compute partial dependence of Y at X = 1. We first set X equal to 1 for all rows in the training data (or whatever data we are using to compute partial dependence) and then compute the predicted value of Y for each row in the modified data, leaving us with N predictions if we have N observations. The point estimate for partial dependence is the mean of the predictions we compute, but you could also compute the 25th percentile of those predictions, the median, the 75th percentile, or whatever summary you like. I definitely wouldn't think of those percentiles as confidence bounds, but they do a nice job of quantifying how much the predictions vary across the training data when X is fixed at 1. Does that make sense? Perhaps we could add a definition like this into the partial dependence vignette?
