+++ man/aorsf-package.Rd
@@ -26,6 +26,7 @@ Other contributors:
 \itemize{
   \item Nicholas Pajewski [contributor]
   \item Sawyer Welden \email{swelden@wakehealth.edu} [contributor]
+  \item Christopher Jackson \email{chris.jackson@mrc-bsu.cam.ac.uk} [reviewer]
 }
 
 }
