+++ DESCRIPTION
@@ -1,14 +1,23 @@
 Package: aorsf
 Title: Accelerated Oblique Random Survival Forests
-Version: 0.0.0.9000
+Version: 0.0.1
 Authors@R: c(
     person(given = "Byron",
            family = "Jaeger",
            role = c("aut", "cre"),
            email = "bjaeger@wakehealth.edu",
            comment = c(ORCID = "0000-0001-7399-2299")),
-    person(given = "Nicholas",  family = "Pajewski", role = "ctb"),
-    person(given = "Sawyer", family = "Welden", role = "ctb", email = "swelden@wakehealth.edu")
+    person(given = "Nicholas",  
+           family = "Pajewski", 
+           role = "ctb"),
+    person(given = "Sawyer",
+           family = "Welden", 
+           role = "ctb", 
+           email = "swelden@wakehealth.edu"),
+    person(given = "Christopher", 
+           family = "Jackson",
+           email = "chris.jackson@mrc-bsu.cam.ac.uk",
+           role = "rev")
     )
 Description: Fit, interpret, and make predictions with oblique random
     survival forests. Oblique decision trees are notoriously slow compared
