### 2022-07-22 4d962eb
api changes
data_train -> data, since this is consistent with almost every other
modeling package.

oobag_time -> oobag_pred_horizon, since pred_horizon is the term used by
prediction functions.
