+++ R/orsf_attr.R
@@ -38,7 +38,7 @@ get_n_retry          <- function(object) attr(object, 'n_retry')
 get_f_oobag_eval     <- function(object) attr(object, 'f_oobag_eval')
 get_type_oobag_eval  <- function(object) attr(object, 'type_oobag_eval')
 get_oobag_pred       <- function(object) attr(object, 'oobag_pred')
-get_oobag_time       <- function(object) attr(object, 'oobag_time')
+get_oobag_pred_horizon       <- function(object) attr(object, 'oobag_pred_horizon')
 get_oobag_eval_every <- function(object) attr(object, 'oobag_eval_every')
 get_importance       <- function(object) attr(object, 'importance')
 get_f_beta           <- function(object) attr(object, 'f_beta')
