### 2022-07-22 #532 @bcjaeger
Excellent! Just pushed the API changes to the main branch. In addition to changing `oobag_time` to `oobag_pred_horizon`, I also changed `data_train` to `data` in `orsf()`. The second change was done to have consistency between `orsf()` and other modeling functions, most of which denote training data as `data` in their inputs.
