### 2022-07-22 1b2eb11
test cleaner
only run the intentional error tests in dev version, and suppress
messages with printed output so that the testthat summary is easier to
read.
