+++ tests/testthat/test-orsf_print.R
@@ -1,8 +1,8 @@
 
 
-test_that("no errors for standard object", {
+test_that("standard object prints fine", {
 
  object <- orsf(pbc_orsf, time+status ~ . - id)
- expect_invisible(print(object))
+ expect_invisible(p <- capture.output(print(object)))
 
 })
