### 2022-07-24 f87a73a
prepare to cite aorsf paper
the benchmark paper will be on arxiv soon enough, so the JOSS paper does
not need to provide complete justification on performance claims. It can
just cite the arxiv print.
