### 2022-07-24 b3afa7a
weights
needed to put weight input into the API before submitting to CRAN
because I am pretty sure the weight input will play a role in future
projects.
