+++ R/orsf_attr.R
@@ -46,6 +46,7 @@ get_orsf_type        <- function(object) attr(object, 'orsf_type')
 get_f_oobag_eval     <- function(object) attr(object, 'f_oobag_eval')
 get_type_oobag_eval  <- function(object) attr(object, 'type_oobag_eval')
 get_tree_seeds       <- function(object) attr(object, 'tree_seeds')
+get_weights_user     <- function(object) attr(object, 'weights_user')
 
 #' ORSF status
 #'

[…]
