### 2022-07-26 a318e9e
prep for cran submit
