+++ NEWS.md
@@ -10,4 +10,4 @@
 
 * Added `orsf_control_custom()`, which allows users to submit custom functions for identifying linear combinations of inputs while growing oblique decision trees.
 
-* To be filled in as ROpenSci review continues
+* Added `weights` input to `orsf`, allowing users to over or under fit `orsf` to specific data in their training set.
