+++ cran-comments.md
@@ -1,0 +1,10 @@
+
+## R CMD check results 
+
+Duration: 3m 52.7s
+
+0 errors v | 0 warnings v | 0 notes v
+
+## Downstream dependencies
+
+I have also run R CMD check on downstream dependencies of aorsf: Rcpp and data.table. Both packages passed.
