### 2022-07-26 #532 @tdhock
I'm not sure if this is in scope for the package review, but I had a look at your aorsf-bench slides/figures, and I saw one figure that compared the computation time of various packages/algos, I think for a single data set/size. That is a good start, but I was wondering if you have done any analysis of the computation time of the packages/algos as a function of data size? (plot time vs number of features and number of observations, is it expected to scale linear for each algo, or..?)
