### 2022-07-26 #532 @bcjaeger
Hi @tdhock - thank you for checking that out!

The figure you are referring to is https://bcjaeger.github.io/aorsf-bench/#32, right? The slide does not make this clear, but the times in the figure for each algo are aggregated across all of the risk prediction tasks on this slide: https://bcjaeger.github.io/aorsf-bench/#29. So I'd say the timing estimates are more general than timing estimates on one dataset, but they still do not show patterns on how each algo scales with more observations or more features.

I very much like the idea of showing computation time as a function of the number of observations and as a function of the number of features. Would it make sense to do this by focusing on one of the larger benchmark datasets, and then using subsets of that data with progressively higher dimensions? e.g., suppose one dataset has 100 features and 10,000 rows. We could run:

1. using the first 1000 observations and all features
2. using the first 2500 observations and all features
3. using the first 5000 observations and all features
4. using the first 7500 observations and all features
5. using the first 10000 observations and all features

And then do a similar routine for features, holding the number of observations fixed.
