### 2022-07-26 #532 @tdhock
Hi @bcjaeger yes that sounds good, but I would suggest avoiding linear scales (2500, 5000, 7500). Typically for this kind of analysis I would suggest varying the number of rows/columns on a log scale, like (100, 1000, 10000) or `N=as.integer(seq(2, 4, by=0.5))`
For an example of what kind of figure I am thinking of see
![image](https://user-images.githubusercontent.com/932850/181114693-9d486984-559e-4c5f-8d8b-902fbb88490a.png)
which is from Fig 7 of https://jmlr.org/papers/volume21/18-843/18-843.pdf
