### 2022-07-26 #532 @bcjaeger
Ahh, that makes perfect sense. I will add this to the paper I am working on in the aorsf-bench repo.
