+++ man/pbc_orsf.Rd
@@ -25,7 +25,7 @@ A data frame with 276 rows and 20 variables:
 \item{ast}{aspartate aminotransferase, once called SGOT (U/ml)}
 \item{trig}{triglycerides (mg/dl)}
 \item{platelet}{platelet count}
-\item{protime}{standardised blood clotting time}
+\item{protime}{standardized blood clotting time}
 \item{stage}{histologic stage of disease (needs biopsy)}
 }
 }
