### 2022-07-27 #532 @bcjaeger
So it will take me awhile to get this result written into the paper but here is a very rough figure to be responsive to @tdhock's earlier suggestion. Each facet in the figure has the number of predictor variables printed at the top strip (10, 100, or 1,000). Each point is the mean value of time to fit the corresponding model, taken over 10 independent runs and allowing up to 4 CPUs for parallel computing.  Overall, the `ranger` package tends to do best with lower values of n, but its computing time scales up very fast. This may be due to memory usage (I noticed my working memory was completely used up while ranger was running). the `aorsf` package tends to be similar to `randomForestSRC` (`rsf_rfsrc` in the figure), with `aorsf` having a narrow advantage when p = 10 or 100 and `randomForestSRC` winning out when p = 1000. the `party` package scales consistently but is generally slower than `ranger`, `randomForestSRC`, and `aorsf`. I am puzzled by the way that `randomForestSRC` scales better when p is 1000 but not when p is 100 or 10.  Perhaps a special computational routine is activated in `randomForestSRC` when p is very large? It seems like a different computation may be taking place in `randomForestSRC` with n>1000 as well.

The data used here were simulated, and all predictors were continuous.

![image](https://user-images.githubusercontent.com/11821781/181296577-b4bc5a31-ecff-447f-954a-8357ce44e4bf.png)
