### 2022-07-27 #532 @tdhock
In response to that timing figure, that looks like a good start. Is the Y axis seconds? It would be helpful to add error bands to see the variance in timings (geom_ribbon with ymin=min or lower quartile over 5 or 10 timings, etc).
On the log-log plot the different asymptotic complexity classes show up as different slopes on the right, for large n_obs values. Looks like party and aorsf have the same complexity class (linear?). For the others, perhaps your n_obs is not large enough to see the asymptotic complexity class clearly, so you may try increasing the max n_obs to see that.
