### 2022-07-31 #532 @bcjaeger
Sorry for the delay - I am happy to share an update:

- increased the max sample size to be slightly over 10k (100k caused rstudio to crash, likely b/c of memory)
- reduced the number of trees per benchmark run to 1 (down from 50) in order to make memory crashes less likely.
- increased the number of times we replicated the benchmark to 50 (up from 10)
- increased the number of sample sizes we examined (now showing 8 different sample sizes: `round(10^seq(from = 2, to = 4+1/3, by = 1/3))`)
- ribbons based on the 2.5th and 97.5th percentile of the run times and better labels for R packages, panels, and axes.

As far as my interpretation goes, I think `ranger` has different computational engines that get used depending on the dimension of the input data.  `aorsf`, `randomForestSRC`, and `party` have pretty consistent patterns, and `aorsf` does well overall.

What do you think?

![temp](https://user-images.githubusercontent.com/11821781/182029369-de734987-3819-4257-afdf-10d4a3c599a0.png)
