### 2022-08-04 #532 @tdhock
Hi, thanks for sharing. That figure/result looks very interesting, good job. Based on the figure, which has same slopes for each line on that log log plot, it looks like all of the packages have the same asymptotic complexity class, which should be linear, right?
