### 2022-08-04 #532 @bcjaeger
Yes, I think so! Thank you very much for your help improving the figure. I intend to include this work in the paper describing `aorsf`'s methods and efficiency. I have two questions:

1. would it be reasonable/helpful for me to include a note to you or to rOpenSci in the acknowledgments section of the paper I have referred to above (particularly for your help guiding the timing benchmark above)?
2. do you think this review of `aorsf` is complete?

I am very happy with the feedback I've gotten and feel ready to proceed with submissions to JOSS and CRAN.
