+++ joss/paper.bib
@@ -28,6 +28,24 @@
 	note = {original-date: 2016-04-28T18:05:02Z},
 }
 
+@misc{aorsf_arxiv,
+  doi = {10.48550/ARXIV.2208.01129},
+
+  url = {https://arxiv.org/abs/2208.01129},
+
+  author = {Jaeger, Byron and Welden, Sawyer and Lenoir, Kristin and Speiser, Jaime L. and Segar, Matthew W. and Pandey, Ambarish and Pajewski, Nicholas M.},
+
+  keywords = {Methodology (stat.ME), Machine Learning (stat.ML), FOS: Computer and information sciences, FOS: Computer and information sciences},
+
+  title = {Accelerated and interpretable oblique random survival forests},
+
+  publisher = {arXiv},
+
+  year = {2022},
+
+  copyright = {Creative Commons Attribution 4.0 International}
+}
+
 @article{breiman2001random,
   title={Random forests},
   author={Breiman, Leo},
@@ -50,7 +68,7 @@
 
 @article{jaeger2019oblique,
   title={Oblique random survival forests},
-  author={Jaeger, Byron C and Long, D Leann and Long, Dustin M and Sims, Mario and Szychowski, Jeff M and Min, Yuan-I and Mcclure, Leslie A and Howard, George and Simon, Noah},
+  author={Jaeger, Byron and Long, D Leann and Long, Dustin M and Sims, Mario and Szychowski, Jeff M and Min, Yuan-I and Mcclure, Leslie A and Howard, George and Simon, Noah},
   journal={The Annals of Applied Statistics},
   volume={13},
   number={3},
