### 2022-08-08 6f68a93
pred_type for future API
the type of prediction needs to be a character input so that I can add
more types in the future without breaking the API. Also the pd functions
that call predict need to have the same type argument and the name
'type' is ambiguous in functions that aren't predict. So...call it
pred_type
