+++ R/orsf_predict.R
@@ -18,22 +18,26 @@
 #' @srrstats {G2.1a} explicit secondary documentation of expectations on data types of all vector inputs
 #'
 #' @param pred_horizon (_double_) a single time or a vector of times
-#'   indicating the prediction horizon. Predicted risk or survival values
-#'   will indicate the probability of having an event or surviving from
-#'   baseline to the prediction horizon, respectively. All `pred_horizon`
-#'   values must not exceed the maximum follow-up time in `object`'s
-#'   training data. Also, `pred_horizon` values must be entered in
-#'   ascending order.
+#'   indicating the time(s) that predicted risk or survival probabilities
+#'   will be computed at.
 #'
-#' @param risk (_logical_) if `TRUE`, predicted risk is returned. If `FALSE`,
-#'   predicted survival (i.e., 1-risk) is returned.
+#' @param pred_type (_character_) the type of predictions to return. Valid
+#'   options are
 #'
+#'   - 'risk' : probability of having an event at or before `pred_horizon`.
+#'   - 'survival' : 1 - risk.
 #'
 #' @param ... not used.
 #'
 #' @return a `matrix` of predictions. Column `j` of the matrix corresponds
-#'   to value `j` in `pred_horizon`. Row `i` of the matrix corresponds to row `i`
-#'   in `new_data`.
+#'   to value `j` in `pred_horizon`. Row `i` of the matrix corresponds to
+#'   row `i` in `new_data`.
+#'
+#' @details
+#'
+#' `pred_horizon` values must not exceed the maximum follow-up time in
+#'   `object`'s training data. Also, `pred_horizon` values must be entered
+#'   in ascending order.
 #'
 #' @export
 #'
@@ -59,7 +63,7 @@
 predict.aorsf <- function(object,
                           new_data,
                           pred_horizon,
-                          risk = TRUE,
+                          pred_type = 'risk',
                           ...){
 
  #' @srrstats {G2.13} *check for missing data as part of initial pre-processing prior to passing data to analytic algorithms.*
@@ -73,13 +77,15 @@ predict.aorsf <- function(object,
 
  #' @srrstats {G2.8} *As part of initial pre-processing, run checks on inputs to ensure that all other sub-functions receive inputs of a single defined class or type.*
 
- check_predict(object, new_data, pred_horizon, risk)
+ check_predict(object, new_data, pred_horizon, pred_type)
 
  x_new <- as.matrix(
   ref_code(x_data = new_data,
            fi = get_fctr_info(object),
            names_x_data = names_x_data)
  )
+
+ risk <- pred_type == 'risk'
 
  if(length(pred_horizon) == 1L)
   return(orsf_pred_uni(object$forest, x_new, pred_horizon, risk))
