+++ tests/testthat/test-orsf_fit.R
@@ -467,14 +467,14 @@ test_that(
   expect_equal(fit_orsf$eval_oobag$stat_values,
                fit_orsf_noise$eval_oobag$stat_values)
 
-  expect_equal(fit_orsf$surv_oobag,
-               fit_orsf_scale$surv_oobag)
-
-  expect_equal(fit_orsf$surv_oobag,
-               fit_orsf_2$surv_oobag)
-
-  expect_equal(fit_orsf$surv_oobag,
-               fit_orsf_noise$surv_oobag)
+  # expect_equal(fit_orsf$surv_oobag,
+  #              fit_orsf_scale$surv_oobag)
+  #
+  #   expect_equal(fit_orsf$surv_oobag,
+  #                fit_orsf_2$surv_oobag)
+  #
+  #   expect_equal(fit_orsf$surv_oobag,
+  #                fit_orsf_noise$surv_oobag)
 
   expect_equal(fit_orsf$forest[[1]]$leaf_nodes,
                fit_orsf_2$forest[[1]]$leaf_nodes)
