### 2022-08-08 2519e45
random seed before noise added
