### 2022-08-10 1ae07de
more flexible data input
Max Ks guidelines for R packages suggest lists and recipes should be
allowed as inputs. This update keeps the main orsf function intact but
adds a function to transform data from either recipe or list objects
into the format that orsf expect.
