### 2022-08-10 b076611
dots for backward compatibility
in case i add args to orsf() or predict() in the future, the ... is
placed in to these functions for backward compatibility.
