### 2022-08-11 ff3e9b0
updated refs reflect arxiv paper
