+++ src-i386/orsf.cpp
@@ -1,0 +1,3907 @@
+#include <RcppArmadillo.h>
+#include <RcppArmadilloExtensions/sample.h>
+
+// [[Rcpp::depends(RcppArmadillo)]]
+
+
+using namespace Rcpp;
+using namespace arma;
+
+// ----------------------------------------------------------------------------
+// ---------------------------- global parameters -----------------------------
+// ----------------------------------------------------------------------------
+
+// special note: dont change these doubles to uword,
+//               even though some of them could be uwords;
+//               operations involving uwords and doubles are not
+//               straightforward and may break the routine.
+// also: double + uword is slower than double + double.
+
+double
+ weight_avg,
+ weight_events,
+ w_node_sum,
+ denom_events,
+ denom,
+ cph_eps,
+ // the n_ variables could be integers but it
+ // is safer and faster when they are doubles
+ n_events,
+ n_events_total,
+ n_events_right,
+ n_events_left,
+ n_risk,
+ n_risk_right,
+ n_risk_left,
+ n_risk_sub,
+ g_risk,
+ temp1,
+ temp2,
+ halving,
+ stat_current,
+ stat_best,
+ w_node_person,
+ xb,
+ risk,
+ loglik,
+ cutpoint,
+ observed,
+ expected,
+ V,
+ leaf_min_obs,
+ leaf_min_events,
+ split_min_events,
+ split_min_obs,
+ split_min_stat,
+ time_pred,
+ ll_second,
+ ll_init,
+ net_alpha;
+
+int
+ // verbose=0,
+ max_retry,
+ n_retry,
+ tree,
+ mtry_int,
+ net_df_target,
+ oobag_eval_every;
+
+char type_beta, type_oobag_eval, oobag_importance_type;
+
+// armadillo unsigned integers
+uword
+ i,
+ j,
+ k,
+ iter,
+ mtry,
+ mtry_temp,
+ person,
+ person_leaf,
+ person_ref_index,
+ n_vars,
+ n_rows,
+ cph_method,
+ cph_iter_max,
+ n_split,
+ nodes_max_guess,
+ nodes_max_true,
+ n_cols_to_sample,
+ nn_left,
+ leaf_node_counter,
+ leaf_node_index_counter,
+ oobag_eval_counter;
+
+bool
+ break_loop, // a delayed break statement
+ oobag_pred,
+ oobag_importance,
+ use_tree_seed,
+ cph_do_scale;
+
+// armadillo vectors (doubles)
+vec
+ vec_temp,
+ times_pred,
+ eval_oobag,
+ node_assignments,
+ nodes_grown,
+ surv_pvec,
+ surv_pvec_output,
+ denom_pred,
+ beta_current,
+ beta_new,
+ beta_fit,
+ vi_pval_numer,
+ vi_pval_denom,
+ cutpoints,
+ w_input,
+ w_inbag,
+ w_user,
+ w_node,
+ group,
+ u,
+ a,
+ a2,
+ XB,
+ Risk;
+
+// armadillo unsigned integer vectors
+uvec
+ iit_vals,
+ jit_vals,
+ rows_inbag,
+ rows_oobag,
+ rows_node,
+ rows_leaf,
+ rows_node_combined,
+ cols_to_sample_01,
+ cols_to_sample,
+ cols_node,
+ leaf_node_index,
+ nodes_to_grow,
+ nodes_to_grow_next,
+ obs_in_node,
+ children_left,
+ leaf_pred;
+
+// armadillo iterators for unsigned integer vectors
+uvec::iterator
+ iit,
+ iit_best,
+ jit,
+ node;
+
+// armadillo matrices (doubles)
+mat
+ x_input,
[…]
