+++ README.Rmd
@@ -34,7 +34,7 @@ knitr::opts_chunk$set(
 
 ## Why aorsf?
 
-- over 400 times faster than `obliqueRSF`.
+- over 400 times faster than `obliqueRSF` (see Jaeger, 2019).
 
 - accurate predictions for time-to-event outcomes.
 
@@ -98,9 +98,15 @@ How about interpreting the fit?
 
   ```
 
+## Comparison to existing software
+
+Jaeger (2022) describes `aorsf` in detail, with a summary of the procedures used in the tree fitting algorithm and a general benchmark comparing `aorsf` with `obliqueRSF` (and several other learners) in terms of prediction accuracy and computational efficiency. 
+
 ## References
 
-Byron C. Jaeger, D. Leann Long, Dustin M. Long, Mario Sims, Jeff M. Szychowski, Yuan-I Min, Leslie A. Mcclure, George Howard, Noah Simon (2019). Oblique Random Survival Forests. Ann. Appl. Stat. 13(3): 1847-1883. URL https://doi.org/10.1214/19-AOAS1261 DOI: 10.1214/19-AOAS1261
+Jaeger BC, Long DL, Long DM, Sims M, Szychowski JM, Min YI, Mcclure LA, Howard G, Simon N. Oblique random survival forests. The Annals of Applied Statistics. 2019 Sep;13(3):1847-83. URL: https://doi.org/10.1214/19-AOAS1261 DOI: 10.1214/19-AOAS1261
+
+Jaeger BC, Welden S, Lenoir K, Speiser JL, Segar M, Pandey A, Pajewski NM. Accelerated and interpretable oblique random survival forests. arXiv e-prints. 2022 Aug 3:arXiv-2208. URL: https://arxiv.org/abs/2208.01129
 
 ## Funding
 

[…]
