+++ README.md
@@ -76,14 +76,14 @@ How about interpreting the fit?
 
     ``` r
     orsf_vi_negate(fit)
-    #>          bili           sex       protime           age         stage 
-    #>  1.187747e-02  6.720150e-03  6.407585e-03  6.042926e-03  5.730360e-03 
-    #>        copper       spiders       ascites         edema           ast 
-    #>  5.365701e-03  3.177745e-03  2.865180e-03  2.421139e-03  2.135862e-03 
-    #>        hepato          chol      alk.phos           trt       albumin 
-    #>  1.041884e-03  5.209419e-05 -5.209419e-04 -2.135862e-03 -3.021463e-03 
-    #>      platelet 
-    #> -5.834549e-03
+    #>          bili           age        copper           sex       ascites 
+    #>  0.0143259012  0.0135965826  0.0082829756  0.0078662221  0.0050010419 
+    #>         stage       spiders       protime           ast         edema 
+    #>  0.0041675349  0.0040112523  0.0036986872  0.0036465930  0.0027833180 
+    #>          chol      alk.phos        hepato       albumin          trig 
+    #>  0.0007293186  0.0006251302  0.0005730360  0.0004688477 -0.0017712023 
+    #>      platelet           trt 
+    #> -0.0032298395 -0.0036986872
     ```
 
 -   use `orsf_pd_ice()` or `orsf_pd_summary()` for individual or
@@ -93,11 +93,11 @@ How about interpreting the fit?
     orsf_pd_summary(fit, pd_spec = list(bili = c(1:5)))
     #>     bili      mean        lwr      medn       upr
     #>    <int>     <num>      <num>     <num>     <num>
-    #> 1:     1 0.2448441 0.01284079 0.1312983 0.8747557
-    #> 2:     2 0.2812156 0.02878862 0.1667878 0.8927174
-    #> 3:     3 0.3202419 0.04351310 0.2171733 0.9050862
-    #> 4:     4 0.3584524 0.06467319 0.2707156 0.9151960
-    #> 5:     5 0.3910040 0.08913003 0.3138446 0.9204218
+    #> 1:     1 0.2469777 0.01345288 0.1315400 0.8857421
+    #> 2:     2 0.2839531 0.02911864 0.1684011 0.8986682
+    #> 3:     3 0.3241288 0.04961583 0.2261766 0.9148388
+    #> 4:     4 0.3645008 0.07139964 0.2834042 0.9188526
+    #> 5:     5 0.3974148 0.09856039 0.3182883 0.9282819
     ```
 
 -   use `orsf_summarize_uni()` to show the top predictor variables in an
@@ -116,45 +116,43 @@ How about interpreting the fit?
     #>         |---------------- risk ----------------|
     #>   Value      Mean    Median     25th %    75th %
     #>  <char>     <num>     <num>      <num>     <num>
-    #>    0.80 0.2405862 0.1252352 0.04384391 0.3843450
-    #>     1.4 0.2568429 0.1461177 0.05419999 0.4047258
-    #>     3.5 0.3412388 0.2453132 0.12975781 0.5353945
+    #>    0.80 0.2426995 0.1308204 0.05059714 0.3869034
+    #>     1.4 0.2594904 0.1421002 0.06048081 0.4252482
+    #>     3.5 0.3457362 0.2538523 0.13447505 0.5257414
     #> 
-    #> -- sex (VI Rank: 2) -----------------------------
-    #> 
-    #>         |---------------- risk ----------------|
-    #>   Value      Mean    Median     25th %    75th %
-    #>  <char>     <num>     <num>      <num>     <num>
-    #>       m 0.3507520 0.2415970 0.11829094 0.5454135
[…]
