### 2022-08-11 1106ac6
refactor, expand docs, add wraps
there are a couple of changes to make the api easier to engage with, and
a number of notes added to details, examples, or inputs. All of this is
generally done in the sense of making the api feel more intuitive and
flexible.
