### 2022-08-11 517fbdc
include new function for pkgdown
