### 2022-08-11 73dd14c
simplify, don't fixate on interactions
