+++ R/orsf_print.R
@@ -74,7 +74,6 @@ print.aorsf <- function(x, ...){
      '\n-----------------------------------------',
      sep = '\n')
 
-
-
+ invisible(x)
 
 }
