### 2022-08-12 1217218
negation stable, permutation included
