### 2022-08-12 ea12949
consistent defn of ... in docs
