### 2022-08-12 33e4460
cleaner description of oobag_fun
