+++ vignettes/aorsf.Rmd
@@ -22,7 +22,7 @@ This article covers core features of the `aorsf` package. Separate articles will
 
 ## Background: ORSF
 
-The oblique random survival forest (ORSF) is an extension of the axis-based RSF algorithm that can be found in several R packages, including `RandomForestSRC`, `party`, and `ranger`. 
+The oblique random survival forest (ORSF) is an extension of the axis-based RSF algorithm. Axis based RSFs can be fit with several R packages, including `RandomForestSRC`, `party`, and `ranger`. However, apart from `aorsf` and `obliqueRSF`, few R packages fit ORSFs.
 
 ## What is an oblique decision tree?
 
@@ -36,7 +36,7 @@ knitr::include_graphics('tree_axis_v_oblique.png')
 
 ## Accelerated ORSF
 
-The purpose of `aorsf` ('a' is short for accelerated) is to provide routines to fit ORSFs that will scale adequately to large data sets. For example, `aorsf::orsf()` runs about 500 times faster than its predecessor, `obliqueRSF::ORSF()`.
+The purpose of `aorsf` ('a' is short for accelerated) is to provide routines to fit ORSFs that will scale adequately to large data sets. For example, `aorsf::orsf()` runs hundreds of times faster than its predecessor, `obliqueRSF::ORSF()`.
 
 To fit an accelerated ORSF model, use the `orsf` function:
 
@@ -66,7 +66,9 @@ orsf_fit <- pbc_orsf |>
 
 ```
 
-In addition to their reputation for slow computing, oblique decision trees are considered less interpretable than their axis-based counterparts. `aorsf` includes routines that use coefficients from linear combinations of input variables to estimate variable importance.
+## Interpretation
+
+`aorsf` includes several functions dedicated to interpretation of ORSFs, both through estimation of partial dependence and variable importance.
 
 ### Variable importance
 
@@ -74,13 +76,29 @@ In addition to their reputation for slow computing, oblique decision trees are c
 
 - To compute negation importance, ORSF multiplies each coefficient of that variable by -1 and then re-computes the out-of-sample (sometimes referred to as out-of-bag) accuracy of the ORSF model. 
 
-```{r}
+    ```{r}
 
-variable_importance <- orsf_vi_negate(orsf_fit)
+    orsf_vi_negate(orsf_fit)
 
-variable_importance
+    ```
 
-```
+- You can also compute variable importance using permutation, a more classical approach.
+
+    ```{r}
+  
+    orsf_vi_permute(orsf_fit)
+  
+    ```
+
+- A faster alternative to permutation and negation importance is ANOVA importance, which computes the proportion of times each variable obtains a low p-value (p < 0.01) while the forest is grown.
+
+    ```{r}
+    
+    orsf_vi_anova(orsf_fit)
+    
+    ```
+
+
 
 ### Partial dependence
 
