+++ R/orsf.R
@@ -1,3 +1,18 @@
+#' Oblique Random Survival Forest (ORSF)
+#'
+#' The oblique random survival forest (ORSF) is an extension of the RSF
+#'   algorithm developed by Ishwaran et al and maintained in the
+#'   `RandomForestSRC` package. The difference between ORSF and RSF is
+#'   that ORSF uses linear combinations of input variables whereas RSF
+#'   uses a single variable when growing new nodes in survival decision trees.
+#'   A linear combination is an expression constructed from a set of terms
+#'   by multiplying each term by a constant and adding the results (e.g.
+#'   a linear combination of x and y would be any expression of the form
+#'   ax + by, where a and b are constants). For more details on the ORSF
+#'   algorithm, see Jaeger et al, 2019. The `orsf()` function implements a
+#'   novel algorithm that speeds up the ORSF algorithm described by Jaeger
+#'   et al (see details).
+#'
 #' @srrstats {G1.4} *documented with Roxygen*
 #' @srrstats {G1.1} *aorsf is an improvement of the ORSF algorithm implemented in obliqueRSF, which was an extension of Hemant Ishwaran's random survival forest.*
 #' @srrstats {G1.3} *linear combinations of inputs defined.*
@@ -22,22 +37,13 @@
 #' @srrstats {ML3.0} *Model specification can be implemented prior to actual model fitting or training*
 #' @srrstats {ML3.0a} *As pre-processing, model specification, and training are controlled by the orsf() function, an input parameter (no_fit) enables models to be specified yet not fitted.*
 #' @srrstats {ML3.0c} *when no_fit=TRUE, orsf() will return an object that can be directly trained using orsf_train().*
-
-
-#' Oblique Random Survival Forest (ORSF)
+#' @srrstats {ML1.6a} *Explain why missing values are not admitted.*
+#' @srrstats {G1.0} *Jaeger et al describes the ORSF algorithm that aorsf is based on. Note: aorsf uses a different approach to create linear combinations of inputs for speed reasons, but orsf_control_net() allows users to make ensembles that are very similar to obliqueRSF::ORSF().*
+#' @srrstats {ML1.6b} *Explicit example showing how missing values may be imputed rather than discarded.*
+#' @srrstats {ML6.1} *clearly document how aorsf can be embedded within a typical full ML workflow.*
+#' @srrstats {ML6.1a} *Embed aorsf within a full workflow using tidymodels, tidyverse, and survivalROC.*
+#' @srrstats {ML5.2b} *Documentation includes examples of how to save and re-load trained model objects for their re-use.*
 #'
-#' The oblique random survival forest (ORSF) is an extension of the RSF
-#'   algorithm developed by Ishwaran et al and maintained in the
-#'   `RandomForestSRC` package. The difference between ORSF and RSF is
-#'   that ORSF uses linear combinations of input variables whereas RSF
-#'   uses a single variable when growing new nodes in survival decision trees.
-#'   A linear combination is an expression constructed from a set of terms
-#'   by multiplying each term by a constant and adding the results (e.g.
-#'   a linear combination of x and y would be any expression of the form
-#'   ax + by, where a and b are constants). For more details on the ORSF
-#'   algorithm, see Jaeger et al, 2019. The `orsf()` function implements a
-#'   novel algorithm that speeds up the ORSF algorithm described by Jaeger
-#'   et al (see details).
 #'
 #' @param data (_data.frame_) that contains the relevant variables.
 #'
@@ -228,10 +234,7 @@
[…]
