+++ man/orsf.Rd
@@ -137,14 +137,25 @@ directly submitted to \code{orsf_train()} so long as \code{attach_data} is \code
 \item{...}{Further arguments passed to or from other methods
 (not currently used).}
 
-\item{object}{an untrained aorsf object, created by setting
+\item{object}{an untrained 'aorsf' object, created by setting
 \code{no_fit = TRUE} in \code{orsf()}.}
 }
 \value{
 an accelerated oblique RSF object (\code{aorsf})
 }
 \description{
-Oblique Random Survival Forest (ORSF)
+The oblique random survival forest (ORSF) is an extension of the RSF
+algorithm developed by Ishwaran et al and maintained in the
+\code{RandomForestSRC} package. The difference between ORSF and RSF is
+that ORSF uses linear combinations of input variables whereas RSF
+uses a single variable when growing new nodes in survival decision trees.
+A linear combination is an expression constructed from a set of terms
+by multiplying each term by a constant and adding the results (e.g.
+a linear combination of x and y would be any expression of the form
+ax + by, where a and b are constants). For more details on the ORSF
+algorithm, see Jaeger et al, 2019. The \code{orsf()} function implements a
+novel algorithm that speeds up the ORSF algorithm described by Jaeger
+et al (see details).
 }
 \details{
 This function is based on and highly similar to the \code{ORSF} function
@@ -277,6 +288,60 @@ fit_standard_oobag$eval_oobag$stat_values
 
 fit_custom_oobag$eval_oobag$stat_values
 
+
+\dontrun{requires too many external packages
+
+# --------------------------------------------------------------------------
+# a standard internal validation workflow using aorsf and tidymodels
+# --------------------------------------------------------------------------
+
+
+library(tidymodels)
+library(tidyverse)
+library(survivalROC)
+library(aorsf)
+
+set.seed(329)
+
+# a recipe to impute missing values instead of discarding them.
+# (this is for illustration only. pbc_orsf does not have missing values)
+
+imputer <- recipe(x = pbc_orsf, time + status ~ .) |>
+ step_impute_mean(all_numeric_predictors()) |>
+ step_impute_mode(all_nominal_predictors()) |>
+ step_rm(id)
+
+
+# 10-fold cross validation; make a container for the pre-processed data
+analyses <- vfold_cv(data = pbc_orsf, v = 10) |>
+ mutate(recipe = map(splits, ~prep(imputer, training = training(.x))),
+        data = map(recipe, juice),
+        data_test = map2(splits, recipe, ~bake(.y, new_data = testing(.x))))
+
+# 10-fold cross validation; train models and compute test predictions
+aorsf_data <- analyses |>
+ select(data, data_test) |>
+ mutate(fit = map(data, orsf, formula = time + status ~ .),
+        pred = map2(fit, data_test, predict, pred_horizon = 3500),
+        pred = map(pred, as.numeric))
+
+# testing sets are small, so pool them and compute 1 overall C-stat.
+aorsf_eval <- aorsf_data |>
+ select(data_test, pred) |>
+ unnest(cols = everything()) |>
+ summarize(
+  auc = survivalROC(Stime = time,
+                    status = status,
+                    marker = pred,
+                    predict.time = 3500,
+                    span = 0.25*n()^(-0.20)) |>
+   getElement('AUC')
+ )
+
+# C-stat: 0.81465
+aorsf_eval$auc
+
+}
 }
 \references{
 Breiman L. Random forests. \emph{Machine learning}. 2001 Oct;45(1):5-32.
