+++ R/orsf_attr.R
@@ -11,42 +11,42 @@
 #' @return varies depending on which get_ function is used.
 #'
 #' @noRd
-get_mtry             <- function(object) attr(object, 'mtry')
-get_n_obs            <- function(object) attr(object, 'n_obs')
-get_n_tree           <- function(object) attr(object, 'n_tree')
-get_names_y          <- function(object) attr(object, 'names_y')
-get_types_x          <- function(object) attr(object, 'types_x')
-get_n_events         <- function(object) attr(object, 'n_events')
-get_max_time         <- function(object) attr(object, 'max_time')
-get_unit_info        <- function(object) attr(object, 'unit_info')
-get_fctr_info        <- function(object) attr(object, 'fctr_info')
-get_n_leaves_mean    <- function(object) attr(object, 'n_leaves_mean')
-get_n_split          <- function(object) attr(object, 'n_split')
-get_leaf_min_events  <- function(object) attr(object, 'leaf_min_events')
-get_leaf_min_obs     <- function(object) attr(object, 'leaf_min_obs')
-get_split_min_events <- function(object) attr(object, 'split_min_events')
-get_split_min_obs    <- function(object) attr(object, 'split_min_obs')
-get_split_min_stat   <- function(object) attr(object, 'split_min_stat')
-get_cph_method       <- function(object) attr(object, 'cph_method')
-get_cph_eps          <- function(object) attr(object, 'cph_eps')
-get_cph_iter_max     <- function(object) attr(object, 'cph_iter_max')
-get_cph_do_scale     <- function(object) attr(object, 'cph_do_scale')
-get_net_alpha        <- function(object) attr(object, 'net_alpha')
-get_net_df_target    <- function(object) attr(object, 'net_df_target')
-get_numeric_bounds   <- function(object) attr(object, 'numeric_bounds')
-get_n_retry          <- function(object) attr(object, 'n_retry')
-get_f_oobag_eval     <- function(object) attr(object, 'f_oobag_eval')
-get_type_oobag_eval  <- function(object) attr(object, 'type_oobag_eval')
-get_oobag_pred       <- function(object) attr(object, 'oobag_pred')
-get_oobag_pred_horizon       <- function(object) attr(object, 'oobag_pred_horizon')
-get_oobag_eval_every <- function(object) attr(object, 'oobag_eval_every')
-get_importance       <- function(object) attr(object, 'importance')
-get_f_beta           <- function(object) attr(object, 'f_beta')
-get_orsf_type        <- function(object) attr(object, 'orsf_type')
-get_f_oobag_eval     <- function(object) attr(object, 'f_oobag_eval')
-get_type_oobag_eval  <- function(object) attr(object, 'type_oobag_eval')
-get_tree_seeds       <- function(object) attr(object, 'tree_seeds')
-get_weights_user     <- function(object) attr(object, 'weights_user')
+get_mtry               <- function(object) attr(object, 'mtry')
+get_n_obs              <- function(object) attr(object, 'n_obs')
+get_n_tree             <- function(object) attr(object, 'n_tree')
+get_names_y            <- function(object) attr(object, 'names_y')
+get_types_x            <- function(object) attr(object, 'types_x')
+get_n_events           <- function(object) attr(object, 'n_events')
+get_max_time           <- function(object) attr(object, 'max_time')
+get_unit_info          <- function(object) attr(object, 'unit_info')
+get_fctr_info          <- function(object) attr(object, 'fctr_info')
+get_n_leaves_mean      <- function(object) attr(object, 'n_leaves_mean')
+get_n_split            <- function(object) attr(object, 'n_split')
[…]
