### 2022-08-13 cbfb64e
codifying inputs/refs
