+++ R/orsf.R
@@ -1,3 +1,4 @@
+
 #' Oblique Random Survival Forest (ORSF)
 #'
 #' Fit an oblique random survival forest
@@ -52,17 +53,18 @@
 #'    criteria is met.
 #'
 #'  - [orsf_control_net] uses `glmnet` to identify linear combinations of
-#'    predictors, similar to [obliqueRSF][obliqueRSF::ORSF].
+#'    predictors, similar to Jaeger (2019).
 #'
 #'  - [orsf_control_custom] allows the user to apply their own function
 #'    to create linear combinations of predictors.
 #'
-#' @param weights (_numeric vector_) Optional. If given, this
-#'   input should be a vector with length equal to the number of rows in
-#'   `data`. The values in `weights` are treated like replication weights,
-#'   i.e., a weight value of 2 is the same thing as having 2 observations
-#'   in `data`, each containing a copy of the corresponding person's data.
-#'   __Use this input cautiously__, as `orsf` will count the number of
+#' @param weights (_numeric vector_) Optional. If given, this input should
+#'   have length equal to `nrow(data)`. Values in `weights` are treated like
+#'   replication weights, i.e., a value of 2 is the same thing as having 2
+#'   observations in `data`, each containing a copy of the corresponding
+#'   person's data.
+#'
+#'   *Use* `weights` *cautiously*, as `orsf` will count the number of
 #'   observations and events prior to growing a node for a tree, so higher
 #'   values in `weights` will lead to deeper trees.
 #'
@@ -114,18 +116,22 @@
 #'   is `oobag_eval_every = n_tree`, so that out-of-bag performance is
 #'   assessed once after growing all the trees.
 #'
-#' @param oobag_fun (_function_) A function to be measure the accuracy of
-#'   out-of-bag predictions. When `oobag_fun` = `NULL` (the default),
-#'   out-of-bag predictions are evaluated using Harrell's C-statistic.
-#'   For more details see the
-#'   [vignette](https://bcjaeger.github.io/aorsf/articles/oobag.html#user-supplied-out-of-bag-evaluation-functions)
-#'   on out-of-bag predictions:
+#' @param oobag_fun `r roxy_oobag_fun_header()` every `oobag_eval_every` trees.
+#' - `r roxy_oobag_fun_default()`
+#' - `r roxy_oobag_fun_user()`
+#'     + `r roxy_oobag_fun_inputs()`
+#'     + `r roxy_oobag_fun_ymat()`
+#'     + `r roxy_oobag_fun_svec()`
+#'     + `r roxy_oobag_fun_return()`
+#' For more details, see the out-of-bag [vignette](https://bcjaeger.github.io/aorsf/articles/oobag.html#user-supplied-out-of-bag-evaluation-functions).
 #'
-#' @param importance (_character_) Indicate method for variable importance:
-#'   - 'none': no variable importance is computed.
-#'   - 'anova': use the analysis of variance (ANOVA) method
-#'   - 'negate': compute negation importance
-#'   - 'permute': compute permutation importance
+#' @param importance `r roxy_importance_header()`
+#' - `r roxy_importance_none()`
+#' - `r roxy_importance_anova()`
+#' - `r roxy_importance_negate()`
+#' - `r roxy_importance_permute()`
+#'
+#' For details on these methods, see [orsf_vi].
 #'
 #' @param tree_seeds (_integer vector_) if specified, random seeds will be set
[…]
