+++ R/orsf_interaction.R
@@ -6,7 +6,7 @@
 #'
 #' @srrstats {G1.4} *documented with Roxygen*
 #'
-#' @inheritParams predict.aorsf
+#' @inheritParams predict.orsf_fit
 #'
 #' @param min_pairwise_obs (_integer_) minimum number of observations
 #'   where both variables were included in a linear combination together.
