+++ R/orsf_predict.R
@@ -1,24 +1,26 @@
 
 
-#' Predict using oblique random survival forest
+#' Compute predictions using ORSF
+#'
+#' Predicted risk or survival (someday also hazard or mortality)
+#'   from an ORSF model.
 #'
 #' @srrstats {G1.4} *documented with Roxygen*
 #' @srrstats {ML1.1} *using the terms 'train' and 'test'.*
 #' @srrstats {G2.0a} *specified expectations for length of `pred_horizon`. In general, inputs of length > 1 have the term 'vector' in their description, and inputs of length 1 just have the expected type.*
+#' @srrstats {G2.1a} explicit secondary documentation of expectations on data types of all vector inputs
 #' @srrstats {ML1.1} *The term 'new_data' are used instead of data_test. There are two reasons for this. First, I am making an effort to be consistent with tidymodels. Second, there is a possibility that users will use predict() without the intention of testing their model, e.g., for interpretation.*
 #'
-#' @param object (*orsf_fit*) an oblique random survival forest (ORSF; see [orsf]).
+#' @param object (*orsf_fit*) a trained oblique random survival forest
+#'   (see [orsf]).
 #'
-#' @param new_data (_data.frame_) data to compute predictions for. Must have
-#'   the same columns with equivalent types as the data used to train `object`.
-#'   Also, factors in `new_data` must not have levels that were not in the
-#'   data used to train `object`. Last, missing data are not supported.
+#' @param new_data a `r roxy_data_allowed()` to compute predictions in.
+#'   Missing data are not currently allowed
 #'
-#' @srrstats {G2.1a} explicit secondary documentation of expectations on data types of all vector inputs
-#'
-#' @param pred_horizon (_double_) a single time or a vector of times
-#'   indicating the time(s) that predicted risk or survival probabilities
-#'   will be computed at.
+#' @param pred_horizon (_double_) a value or vector indicating the time(s)
+#'   that predictions will be calibrated to. E.g., if you were predicting
+#'   risk of incident heart failure within the next 10 years, then
+#'   `pred_horizon = 10`
 #'
 #' @param pred_type (_character_) the type of predictions to return. Valid
 #'   options are
@@ -33,6 +35,10 @@
 #'   row `i` in `new_data`.
 #'
 #' @details
+#'
+#' `new_data` must have the same columns with equivalent types as the data
+#'   used to train `object`. Also, factors in `new_data` must not have levels
+#'   that were not in the data used to train `object`.
 #'
 #' `pred_horizon` values must not exceed the maximum follow-up time in
 #'   `object`'s training data. Also, `pred_horizon` values must be entered
