+++ R/orsf_control.R
@@ -231,26 +231,7 @@ orsf_control_net <- function(alpha = 1/2,
 #'
 #' @family orsf_control
 #'
-#' @examples
-#'
-#' # fit an oblique random survival forest using random coefficients to
-#' # generate linear combinations of predictor variables. First, define
-#' # a function that supplies the random coefficients:
-#'
-#' f <- function(x_node, y_node, w_node) { matrix(runif(ncol(x_node)), ncol=1) }
-#'
-#' # next, plug the function into orsf_control_custom(), which is in turn
-#' # passed into orsf():
-#'
-#' fit_rando <- orsf(pbc_orsf,
-#'                   Surv(time, status) ~ .,
-#'                   control = orsf_control_custom(beta_fun = f),
-#'                   n_tree = 500)
-#'
-#' # last, check the out-of-bag performance.
-#' # it's surprising how well the random approach works.
-#'
-#' fit_rando$eval_oobag
+#' @includeRmd Rmd/orsf_control_custom.Rmd
 
 orsf_control_custom <- function(beta_fun, ...){
 
@@ -265,6 +246,4 @@ orsf_control_custom <- function(beta_fun, ...){
 
 
 }
-
-
 
