+++ Rmd/orsf_control_custom.Rmd
@@ -1,0 +1,84 @@
+
+# examples
+
+## Random coefficients
+
+Want to fit extremely randomized oblique survival forests? Great! First, define a function that supplies the random coefficients:
+
+```{r}
+
+f_rando <- function(x_node, y_node, w_node){
+ matrix(runif(ncol(x_node)), ncol=1) 
+}
+
+```
+
+next, plug the function into orsf_control_custom(), which is in turn passed into orsf():
+
+```{r}
+
+library(aorsf)
+
+fit_rando <- orsf(pbc_orsf,
+                  Surv(time, status) ~ . - id,
+                  control = orsf_control_custom(beta_fun = f_rando),
+                  n_tree = 500)
+
+fit_rando
+
+```
+
+## Principal components
+
+Want to use principal component analysis to identify linear combinations of predictors? Follow the same steps as above, starting with the function to get coefficients:
+
+```{r}
+
+# use the second principal component to split the node
+
+f_pca <- function(x_node, y_node, w_node) { 
+ stats::prcomp(x_node, rank. = 2)$rotation[, 2, drop = FALSE]
+}
+
+```
+
+Then pass that function into `orsf_control_custom()`, which can be passed into `orsf()`
+
+```{r}
+
+fit_pca <- orsf(pbc_orsf,
+                Surv(time, status) ~ . - id,
+                control = orsf_control_custom(beta_fun = f_pca),
+                n_tree = 500)
+
+```
+
+## Evaluate
+
+How well do our two customized ORSFs do? Let's compute their indices of prediction accuracy based on out-of-bag predictions:
+
+```{r}
+
+library(riskRegression)
+library(survival)
+
+risk_preds <- list(rando = 1 - fit_rando$surv_oobag,
+                    pca = 1 - fit_pca$surv_oobag)
+
+sc <- Score(object = risk_preds, 
+            formula = Surv(time, status) ~ 1, 
+            data = pbc_orsf, 
+            summary = 'IPA',
+            times = fit_pca$pred_horizon)
+
+```
+
+The PCA ORSF does quite well!
+
+```{r}
+
+sc$Brier
+
+```
+
+
