+++ man/orsf_control_custom.Rd
@@ -30,27 +30,125 @@ an input for the \code{control} argument of \link{orsf}.
 \description{
 Custom ORSF control
 }
-\examples{
+\section{examples}{
+\subsection{Random coefficients}{
 
-# fit an oblique random survival forest using random coefficients to
-# generate linear combinations of predictor variables. First, define
-# a function that supplies the random coefficients:
+Want to fit extremely randomized oblique survival forests? Great! First,
+define a function that supplies the random coefficients:
 
-f <- function(x_node, y_node, w_node) { matrix(runif(ncol(x_node)), ncol=1) }
+\if{html}{\out{<div class="sourceCode r">}}\preformatted{f_rando <- function(x_node, y_node, w_node)\{
+ matrix(runif(ncol(x_node)), ncol=1) 
+\}
+}\if{html}{\out{</div>}}
 
-# next, plug the function into orsf_control_custom(), which is in turn
-# passed into orsf():
+next, plug the function into orsf_control_custom(), which is in turn
+passed into orsf():
+
+\if{html}{\out{<div class="sourceCode r">}}\preformatted{library(aorsf)
 
 fit_rando <- orsf(pbc_orsf,
-                  Surv(time, status) ~ .,
-                  control = orsf_control_custom(beta_fun = f),
+                  Surv(time, status) ~ . - id,
+                  control = orsf_control_custom(beta_fun = f_rando),
                   n_tree = 500)
 
-# last, check the out-of-bag performance.
-# it's surprising how well the random approach works.
+fit_rando
+}\if{html}{\out{</div>}}
 
-fit_rando$eval_oobag
+\if{html}{\out{<div class="sourceCode">}}\preformatted{## ---------- Oblique random survival forest
+## 
+##      Linear combinations: Custom user function
+##           N observations: 276
+##                 N events: 111
+##                  N trees: 500
+##       N predictors total: 17
+##    N predictors per node: 5
+##  Average leaves per tree: 21
+## Min observations in leaf: 5
+##       Min events in leaf: 1
+##           OOB stat value: 0.79
+##            OOB stat type: Harrell's C-statistic
+##      Variable importance: anova
+## 
+## -----------------------------------------
+}\if{html}{\out{</div>}}
 }
+
+\subsection{Principal components}{
+
+Want to use principal component analysis to identify linear combinations
+of predictors? Follow the same steps as above, starting with the
+function to get coefficients:
+
+\if{html}{\out{<div class="sourceCode r">}}\preformatted{# use the second principal component to split the node
+
+f_pca <- function(x_node, y_node, w_node) \{ 
+ stats::prcomp(x_node, rank. = 2)$rotation[, 2, drop = FALSE]
+\}
+}\if{html}{\out{</div>}}
+
+Then pass that function into \code{orsf_control_custom()}, which can be
+passed into \code{orsf()}
+
+\if{html}{\out{<div class="sourceCode r">}}\preformatted{fit_pca <- orsf(pbc_orsf,
+                Surv(time, status) ~ . - id,
+                control = orsf_control_custom(beta_fun = f_pca),
+                n_tree = 500)
+}\if{html}{\out{</div>}}
+}
+
+\subsection{Evaluate}{
+
+How well do our two customized ORSFs do? Let’s compute their indices of
+prediction accuracy based on out-of-bag predictions:
+
+\if{html}{\out{<div class="sourceCode r">}}\preformatted{library(riskRegression)
+}\if{html}{\out{</div>}}
+
[…]
