+++ README.md
@@ -11,7 +11,6 @@ public.](https://www.repostatus.org/badges/latest/wip.svg)](https://www.repostat
 [![Codecov test
 coverage](https://codecov.io/gh/bcjaeger/aorsf/branch/master/graph/badge.svg)](https://app.codecov.io/gh/bcjaeger/aorsf?branch=master)
 [![R-CMD-check](https://github.com/bcjaeger/aorsf/workflows/R-CMD-check/badge.svg)](https://github.com/bcjaeger/aorsf/actions)
-[![pkgcheck](https://github.com/bcjaeger/aorsf/workflows/pkgcheck/badge.svg)](https://github.com/bcjaeger/aorsf/actions?query=workflow%3Apkgcheck)
 [![Status at rOpenSci Software Peer
 Review](https://badges.ropensci.org/532_status.svg)](https://github.com/ropensci/software-review/issues/532)
 <!-- badges: end -->
@@ -56,7 +55,7 @@ fit <- orsf(data = pbc_orsf,
 print(fit)
 #> ---------- Oblique random survival forest
 #> 
-#>                ORSF type: Accelerated
+#>      Linear combinations: Accelerated
 #>           N observations: 276
 #>                 N events: 111
 #>                  N trees: 500
@@ -78,14 +77,14 @@ How about interpreting the fit?
 
     ``` r
     orsf_vi_negate(fit)
-    #>          bili           age       ascites       protime       albumin 
-    #>  0.0171910815  0.0149510315  0.0057303605  0.0052615128  0.0034903105 
-    #>         stage       spiders      platelet        copper         edema 
-    #>  0.0034382163  0.0033340279  0.0030214628  0.0027609919  0.0022425307 
-    #>           ast        hepato          trig           trt      alk.phos 
-    #>  0.0021358616  0.0006251302 -0.0002604709 -0.0005730360 -0.0013544488 
-    #>           sex          chol 
-    #> -0.0017191081 -0.0027088977
+    #>          bili           age       protime       spiders       ascites 
+    #>  0.0189101896  0.0156282559  0.0073973745  0.0051573244  0.0044280058 
+    #>        copper         stage           ast          trig           sex 
+    #>  0.0040633465  0.0034903105  0.0025526151  0.0025005209  0.0018753907 
+    #>         edema        hepato      platelet      alk.phos          chol 
+    #>  0.0013978607  0.0007293186 -0.0006772244 -0.0009376954 -0.0014065430 
+    #>           trt 
+    #> -0.0019274849
     ```
 
 -   use `orsf_pd_ice()` or `orsf_pd_summary()` for individual or
@@ -95,11 +94,11 @@ How about interpreting the fit?
     orsf_pd_summary(fit, pd_spec = list(bili = c(1:5)))
     #>     bili      mean        lwr      medn       upr
     #>    <int>     <num>      <num>     <num>     <num>
-    #> 1:     1 0.2365249 0.01101961 0.1321714 0.8670140
-    #> 2:     2 0.2871905 0.03974545 0.1744483 0.8973028
-    #> 3:     3 0.3427409 0.07308052 0.2560156 0.9234381
-    #> 4:     4 0.3940760 0.10322698 0.3167858 0.9324083
-    #> 5:     5 0.4348478 0.12844698 0.3807357 0.9380762
+    #> 1:     1 0.2356302 0.01203499 0.1233501 0.8728360
+    #> 2:     2 0.2872993 0.03491402 0.1809715 0.8968723
+    #> 3:     3 0.3428164 0.05744697 0.2435341 0.9165324
+    #> 4:     4 0.3911216 0.09348188 0.3115675 0.9336801
[…]
