+++ R/roxy.R
@@ -197,3 +197,18 @@ roxy_dots <- function(){
 }
 
 
+
+# variable importance descriptions ----------------------------------------
+
+roxy_vi_describe <- function(type){
+
+ switch(type,
+        'negate' = "Each variable is assessed separately by multiplying the variable's coefficients by -1 and then determining how much the model's performance changes. The worse the model's performance after negating coefficients for a given variable, the more important the variable.",
+        'permute' = "Each variable is assessed separately by randomly permuting the variable's values and then determining how much the model's performance changes. The worse the model's performance after permuting the values of a given variable, the more important the variable.",
+        'anova' = "A p-value is computed for each coefficient in each linear combination of variables in each decision tree. Importance for an individual predictor variable is the proportion of times a p-value for its coefficient is < 0.01.")
+
+}
+
+
+
+
