+++ README.Rmd
@@ -33,11 +33,11 @@ knitr::opts_chunk$set(
 
 ## Why aorsf?
 
-- Hundreds of times faster than `obliqueRSF` (see Jaeger, 2019).
+- Hundreds of times faster than `obliqueRSF`.^1^
 
-- accurate predictions for time-to-event outcomes.
+- Fast and accurate predictions for censored outcomes.^2^
 
-- negation importance, a novel technique to estimate variable importance for ORSFs.
+- negation importance, a novel technique to estimate variable importance for ORSFs.^2^
 
 - intuitive API with formula based interface.
 
@@ -51,24 +51,37 @@ You can install the development version of aorsf from [GitHub](https://github.co
 # install.packages("remotes")
 remotes::install_github("bcjaeger/aorsf")
 ```
+
 ## Example
 
-The `orsf()` function is used to fit ORSFs. Printing the output from `orsf()` will give some descriptive statistics about the ensemble.
+The `orsf()` function is used to fit ORSF ensembles: 
 
-```{r example}
+```{r}
 
 library(aorsf)
 
-fit <- orsf(data = pbc_orsf,
+fit <- orsf(data = pbc_orsf, 
             formula = Surv(time, status) ~ . - id)
+
+```
+
+The default routine to fit ORSF ensembles is the 'accelerated' ORSF - an algorithm based on Newton Raphson scoring that does very well in benchmarks of prediction accuracy and computational efficiency.^2^ In addition to the accelerated ORSF, `aorsf` can fit a broad range of ORSF ensembles (see ORSF CONTROL VIGNETTE (not yet written)). 
+
+### Inspect
+
+Printing the output from `orsf()` will give some descriptive statistics about the ensemble.
+
+```{r}
 
 print(fit)
 
 ```
 
-How about interpreting the fit? 
+### Variable importance
 
-- use `orsf_vi_negate()` and `orsf_vi_anova()` for variable importance
+The importance of individual variables can be estimated in three ways using `aorsf`:
+
+- **negation**: `r aorsf:::roxy_vi_describe('negate')`
 
   ```{r}
   
@@ -76,7 +89,31 @@ How about interpreting the fit?
   
   ```
 
-- use `orsf_pd_ice()` or `orsf_pd_summary()` for individual or aggregated partial dependence values. 
+- **permutation**: `r aorsf:::roxy_vi_describe('permute')`
+
+  ```{r}
+  
+  orsf_vi_permute(fit)
+  
+  ```
+  
+- **analysis of variance (ANOVA)^3^**: `r aorsf:::roxy_vi_describe('anova')`
+
+  ```{r}
+  
+  orsf_vi_anova(fit)
+  
+  ```
+
+You can also supply your own R function to estimate out-of-bag error when using negation or permutation importance (see [oob vignette](https://bcjaeger.github.io/aorsf/articles/oobag.html)). 
+
+### Partial dependence
+
+`aorsf` can generate individual conditional expectation (ICE) and partial dependence: 
+
+- ICE is the expected predicted value of an ORSF ensemble for an individual observation.
+
+- partial dependence is a multi-variable adjusted expected predicted value of an ORSF ensemble.
 
   ```{r}
   
@@ -105,9 +142,16 @@ Jaeger (2022) describes `aorsf` in detail, with a summary of the procedures used
 
 ## References
 
[…]
