+++ README.md
@@ -20,12 +20,12 @@ predictions with oblique random survival forests (ORSFs).
 
 ## Why aorsf?
 
--   Hundreds of times faster than `obliqueRSF` (see Jaeger, 2019).
+-   Hundreds of times faster than `obliqueRSF`.<sup>1</sup>
 
--   accurate predictions for time-to-event outcomes.
+-   Fast and accurate predictions for censored outcomes.<sup>2</sup>
 
 -   negation importance, a novel technique to estimate variable
-    importance for ORSFs.
+    importance for ORSFs.<sup>2</sup>
 
 -   intuitive API with formula based interface.
 
@@ -43,15 +43,28 @@ remotes::install_github("bcjaeger/aorsf")
 
 ## Example
 
-The `orsf()` function is used to fit ORSFs. Printing the output from
-`orsf()` will give some descriptive statistics about the ensemble.
+The `orsf()` function is used to fit ORSF ensembles:
 
 ``` r
 library(aorsf)
 
-fit <- orsf(data = pbc_orsf,
+fit <- orsf(data = pbc_orsf, 
             formula = Surv(time, status) ~ . - id)
+```
 
+The default routine to fit ORSF ensembles is the ‘accelerated’ ORSF - an
+algorithm based on Newton Raphson scoring that does very well in
+benchmarks of prediction accuracy and computational
+efficiency.<sup>2</sup> In addition to the accelerated ORSF, `aorsf` can
+fit a broad range of ORSF ensembles (see ORSF CONTROL VIGNETTE (not yet
+written)).
+
+### Inspect
+
+Printing the output from `orsf()` will give some descriptive statistics
+about the ensemble.
+
+``` r
 print(fit)
 #> ---------- Oblique random survival forest
 #> 
@@ -71,34 +84,86 @@ print(fit)
 #> -----------------------------------------
 ```
 
-How about interpreting the fit?
+### Variable importance
 
--   use `orsf_vi_negate()` and `orsf_vi_anova()` for variable importance
+The importance of individual variables can be estimated in three ways
+using `aorsf`:
+
+-   **negation**: Each variable is assessed separately by multiplying
+    the variable’s coefficients by -1 and then determining how much the
+    model’s performance changes. The worse the model’s performance after
+    negating coefficients for a given variable, the more important the
+    variable.
 
     ``` r
     orsf_vi_negate(fit)
-    #>          bili           age       protime       spiders       ascites 
-    #>  0.0189101896  0.0156282559  0.0073973745  0.0051573244  0.0044280058 
-    #>        copper         stage           ast          trig           sex 
-    #>  0.0040633465  0.0034903105  0.0025526151  0.0025005209  0.0018753907 
-    #>         edema        hepato      platelet      alk.phos          chol 
-    #>  0.0013978607  0.0007293186 -0.0006772244 -0.0009376954 -0.0014065430 
-    #>           trt 
-    #> -0.0019274849
+    #>           age          bili        copper         stage       protime 
+    #>  0.0151594082  0.0133361117  0.0065638675  0.0060950198  0.0060429256 
+    #>       albumin           sex       spiders       ascites           ast 
+    #>  0.0059387372  0.0058345489  0.0049489477  0.0048447593  0.0035424047 
+    #>         edema        hepato          chol          trig           trt 
+    #>  0.0019498110  0.0006772244 -0.0014065430 -0.0015107314 -0.0019795791 
[…]
