+++ R/roxy.R
@@ -211,4 +211,23 @@ roxy_vi_describe <- function(type){
 
 
 
+# partial dependence ------------------------------------------------------
+
+roxy_pd_oob_explain <- function(label){
+
+ paste0("You can compute ", label, " ",
+        "two ways using a random forest: \n",
+        "- using out-of-bag predictions for the training data\n",
+        "- using predictions for a new set of data\n\n",
+        "See examples for more details")
+
+}
+
+roxy_pd_explain <- function(){
+ "Partial dependence shows the expected prediction from a model as a function of a single predictor or multiple predictors. The expectation is marginalized over the values of all other predictors, giving something like a multivariable adjusted estimate of the model's prediction."
+}
+
+roxy_ice_explain <- function(){
+ "Unlike partial dependence, which shows the expected prediction as a function of one or multiple predictors, individual conditional expectations shows the prediction for an individual observation as a function of a predictor."
+}
 
