+++ R/orsf_pd.R
@@ -6,19 +6,14 @@
 #' @srrstats {G2.1a} *explicit secondary documentation of expectations on data types of all vector inputs*
 #' @srrstats {G2.0a} *documenting length by indicating these inputs are vectors and that they must have the same length as the other.*
 #'
-#' Compute partial dependence using oblique random survival forests. A
-#'  [vignette](https://bcjaeger.github.io/aorsf/articles/pd.html)
-#'  covers the basics of partial dependence with extended examples.
+#' Compute partial dependence for an ORSF model.
+#' `r roxy_pd_explain()`
+#'
+#' `r roxy_pd_oob_explain('partial dependence')`
 #'
 #' @inheritParams predict.orsf_fit
 #'
-#' @param pd_data (_data frame_) that will be used to compute partial
-#'   dependence. If `NULL`, then the training data of `object` will be
-#'   used. If the training data were not attached to `object`
-#'   (see `attach_data` input in [orsf]), an error will be triggered.
-#'
-#'
-#' @param pd_spec (_named list_ or _data.frame_).
+#' @param pd_spec (*named list* or _data.frame_).
 #'
 #'  -  If `pd_spec` is a named list,
 #'   Each item in the list should be a vector of values that will be used as
@@ -48,11 +43,6 @@
 #'   value in `prob_values` should be labelled as in summarized outputs.
 #'   `prob_labels` should have the same length as `prob_values`.
 #'
-#' @param oobag (_logical_) if `TRUE`, then partial dependence will be
-#'   computed using the out of bag training data. You should set
-#'   `oobag = TRUE` if you are computing partial dependence using the
-#'   training data for `object`.
-#'
 #' @param boundary_checks (_logical_) if `TRUE`, `pd_spec` will be vetted
 #'  to make sure the requested values are between the 10th and 90th
 #'  percentile in the object's training data. If `FALSE`, these checks are
@@ -60,164 +50,139 @@
 #'
 #' @param ... `r roxy_dots()`
 #'
-#' @return a `data.table` containing
-#'  - `orsf_ice`: individual conditional expectation (ICE) values.
-#'  - `orsf_pd`: summarized ICE values (i.e., partial dependence).
+#' @return a [data.table][data.table::data.table-package] containing
+#'   partial dependence values for the specified variable(s) at the
+#'   specified prediction horizon(s).
+#'
+#' @includeRmd Rmd/orsf_pd_examples.Rmd
 #'
 #' @export
 #'
-#' @examples
 #'
-#' fit <- orsf(pbc_orsf, Surv(time, status) ~ . - id)
-#'
-#' orsf_pd(fit, pd_spec = list(bili = c(1,2,3,4,5,6)), pred_horizon = 1000)
-#'
-#' # more points for a plot
-#' pd_spec <- list(bili = seq(1, 6, length.out = 20))
-#' data_ice <- orsf_ice(fit, pd_spec = pd_spec, pred_horizon = c(1000))
-#'
-#' head(data_ice)
-#'
-#' library(ggplot2)
-#'
-#' ggplot(data_ice) +
-#'  aes(x = bili, y = pred, group = id_row) +
-#'  geom_line(alpha = 0.4, color = 'grey') +
-#'  geom_smooth(aes(group = 1), color = 'black', se = FALSE) +
-#'  theme_bw() +
-#'  theme(panel.grid = element_blank())
-
-orsf_pd <- function(object,
-                            pd_data = NULL,
-                            pd_spec,
-                            pred_horizon = NULL,
-                            pred_type = 'risk',
-                            expand_grid = TRUE,
-                            prob_values = c(0.025, 0.50, 0.975),
-                            prob_labels = c('lwr', 'medn', 'upr'),
[…]

[…]
