+++ README.md
@@ -41,7 +41,29 @@ You can install the development version of aorsf from
 remotes::install_github("bcjaeger/aorsf")
 ```
 
-## Example
+## What is an oblique decision tree?
+
+Decision trees are developed by splitting a set of training data into
+two new subsets, with the goal of having more similarity within the new
+subsets than between them. The splitting process is repeated on
+resulting subsets of data until a stopping criterion is met.
+
+When the new subsets of data are formed based on a single predictor, the
+decision tree is said to be *axis-based* because the splits of the data
+appear perpendicular to the axis of the predictor. When linear
+combinations of variables are used instead of a single variable, the
+tree is *oblique* because the splits of the data are neither parallel
+nor at a right angle to the axis.
+
+**Figure**: Decision trees for classification with axis-based splitting
+(left) and oblique splitting (right). Cases are orange squares; controls
+are purple circles. Both trees partition the predictor space defined by
+variables X1 and X2, but the oblique splits do a better job of
+separating the two classes.
+
+<img src="man/figures/tree_axis_v_oblique.png" width="100%" />
+
+## Examples
 
 The `orsf()` function can fit several types of ORSF ensembles. My
 personal favorite is the accelerated ORSF because it has a great
@@ -51,8 +73,16 @@ combination of prediction accuracy and computational efficiency (see
 ``` r
 library(aorsf)
 
-fit <- orsf(data = pbc_orsf, 
-            formula = Surv(time, status) ~ . - id)
+set.seed(329730)
+
+index_train <- sample(nrow(pbc_orsf), 150) 
+
+pbc_orsf_train <- pbc_orsf[index_train, ]
+pbc_orsf_test <- pbc_orsf[-index_train, ]
+
+fit <- orsf(data = pbc_orsf_train, 
+            formula = Surv(time, status) ~ . - id,
+            oobag_pred_horizon = 365.25 * 5)
 ```
 
 ORSF CONTROL VIGNETTE (not yet written).
@@ -63,19 +93,19 @@ Printing the output from `orsf()` will give some information and
 descriptive statistics about the ensemble.
 
 ``` r
-print(fit)
+fit
 #> ---------- Oblique random survival forest
 #> 
 #>      Linear combinations: Accelerated
-#>           N observations: 276
-#>                 N events: 111
+#>           N observations: 150
+#>                 N events: 52
 #>                  N trees: 500
 #>       N predictors total: 17
 #>    N predictors per node: 5
-#>  Average leaves per tree: 24
+#>  Average leaves per tree: 12
 #> Min observations in leaf: 5
 #>       Min events in leaf: 1
-#>           OOB stat value: 0.84
+#>           OOB stat value: 0.83
 #>            OOB stat type: Harrell's C-statistic
 #>      Variable importance: anova
 #> 
@@ -91,7 +121,7 @@ for a description of each line in the printed output.
 The importance of individual variables can be estimated in three ways
 using `aorsf`:
 
--   **negation<sup>2</sup>**: Each variable is assessed separately by
+-   **negation**<sup>2</sup>: Each variable is assessed separately by
     multiplying the variable’s coefficients by -1 and then determining
     how much the model’s performance changes. The worse the model’s
     performance after negating coefficients for a given variable, the
@@ -99,14 +129,12 @@ using `aorsf`:
 
     ``` r
     orsf_vi_negate(fit)
[…]
