+++ vignettes/pd.Rmd
@@ -32,49 +32,6 @@ docs_ice <- c("D:/aorsf/Rmd/orsf_ice_content_1.Rmd",
 `r aorsf:::roxy_pd_explain()`
 
 
-Begin by fitting an ORSF ensemble:
-
-```{r, child = 'orsf-fit-accelerated.Rmd'}
-
-```
-
-## Partial dependence two ways
-
-You can compute partial dependence two ways using a random forest:
-
-- using out-of-bag predictions for the training data
-
-  ```{r}
-  
-  pd_train <- orsf_pd_oob(fit, pd_spec = list(bili = 1:5))
-  
-  pd_train
-  
-  ```
-
-- using predictions for a new set of data
-
-  ```{r}
-  
-  pd_test <- orsf_pd_new(fit, 
-                         new_data = pbc_orsf_test, 
-                         pd_spec = list(bili = 1:5))
-  
-  pd_test
-  
-  ```
-
-Which should you use? Both!
-
-- out-of-bag partial dependence indicates relationships that the model has learned during training. This is helpful if your goal is to interpret the model.
-
-- partial dependence for new data shows how the model predicts outcomes for observations it has not seen. This is helpful if you want to test your model's reliability or fairness.
-
-A reliable model should give roughly the same partial dependence values in its training and testing data, given the size of the training and testing data are both sufficient. 
-
-
-
-
 ## Individual conditional expectations
 
 `r aorsf:::roxy_ice_explain()`
