+++ man/orsf.Rd
@@ -258,74 +258,220 @@ A user should impute missing values using an R package with that purpose,
 such as \code{recipes} or \code{mlr3pipelines}.
 }
 
-\examples{
+\section{Examples}{
+the aorsf package is designed to fit into standard modeling analyses,
+tidymodels workflows, and mlr3 pipelines.
+\subsection{standard modeling}{
 
-# standard workflow for model development: fit and interpret
-
-fit <- orsf(pbc_orsf, formula = Surv(time, status) ~ . - id)
-
-print(fit)
-
-# more specific parameters
-
-fit_custom <- orsf(pbc_orsf,
-                   formula = time + status ~ . - id,
-                   mtry = 2,
-                   oobag_pred_horizon = 4000,
-                   oobag_eval_every = 50)
-
-# 10 oobag error values are computed  b/c oob error
-# is assessed every 50 trees (50, 100, ..., 500)
-fit_custom$eval_oobag$stat_values
-
-# make your own oobag function:
-# (this is an R implementation of Harrell's C
-# stat, which is what orsf uses by default)
-
-oobag_c_harrell <- function(y_mat, s_vec){
- time = y_mat[, 1]
- status = y_mat[, 2]
- events = which(status == 1)
- k = nrow(y_mat)
- total <- 0
- concordant <- 0
- for(i in events){
-  if(i+1 <= k){
-   for(j in seq(i+1, k)){
-    if(time[j] > time[i]){
-     total <- total + 1
-     if(s_vec[j] > s_vec[i]){
-      concordant <- concordant + 1
-     } else if (s_vec[j] == s_vec[i]){
-      concordant <- concordant + 0.5
-     }
-    }
-   }
-  }
- }
- concordant / total
+FILL
 }
 
-# tree_seeds lets you grow each tree with the seed it corresponds to
-# (nice way to make sure my R oobag function gives the same answer
-# as the internal C oobag function)
+\subsection{tidymodels}{
 
-fit_custom_oobag <- orsf(pbc_orsf,
-                         formula = Surv(time, status) ~ . - id,
-                         oobag_fun = oobag_c_harrell,
-                         n_tree = 10,
-                         tree_seeds = 1:10)
+\if{html}{\out{<div class="sourceCode r">}}\preformatted{set.seed(329730)
 
-fit_standard_oobag <- orsf(pbc_orsf,
-                           formula = Surv(time, status) ~ . - id,
-                           n_tree = 10,
-                           tree_seeds = 1:10)
+library(aorsf)
+library(survival)
+library(tidymodels)
+library(tidyverse)
+library(randomForestSRC)
+library(ranger)
+library(riskRegression)
+}\if{html}{\out{</div>}}
+\subsection{Comparing learners}{
 
-fit_standard_oobag$eval_oobag$stat_values
+Start with a recipe to pre-process data
 
-fit_custom_oobag$eval_oobag$stat_values
+\if{html}{\out{<div class="sourceCode r">}}\preformatted{imputer <- recipe(pbc_orsf, formula = time + status ~ .) \%>\% 
+ step_impute_mean(all_numeric_predictors()) \%>\%
+ step_impute_mode(all_nominal_predictors()) 
+}\if{html}{\out{</div>}}
+
[…]
