+++ Rmd/orsf_pd_content_1.Rmd
@@ -5,9 +5,19 @@ Begin by fitting an ORSF ensemble:
 
 ```
 
-## Partial dependence two ways
+## Three ways to compute PD and ICE
 
-You can compute partial dependence two ways using a random forest:
+You can compute partial dependence and ICE three ways with `aorsf`: 
+
+- using in-bag predictions for the training data
+
+  ```{r}
+  
+  pd_train <- orsf_pd_inb(fit, pd_spec = list(bili = 1:5))
+  
+  pd_train
+  
+  ```
 
 - using out-of-bag predictions for the training data
 
@@ -31,10 +41,9 @@ You can compute partial dependence two ways using a random forest:
   
   ```
 
-Which should you use? Both!
+- in-bag partial dependence indicates relationships that the model has learned during training. This is helpful if your goal is to interpret the model.
 
-- out-of-bag partial dependence indicates relationships that the model has learned during training. This is helpful if your goal is to interpret the model.
+- out-of-bag partial dependence indicates relationships that the model has learned during training but using the out-of-bag data simulates application of the model to new data. if you want to test your model's reliability or fairness in new data but you don't have access to a large testing set.
 
-- partial dependence for new data shows how the model predicts outcomes for observations it has not seen. This is helpful if you want to test your model's reliability or fairness.
+- new data partial dependence shows how the model predicts outcomes for observations it has not seen. This is helpful if you want to test your model's reliability or fairness.
 
-A reliable model should give roughly the same partial dependence values in its training and testing data, given the size of the training and testing data are both sufficient. 
