+++ vignettes/pd.Rmd
@@ -1,5 +1,5 @@
 ---
-title: "Partial dependence and ICE curves with ORSF"
+title: "PD and ICE curves with ORSF"
 description: >
   Learn how to interpret ORSF models using partial dependence.
 output: rmarkdown::html_vignette
@@ -18,24 +18,21 @@ knitr::opts_chunk$set(
 )
 
 library(aorsf)
-
-docs_pd <- c("D:/aorsf/Rmd/orsf_pd_content_1.Rmd",
-             "D:/aorsf/Rmd/orsf_pd_content_2.Rmd")
-
-docs_ice <- c("D:/aorsf/Rmd/orsf_ice_content_1.Rmd",
-              "D:/aorsf/Rmd/orsf_ice_content_2.Rmd")
+library(ggplot2)
 
 ```
 
-## Partial dependence
+## Partial dependence (PD)
 
 `r aorsf:::roxy_pd_explain()`
 
-Begin by fitting an ORSF ensemble.
+Begin by fitting an ORSF ensemble. Set a prediction horizon of 5 years when we fit the ensemble so that any `aorsf` function that we pass this ensemble to will assume we want to compute predictions at 5 years.
 
 ```{r}
 
 library(aorsf)
+
+pred_horizon <- 365.25 * 5
 
 set.seed(329730)
 
@@ -46,23 +43,34 @@ pbc_orsf_test <- pbc_orsf[-index_train, ]
 
 fit <- orsf(data = pbc_orsf_train, 
             formula = Surv(time, status) ~ . - id,
-            oobag_pred_horizon = 365.25 * 5)
+            oobag_pred_horizon = pred_horizon)
 
 fit
 
 ```
 
-## Partial dependence two ways
 
-You can compute partial dependence two ways using a random forest:
+## Three ways to compute PD
+
+You can compute PD three ways with `aorsf`: 
+
+- using in-bag predictions for the training data
+
+  ```{r}
+  
+  pd_inb <- orsf_pd_inb(fit, pd_spec = list(bili = 1:5))
+  
+  pd_inb
+  
+  ```
 
 - using out-of-bag predictions for the training data
 
   ```{r}
   
-  pd_train <- orsf_pd_oob(fit, pd_spec = list(bili = 1:5))
+  pd_oob <- orsf_pd_oob(fit, pd_spec = list(bili = 1:5))
   
-  pd_train
+  pd_oob
   
   ```
 
@@ -78,28 +86,109 @@ You can compute partial dependence two ways using a random forest:
   
   ```
 
-Which should you use? Both!
+- in-bag PD indicates relationships that the model has learned during training. This is helpful if your goal is to interpret the model.
 
-- out-of-bag partial dependence indicates relationships that the model has learned during training. This is helpful if your goal is to interpret the model.
+- out-of-bag PD indicates relationships that the model has learned during training but using the out-of-bag data simulates application of the model to new data. if you want to test your model's reliability or fairness in new data but you don't have access to a large testing set.
 
-- partial dependence for new data shows how the model predicts outcomes for observations it has not seen. This is helpful if you want to test your model's reliability or fairness.
+- new data PD shows how the model predicts outcomes for observations it has not seen. This is helpful if you want to test your model's reliability or fairness.
 
-A reliable model should give roughly the same partial dependence values in its training and testing data, given the size of the training and testing data are both sufficient. 
-
-## Partial dependence for multiple variables
-
-Let's re-fit our ORSF to all available data before proceeding.
+Let's re-fit our ORSF to all available data before proceeding to the next sections.
 
 ```{r}
 
 set.seed(329730)
 
-fit <- orsf(pbc_orsf, Surv(time, status) ~ . -id)
[…]
