+++ NAMESPACE
@@ -9,8 +9,10 @@ export(orsf_control_cph)
 export(orsf_control_custom)
 export(orsf_control_fast)
 export(orsf_control_net)
+export(orsf_ice_inb)
 export(orsf_ice_new)
 export(orsf_ice_oob)
+export(orsf_pd_inb)
 export(orsf_pd_new)
 export(orsf_pd_oob)
 export(orsf_scale_cph)

[…]
