+++ R/check.R
@@ -1023,7 +1023,7 @@ check_orsf_inputs <- function(data = NULL,
 #' @noRd
 #'
 check_pd_inputs <- function(object,
-                            pd_spec = NULL,
+                            pred_spec = NULL,
                             expand_grid = NULL,
                             prob_values = NULL,
                             prob_labels = NULL,
@@ -1049,30 +1049,30 @@ check_pd_inputs <- function(object,
 
  }
 
- if(!is.null(pd_spec)){
+ if(!is.null(pred_spec)){
 
-  if(is_empty(pd_spec)){
-   stop("pd_spec is empty", call. = FALSE)
+  if(is_empty(pred_spec)){
+   stop("pred_spec is empty", call. = FALSE)
   }
 
-  if(is_empty(names(pd_spec))){
-   stop("pd_spec is unnamed", call. = FALSE)
+  if(is_empty(names(pred_spec))){
+   stop("pred_spec is unnamed", call. = FALSE)
   }
 
-  bad_name_index <- which(is.na(match(names(pd_spec), get_names_x(object))))
+  bad_name_index <- which(is.na(match(names(pred_spec), get_names_x(object))))
 
   if(!is_empty(bad_name_index)){
 
-   bad_names <- names(pd_spec)[bad_name_index]
+   bad_names <- names(pred_spec)[bad_name_index]
 
-   stop("variables in pd_spec are not recognized as predictors in object: ",
+   stop("variables in pred_spec are not recognized as predictors in object: ",
         paste_collapse(bad_names, last = ' and '),
         call. = FALSE)
 
   }
 
   numeric_bounds <- get_numeric_bounds(object)
-  numeric_names <- intersect(colnames(numeric_bounds), names(pd_spec))
+  numeric_names <- intersect(colnames(numeric_bounds), names(pred_spec))
 
   if(is.null(boundary_checks)) boundary_checks <- TRUE
 
@@ -1080,15 +1080,15 @@ check_pd_inputs <- function(object,
 
    for(.name in numeric_names){
 
-    vals_above_stop <- which(pd_spec[[.name]] > numeric_bounds['90%', .name])
-    vals_below_stop <- which(pd_spec[[.name]] < numeric_bounds['10%', .name])
+    vals_above_stop <- which(pred_spec[[.name]] > numeric_bounds['90%', .name])
+    vals_below_stop <- which(pred_spec[[.name]] < numeric_bounds['10%', .name])
 
     boundary_error <- FALSE
     vals_above_list <- vals_below_list <- " "
 
     if(!is_empty(vals_above_stop)){
      vals_above_list <- paste_collapse(
-      round_magnitude(pd_spec[[.name]][vals_above_stop]),
+      round_magnitude(pred_spec[[.name]][vals_above_stop]),
       last = ' and '
      )
 
@@ -1101,7 +1101,7 @@ check_pd_inputs <- function(object,
     if(!is_empty(vals_below_stop)){
 
      vals_below_list <- paste_collapse(
-      round_magnitude(pd_spec[[.name]][vals_below_stop]),
+      round_magnitude(pred_spec[[.name]][vals_below_stop]),
       last = ' and '
      )
 
@@ -1115,12 +1115,12 @@ check_pd_inputs <- function(object,
     if(boundary_error)
      stop("Some values for ",
           .name,
-          " in pd_spec are above",
+          " in pred_spec are above",
           vals_above_list,
           "or below",
           vals_below_list,
           "90th or 10th percentiles in training data.",
-          " Change pd_spec or set boundary_checks = FALSE",
+          " Change pred_spec or set boundary_checks = FALSE",
           " to prevent this error",
           call. = FALSE)
 
