+++ R/orsf_pd.R
@@ -13,22 +13,22 @@
 #'
 #' @inheritParams predict.orsf_fit
 #'
-#' @param pd_spec (*named list* or _data.frame_).
+#' @param pred_spec (*named list* or _data.frame_).
 #'
-#'  -  If `pd_spec` is a named list,
+#'  -  If `pred_spec` is a named list,
 #'   Each item in the list should be a vector of values that will be used as
 #'   points in the partial dependence function. The name of each item in the
 #'   list should indicate which variable will be modified to take the
 #'   corresponding values.
 #'
-#'  - If `pd_spec` is a `data.frame`, columns will
+#'  - If `pred_spec` is a `data.frame`, columns will
 #'   indicate variable names, values will indicate variable values, and
 #'   partial dependence will be computed using the inputs on each row.
 #'
 #' @param expand_grid (_logical_) if `TRUE`, partial dependence will be
-#'   computed at all possible combinations of inputs in `pd_spec`. If
+#'   computed at all possible combinations of inputs in `pred_spec`. If
 #'   `FALSE`, partial dependence will be computed for each variable
-#'   in `pd_spec`, separately.
+#'   in `pred_spec`, separately.
 #'
 #'
 #' @param prob_values (_numeric_) a vector of values between 0 and 1,
@@ -36,14 +36,14 @@
 #'   dependence values at each set of inputs. `prob_values` should
 #'   have the same length as `prob_labels`. The quantiles are calculated
 #'   based on predictions from `object` at each set of values indicated
-#'   by `pd_spec`.
+#'   by `pred_spec`.
 #'
 #' @param prob_labels (_character_) a vector of labels with the same length
 #'   as `prob_values`, with each label indicating what the corresponding
 #'   value in `prob_values` should be labelled as in summarized outputs.
 #'   `prob_labels` should have the same length as `prob_values`.
 #'
-#' @param boundary_checks (_logical_) if `TRUE`, `pd_spec` will be vetted
+#' @param boundary_checks (_logical_) if `TRUE`, `pred_spec` will be vetted
 #'  to make sure the requested values are between the 10th and 90th
 #'  percentile in the object's training data. If `FALSE`, these checks are
 #'  skipped.
@@ -60,7 +60,7 @@
 #'
 #'
 orsf_pd_oob <- function(object,
-                        pd_spec,
+                        pred_spec,
                         pred_horizon = NULL,
                         pred_type = 'risk',
                         expand_grid = TRUE,
@@ -72,7 +72,7 @@ orsf_pd_oob <- function(object,
  check_dots(list(...), orsf_pd_oob)
 
  orsf_pred_dependence(object = object,
-                      pd_spec = pd_spec,
+                      pred_spec = pred_spec,
                       pd_data = NULL,
                       pred_horizon = pred_horizon,
                       pred_type = pred_type,
@@ -88,7 +88,7 @@ orsf_pd_oob <- function(object,
 #' @rdname orsf_pd_oob
 #' @export
 orsf_pd_inb <- function(object,
-                        pd_spec,
+                        pred_spec,
                         pred_horizon = NULL,
                         pred_type = 'risk',
                         expand_grid = TRUE,
@@ -105,7 +105,7 @@ orsf_pd_inb <- function(object,
        "running orsf()?", call. = FALSE)
 
  orsf_pred_dependence(object = object,
-                      pd_spec = pd_spec,
+                      pred_spec = pred_spec,
                       pd_data = object$data,
                       pred_horizon = pred_horizon,
                       pred_type = pred_type,
[…]

[…]
