+++ man/orsf_ice_oob.Rd
@@ -8,7 +8,7 @@
 \usage{
 orsf_ice_oob(
   object,
-  pd_spec,
+  pred_spec,
   pred_horizon = NULL,
   pred_type = "risk",
   expand_grid = TRUE,
@@ -18,7 +18,7 @@ orsf_ice_oob(
 
 orsf_ice_inb(
   object,
-  pd_spec,
+  pred_spec,
   pred_horizon = NULL,
   pred_type = "risk",
   expand_grid = TRUE,
@@ -28,7 +28,7 @@ orsf_ice_inb(
 
 orsf_ice_new(
   object,
-  pd_spec,
+  pred_spec,
   new_data,
   pred_horizon = NULL,
   pred_type = "risk",
@@ -41,14 +41,14 @@ orsf_ice_new(
 \item{object}{(\emph{orsf_fit}) a trained oblique random survival forest
 (see \link{orsf}).}
 
-\item{pd_spec}{(\emph{named list} or \emph{data.frame}).
+\item{pred_spec}{(\emph{named list} or \emph{data.frame}).
 \itemize{
-\item If \code{pd_spec} is a named list,
+\item If \code{pred_spec} is a named list,
 Each item in the list should be a vector of values that will be used as
 points in the partial dependence function. The name of each item in the
 list should indicate which variable will be modified to take the
 corresponding values.
-\item If \code{pd_spec} is a \code{data.frame}, columns will
+\item If \code{pred_spec} is a \code{data.frame}, columns will
 indicate variable names, values will indicate variable values, and
 partial dependence will be computed using the inputs on each row.
 }}
@@ -66,11 +66,11 @@ options are
 }}
 
 \item{expand_grid}{(\emph{logical}) if \code{TRUE}, partial dependence will be
-computed at all possible combinations of inputs in \code{pd_spec}. If
+computed at all possible combinations of inputs in \code{pred_spec}. If
 \code{FALSE}, partial dependence will be computed for each variable
-in \code{pd_spec}, separately.}
+in \code{pred_spec}, separately.}
 
-\item{boundary_checks}{(\emph{logical}) if \code{TRUE}, \code{pd_spec} will be vetted
+\item{boundary_checks}{(\emph{logical}) if \code{TRUE}, \code{pred_spec} will be vetted
 to make sure the requested values are between the 10th and 90th
 percentile in the object's training data. If \code{FALSE}, these checks are
 skipped.}
@@ -121,9 +121,9 @@ random forest:
 \itemize{
 \item using out-of-bag predictions for the training data
 
-\if{html}{\out{<div class="sourceCode r">}}\preformatted{pd_spec <- list(bili = seq(1, 6, length.out = 10))
+\if{html}{\out{<div class="sourceCode r">}}\preformatted{pred_spec <- list(bili = seq(1, 6, length.out = 10))
 
-ice_train <- orsf_ice_oob(fit, pd_spec)
+ice_train <- orsf_ice_oob(fit, pred_spec)
 
 ice_train
 }\if{html}{\out{</div>}}
@@ -146,7 +146,7 @@ ice_train
 
 \if{html}{\out{<div class="sourceCode r">}}\preformatted{ice_test <- orsf_ice_new(fit, 
                          new_data = pbc_orsf_test, 
-                         pd_spec = pd_spec)
+                         pred_spec = pred_spec)
 
 ice_test
 }\if{html}{\out{</div>}}
