+++ Rmd/orsf_control_cph_examples.Rmd
@@ -1,0 +1,16 @@
+
+# Examples
+
+
+```{r, include=FALSE, echo=FALSE}
+
+docs <- c(
+ "orsf_control_cph_content_1.Rmd"
+)
+
+```
+
+```{r child = docs}
+
+```
+

[…]
