+++ README.Rmd
@@ -26,6 +26,7 @@ knitr::opts_chunk$set(
   [![Codecov test coverage](https://codecov.io/gh/bcjaeger/aorsf/branch/master/graph/badge.svg)](https://app.codecov.io/gh/bcjaeger/aorsf?branch=master)
 [![R-CMD-check](https://github.com/bcjaeger/aorsf/workflows/R-CMD-check/badge.svg)](https://github.com/bcjaeger/aorsf/actions)
 [![Status at rOpenSci Software Peer Review](https://badges.ropensci.org/532_status.svg)](https://github.com/ropensci/software-review/issues/532)
+<a href="https://joss.theoj.org/papers/414871f081cd8449007d671a7f7f7c3a"><img src="https://joss.theoj.org/papers/414871f081cd8449007d671a7f7f7c3a/status.svg"></a>
 <!-- badges: end -->
 
 `aorsf` provides optimized software to fit, interpret, and make predictions with oblique random survival forests (ORSFs).  
@@ -74,8 +75,6 @@ The `orsf()` function can fit several types of ORSF ensembles. My personal favor
 
 ```
 
-ORSF CONTROL VIGNETTE (not yet written). 
-
 ### Inspect
 
 Printing the output from `orsf()` will give some information and descriptive statistics about the ensemble.
@@ -86,7 +85,9 @@ fit
 
 ```
 
-See [`print.orsf_fit`](https://bcjaeger.github.io/aorsf/reference/print.orsf_fit.html) for a description of each line in the printed output.
+- See [print.orsf_fit](https://bcjaeger.github.io/aorsf/reference/print.orsf_fit.html) for a description of each line in the printed output.
+
+- See [orsf examples](https://bcjaeger.github.io/aorsf/reference/orsf.html#examples) for more details on controlling ORSF ensemble fits and using them in prediction modeling workflows.
 
 ### Variable importance
 
@@ -116,52 +117,36 @@ The importance of individual variables can be estimated in three ways using `aor
   
   ```
 
-You can supply your own R function to estimate out-of-bag error when using negation or permutation importance (see [oob vignette](https://bcjaeger.github.io/aorsf/articles/oobag.html)). 
+You can supply your own R function to estimate out-of-bag error when using negation or permutation importance. This feature is experimental and may be changed in the future (see [oob vignette](https://bcjaeger.github.io/aorsf/articles/oobag.html))
 
-### Predict
+### Partial dependence (PD) 
 
+`r aorsf:::roxy_pd_explain()`
 
+For more on PD, see the [vignette](https://bcjaeger.github.io/aorsf/articles/pd.html)
 
-### Partial dependence
+### Individual conditional expectations (ICE)
 
-`aorsf` can generate individual conditional expectation (ICE) and partial dependence (PD): 
-
-- ICE shows how the predicted value for an observation changes when a predictor changes. `orsf_ice()`, the function to compute ICE values in `aorsf`, returns ICE for all observations in the data you supply it. (If no data are supplied, the ORSF's training data are used)
-
-  ```{r}
-  
-  # ICE values:
-  # predicted risk for a single observation with respect to bili.
-  
-  orsf_ice_oob(fit, pred_spec = list(bili = c(1:5)))[id_row == 2]
-  
-  ```
-
-- PD shows how the **expected** predicted value changes when a predictor changes. PD estimates the relationship between the model's prediction and a predictor, adjusting for other predictors.
-
-  ```{r}
-  
-  orsf_pd_oob(fit, pred_spec = list(bili = c(1:5)))
-  
-  ```
+`r aorsf:::roxy_ice_explain()`
 
 
[…]

[…]
