### 2022-08-16 3ddba01
add hazard to pred matrix
