+++ src/orsf.cpp
@@ -49,6 +49,7 @@ double
  observed,
  expected,
  V,
+ pred_t0,
  leaf_min_obs,
  leaf_min_events,
  split_min_events,
@@ -68,7 +69,11 @@ int
  net_df_target,
  oobag_eval_every;
 
-char type_beta, type_oobag_eval, oobag_importance_type;
+char
+ type_beta,
+ type_oobag_eval,
+ oobag_importance_type,
+ pred_type_dflt = 'S';
 
 // armadillo unsigned integers
 uword
@@ -92,6 +97,7 @@ uword
  nn_left,
  leaf_node_counter,
  leaf_node_index_counter,
+ leaf_node_col,
  oobag_eval_counter;
 
 bool
@@ -1924,10 +1930,26 @@ List lrt_multi_testthat(NumericMatrix& y_node_,
 
 
 
-void oobag_pred_surv_uni(){
+void oobag_pred_surv_uni(char pred_type){
 
  iit_vals = sort_index(leaf_pred, "ascend");
  iit = iit_vals.begin();
+
+ switch(pred_type){
+
+ case 'S': case 'R':
+
+  leaf_node_col = 1;
+  pred_t0 = 1;
+  break;
+
+ case 'H':
+
+  leaf_node_col = 2;
+  pred_t0 = 0;
+  break;
+
+ }
 
  do {
 
@@ -1955,12 +1977,12 @@ void oobag_pred_surv_uni(){
    for(; i < leaf_node.n_rows; i++){
     if (leaf_node.at(i, 0) > time_pred){
      if(i == 0)
-      temp1 = 1;
+      temp1 = pred_t0;
      else
-      temp1 = leaf_node.at(i-1, 1);
+      temp1 = leaf_node.at(i-1, leaf_node_col);
      break;
     } else if (leaf_node.at(i, 0) == time_pred){
-     temp1 = leaf_node.at(i, 1);
+     temp1 = leaf_node.at(i, leaf_node_col);
      break;
     }
    }
@@ -1968,7 +1990,7 @@ void oobag_pred_surv_uni(){
   } else {
 
    // go here if prediction horizon > max time in current leaf.
-   temp1 = leaf_node.at(leaf_node.n_rows - 1, 1);
+   temp1 = leaf_node.at(leaf_node.n_rows - 1, leaf_node_col);
 
   }
 
@@ -2052,7 +2074,7 @@ double oobag_c_harrell_testthat(NumericMatrix y_mat,
 
 }
 
-void new_pred_surv_multi_mean(){
+void new_pred_surv_multi_mean(char pred_type){
 
  // allocate memory for output
  // surv_pvec.zeros(x_pred.n_rows);
@@ -2060,6 +2082,22 @@ void new_pred_surv_multi_mean(){
  surv_pvec.set_size(times_pred.size());
  iit_vals = sort_index(leaf_pred, "ascend");
  iit = iit_vals.begin();
+
+ switch(pred_type){
+
+ case 'S': case 'R':
+
+  leaf_node_col = 1;
+  pred_t0 = 1;
+  break;
+
+ case 'H':
+
+  leaf_node_col = 2;
+  pred_t0 = 0;
+  break;
+
+ }
 
  do {
 
@@ -2091,15 +2129,15 @@ void new_pred_surv_multi_mean(){
      if (leaf_node(i, 0) > time_pred){
 
       if(i == 0)
-       temp1 = 1;
+       temp1 = pred_t0;
       else
-       temp1 = leaf_node(i-1, 1);
+       temp1 = leaf_node(i-1, leaf_node_col);
 
       break;
 
      } else if (leaf_node(i, 0) == time_pred){
 
[…]

[…]
