+++ Rmd/orsf_ice_content_1.Rmd
@@ -1,34 +1,0 @@
-
-Begin by fitting an ORSF ensemble:
-
-```{r, child = 'orsf-fit-accelerated.Rmd'}
-
-```
-
-## Individual conditional expectations two ways
-
-You can compute individual conditional expectations two ways using a random forest:
-
-- using out-of-bag predictions for the training data
-
-  ```{r}
-  
-  pred_spec <- list(bili = seq(1, 6, length.out = 10))
-  
-  ice_train <- orsf_ice_oob(fit, pred_spec)
-  
-  ice_train
-  
-  ```
-
-- using predictions for a new set of data
-
-  ```{r}
-  
-  ice_test <- orsf_ice_new(fit, 
-                           new_data = pbc_orsf_test, 
-                           pred_spec = pred_spec)
-  
-  ice_test
-  
-  ```
