+++ NEWS.md
@@ -11,3 +11,5 @@
 * Added `orsf_control_custom()`, which allows users to submit custom functions for identifying linear combinations of inputs while growing oblique decision trees.
 
 * Added `weights` input to `orsf`, allowing users to over or under fit `orsf` to specific data in their training set.
+
+* Added `chf` and `mort` options to `predict.orsf_fit()`
