### 2022-08-18 36fe08c
minor doc edits
