+++ tests/testthat/test-orsf_summary.R
@@ -2,8 +2,7 @@
 
 object <- orsf(pbc_orsf,
                Surv(time, status) ~ . - id,
-               n_tree = 100,
-               oobag_pred = TRUE)
+               n_tree = 100)
 
 n_variables <- 3
 
