### 2022-08-20 688b12a
prepping for submission to cran
