+++ cran-comments.md
@@ -1,10 +1,17 @@
 
 ## R CMD check results 
 
-Duration: 3m 52.7s
+Duration: 2m 31.7s
 
 0 errors v | 0 warnings v | 0 notes v
 
+R CMD check succeeded
+
 ## Downstream dependencies
 
-I have also run R CMD check on downstream dependencies of aorsf: Rcpp and data.table. Both packages passed.
+I have also run R CMD check on downstream dependencies of `aorsf`: 
+
+- Rcpp
+- data.table. 
+
+Both packages passed.

[…]
