### 2022-08-21 #532 @bcjaeger
Hi @tdhock, I am not sure if it's my job or yours to say when the review is complete. Do you know?

Also, the application form mentions a badge for the submission:

> What grade of badge are you aiming for? ([bronze, silver, gold](https://stats-devguide.ropensci.org/pkgdev.html#pkgdev-badges))

I made a case for `aorsf` to receive the gold badge in my application. Do you have the final say on the badge it receives?
