+++ cran-comments.md
@@ -15,3 +15,11 @@ I have also run R CMD check on downstream dependencies of `aorsf`:
 - data.table. 
 
 Both packages passed.
+
+## Revision: 2022-08-22
+
+Updates: 
+
+- ensured all links use https instead of http and have added trailing slashes where appropriate.
+
+- added a reference to `DESCRIPTION` that discusses the methods used in `aorsf`.
