+++ man/aorsf-package.Rd
@@ -8,14 +8,14 @@
 \description{
 \if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}
 
-Fit, interpret, and make predictions with oblique random survival forests. Oblique decision trees are notoriously slow compared to their axis based counterparts, but 'aorsf' runs as fast or faster than axis-based decision tree algorithms for right-censored time-to-event outcomes.
+Fit, interpret, and make predictions with oblique random survival forests. Oblique decision trees are notoriously slow compared to their axis based counterparts, but 'aorsf' runs as fast or faster than axis-based decision tree algorithms for right-censored time-to-event outcomes. Methods to accelerate and interpret the oblique random survival forest are described in Jaeger et al., (2022) \href{https://arxiv.org/abs/2208.01129}{arXiv:2208.01129}.
 }
 \seealso{
 Useful links:
 \itemize{
-  \item \url{https://github.com/bcjaeger/aorsf}
-  \item \url{https://bcjaeger.github.io/aorsf}
-  \item Report bugs at \url{https://github.com/bcjaeger/aorsf/issues}
+  \item \url{https://github.com/bcjaeger/aorsf/}
+  \item \url{https://bcjaeger.github.io/aorsf/}
+  \item Report bugs at \url{https://github.com/bcjaeger/aorsf/issues/}
 }
 
 }
