### 2022-08-22 5551352
kick out nans from unused factor levels
