+++ tests/testthat/test-orsf_vi.R
@@ -337,3 +337,21 @@ test_that(
  }
 )
 
+data_with_empty_factor <- pbc_orsf
+
+levels(data_with_empty_factor$sex) <-
+ c(levels(data_with_empty_factor$sex), 'o')
+
+fit <- orsf(data_with_empty_factor, time + status ~ . - id)
+
+test_that(
+ desc = 'unused factor levels have nan for importance, but this gets kicked out of the aggregated factor importance',
+ code = {
+  expect_equal(sum(is.nan(fit$importance)), 1)
+  expect_equal(sum(is.nan(orsf_vi(fit))), 0)
+ }
+)
+
+
+
+
