+++ man/orsf.Rd
@@ -291,9 +291,11 @@ library(tidymodels)
 ## x purrr::discard()      masks scales::discard()
 ## x dplyr::filter()       masks stats::filter()
 ## x dplyr::first()        masks aorsf::first()
+## x purrr::is_null()      masks testthat::is_null()
 ## x recipes::is_trained() masks aorsf::is_trained()
 ## x dplyr::lag()          masks stats::lag()
 ## x dplyr::last()         masks aorsf::last()
+## x tidyr::matches()      masks rsample::matches(), dplyr::matches(), testthat::matches()
 ## x recipes::step()       masks stats::step()
 ## x purrr::transpose()    masks aorsf::transpose()
 ## * Dig deeper into tidy modeling with R at https://www.tmwr.org
@@ -308,16 +310,20 @@ library(tidymodels)
 ## v stringr 1.4.0
 
 ## -- Conflicts ------------------------------------- tidyverse_conflicts() --
-## x dplyr::between()    masks aorsf::between()
-## x readr::col_factor() masks scales::col_factor()
-## x purrr::discard()    masks scales::discard()
-## x dplyr::filter()     masks stats::filter()
-## x dplyr::first()      masks aorsf::first()
-## x stringr::fixed()    masks recipes::fixed()
-## x dplyr::lag()        masks stats::lag()
-## x dplyr::last()       masks aorsf::last()
-## x readr::spec()       masks yardstick::spec()
-## x purrr::transpose()  masks aorsf::transpose()
+## x dplyr::between()       masks aorsf::between()
+## x readr::col_factor()    masks scales::col_factor()
+## x purrr::discard()       masks scales::discard()
+## x readr::edition_get()   masks testthat::edition_get()
+## x dplyr::filter()        masks stats::filter()
+## x dplyr::first()         masks aorsf::first()
+## x stringr::fixed()       masks recipes::fixed()
+## x purrr::is_null()       masks testthat::is_null()
+## x dplyr::lag()           masks stats::lag()
+## x dplyr::last()          masks aorsf::last()
+## x readr::local_edition() masks testthat::local_edition()
+## x tidyr::matches()       masks rsample::matches(), dplyr::matches(), testthat::matches()
+## x readr::spec()          masks yardstick::spec()
+## x purrr::transpose()     masks aorsf::transpose()
 }\if{html}{\out{</div>}}
 
 \if{html}{\out{<div class="sourceCode r">}}\preformatted{library(randomForestSRC)
@@ -623,29 +629,29 @@ glimpse(results)
 
 \if{html}{\out{<div class="sourceCode">}}\preformatted{## Rows: 276
 ## Columns: 23
-## $ id          <int> 7, 36, 52, 65, 67, 72, 79, 88, 89, 100, 127, 154, 161, 167~
[…]
