+++ DESCRIPTION
@@ -19,12 +19,7 @@ Authors@R: c(
            email = "chris.jackson@mrc-bsu.cam.ac.uk",
            role = "rev")
     )
-Description: Fit, interpret, and make predictions with oblique random 
-    survival forests. Oblique decision trees are notoriously slow compared 
-    to their axis based counterparts, but 'aorsf' runs as fast or faster than 
-    axis-based decision tree algorithms for right-censored time-to-event 
-    outcomes. Methods to accelerate and interpret the oblique random survival 
-    forest are described in Jaeger et al., (2022) <arXiv:2208.01129>.
+Description: Fit, interpret, and make predictions with oblique random survival forests. Oblique decision trees are notoriously slow compared to their axis based counterparts, but 'aorsf' runs as fast or faster than axis-based decision tree algorithms for right-censored time-to-event outcomes. Methods to accelerate and interpret the oblique random survival forest are described in Jaeger et al., (2022) <arXiv:2208.01129>.
 License: MIT + file LICENSE
 Encoding: UTF-8
 LazyData: true
