### 2022-08-24 011036a
updates for additional CRAN checks
