+++ joss/paper.bib
@@ -49,7 +49,7 @@
 @article{breiman2001random,
   title={Random forests},
   author={Breiman, Leo},
-  journal={Machine learning},
+  journal={{M}achine {L}earning},
   volume={45},
   number={1},
   pages={5--32},
