+++ man/orsf.Rd
@@ -275,7 +275,7 @@ library(survival)
 library(tidymodels)
 }\if{html}{\out{</div>}}
 
-\if{html}{\out{<div class="sourceCode">}}\preformatted{## -- Attaching packages --------------------------------- tidymodels 0.2.0 --
+\if{html}{\out{<div class="sourceCode">}}\preformatted{## -- Attaching packages ------------------------------------ tidymodels 0.2.0 --
 
 ## v broom        1.0.0     v recipes      1.0.1
 ## v dials        0.1.1     v rsample      0.1.1
@@ -286,44 +286,38 @@ library(tidymodels)
 ## v parsnip      1.0.0     v workflowsets 0.2.1
 ## v purrr        0.3.4     v yardstick    0.0.9
 
-## -- Conflicts ------------------------------------ tidymodels_conflicts() --
+## -- Conflicts --------------------------------------- tidymodels_conflicts() --
 ## x dplyr::between()      masks aorsf::between()
 ## x purrr::discard()      masks scales::discard()
 ## x dplyr::filter()       masks stats::filter()
 ## x dplyr::first()        masks aorsf::first()
-## x purrr::is_null()      masks testthat::is_null()
 ## x recipes::is_trained() masks aorsf::is_trained()
 ## x dplyr::lag()          masks stats::lag()
 ## x dplyr::last()         masks aorsf::last()
-## x tidyr::matches()      masks rsample::matches(), dplyr::matches(), testthat::matches()
 ## x recipes::step()       masks stats::step()
 ## x purrr::transpose()    masks aorsf::transpose()
-## * Dig deeper into tidy modeling with R at https://www.tmwr.org
+## * Use suppressPackageStartupMessages() to eliminate package startup messages
 }\if{html}{\out{</div>}}
 
 \if{html}{\out{<div class="sourceCode r">}}\preformatted{library(tidyverse)
 }\if{html}{\out{</div>}}
 
-\if{html}{\out{<div class="sourceCode">}}\preformatted{## -- Attaching packages ---------------------------------- tidyverse 1.3.1 --
+\if{html}{\out{<div class="sourceCode">}}\preformatted{## -- Attaching packages ------------------------------------- tidyverse 1.3.1 --
 
 ## v readr   2.1.2     v forcats 0.5.1
 ## v stringr 1.4.0
 
-## -- Conflicts ------------------------------------- tidyverse_conflicts() --
-## x dplyr::between()       masks aorsf::between()
-## x readr::col_factor()    masks scales::col_factor()
-## x purrr::discard()       masks scales::discard()
-## x readr::edition_get()   masks testthat::edition_get()
-## x dplyr::filter()        masks stats::filter()
-## x dplyr::first()         masks aorsf::first()
-## x stringr::fixed()       masks recipes::fixed()
-## x purrr::is_null()       masks testthat::is_null()
[…]
