+++ man/orsf_vi.Rd
@@ -156,12 +156,12 @@ To get aggregated values across all levels of each factor, use
 \if{html}{\out{<div class="sourceCode r">}}\preformatted{orsf_vi(fit)
 }\if{html}{\out{</div>}}
 
-\if{html}{\out{<div class="sourceCode">}}\preformatted{##    ascites       bili      edema     copper        age    albumin 
-## 0.35528942 0.27837977 0.24719876 0.19605331 0.18822292 0.17033964 
-##    protime       chol      stage    spiders        ast     hepato 
-## 0.15320911 0.14883599 0.13627743 0.13495783 0.12754159 0.12057626 
-##        sex       trig   alk.phos   platelet        trt 
-## 0.10669014 0.09723320 0.09183673 0.07599581 0.06846999
+\if{html}{\out{<div class="sourceCode">}}\preformatted{##    ascites       bili      edema     copper        age    albumin    protime 
+## 0.35528942 0.27837977 0.24719876 0.19605331 0.18822292 0.17033964 0.15320911 
+##       chol      stage    spiders        ast     hepato        sex       trig 
+## 0.14883599 0.13627743 0.13495783 0.12754159 0.12057626 0.10669014 0.09723320 
+##   alk.phos   platelet        trt 
+## 0.09183673 0.07599581 0.06846999
 }\if{html}{\out{</div>}}
 }
 
