+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: aorsf
 Title: Accelerated Oblique Random Survival Forests
-Version: 0.0.1.9000
+Version: 0.0.2
 Authors@R: c(
     person(given = "Byron",
            family = "Jaeger",
