### 2022-08-24 99d2891
fix for https://github.com/openjournals/joss-reviews/issues/4705#issuecomment-1225786801
