+++ joss/paper.bib
@@ -1,7 +1,7 @@
 @article{ishwaran2008random,
   title={Random survival forests},
   author={Ishwaran, Hemant and Kogalur, Udaya B and Blackstone, Eugene H and Lauer, Michael S},
-  journal={The annals of applied statistics},
+  journal={{T}he {A}nnals of {A}pplied {S}tatistics},
   volume={2},
   number={3},
   pages={841--860},
@@ -69,7 +69,7 @@
 @article{jaeger2019oblique,
   title={Oblique random survival forests},
   author={Jaeger, Byron and Long, D Leann and Long, Dustin M and Sims, Mario and Szychowski, Jeff M and Min, Yuan-I and Mcclure, Leslie A and Howard, George and Simon, Noah},
-  journal={The Annals of Applied Statistics},
+  journal={{T}he {A}nnals of {A}pplied {S}tatistics},
   volume={13},
   number={3},
   pages={1847--1883},
