### 2022-08-24 e00920d
drop the ? from parenthetical. https://github.com/openjournals/joss-reviews/issues/4705#issuecomment-1225783697
