### 2022-08-24 755db8e
fixes for problems found by brian R
