+++ NEWS.md
@@ -13,3 +13,7 @@
 * Added `weights` input to `orsf`, allowing users to over or under fit `orsf` to specific data in their training set.
 
 * Added `chf` and `mort` options to `predict.orsf_fit()`. Mortality predictions are not fully implemented yet - they are not supported in partial dependence or out-of-bag error estimates. These features will be added in a future update.
+
+# aorsf 0.0.2
+
+* Modified unit tests for compatibility with extra checks run through CRAN.
