+++ joss/paper.bib
@@ -66,6 +66,15 @@
   organization={Springer}
 }
 
+@inproceedings{menze2011oblique,
+  title={On oblique random forests},
+  author={Menze, Bjoern H and Kelm, B Michael and Splitthoff, Daniel N and Koethe, Ullrich and Hamprecht, Fred A},
+  booktitle={Joint European Conference on Machine Learning and Knowledge Discovery in Databases},
+  pages={453--469},
+  year={2011},
+  organization={Springer}
+}
+
 @article{jaeger2019oblique,
   title={Oblique random survival forests},
   author={Jaeger, Byron and Long, D Leann and Long, Dustin M and Sims, Mario and Szychowski, Jeff M and Min, Yuan-I and Mcclure, Leslie A and Howard, George and Simon, Noah},
