+++ joss/paper.bib
@@ -6,6 +6,7 @@
   number={3},
   pages={841--860},
   year={2008},
+  doi={10.1214/08-AOAS169},
   publisher={Institute of Mathematical Statistics}
 }
 
@@ -46,16 +47,19 @@
   copyright = {Creative Commons Attribution 4.0 International}
 }
 
+
 @article{breiman2001random,
-  title={Random forests},
+  title={Random Forests},
   author={Breiman, Leo},
   journal={{M}achine {L}earning},
   volume={45},
   number={1},
   pages={5--32},
   year={2001},
+  doi={10.1023/A:1010933404324},
   publisher={Springer}
 }
+
 
 @inproceedings{menze2011oblique,
   title={On oblique random forests},
@@ -83,6 +87,7 @@
   number={3},
   pages={1847--1883},
   year={2019},
+  doi={10.1214/19-AOAS1261},
   publisher={Institute of Mathematical Statistics}
 }
 
@@ -94,8 +99,10 @@
   number={24},
   pages={2370--2383},
   year={2021},
+  doi={10.1161/circulationaha.120.053134},
   publisher={Am Heart Assoc}
 }
+
 
 @article{katuwal2020heterogeneous,
   title={Heterogeneous oblique random forest},
@@ -104,5 +111,6 @@
   volume={99},
   pages={107078},
   year={2020},
+  doi={10.1016/j.patcog.2019.107078},
   publisher={Elsevier}
 }
