### 2022-08-24 057cf2b
cran badge for 0.0.1
