### 2022-08-25 6ec0156
make statement of need less dependent on background
