### 2022-08-25 #532 @danielskatz
👋 @tdhock - please ping me when this is complete so I can also complete the corresponding JOSS acceptance and publication (in https://github.com/openjournals/joss-reviews/issues/4705)
