### 2022-08-31 #532 @bcjaeger
Cool, thanks @jemus42 and @mnwright for agreeing to review! In case it is helpful, https://github.com/bcjaeger/aorsf-bench contains the paper and code to evaluate `aorsf` in terms of prediction accuracy, variable importance, and efficiency (paper is here: https://github.com/bcjaeger/aorsf-bench/blob/main/paper/jmlr/main.pdf)

Unfortunately, `aorsf` is now on CRAN. I am very sorry for this inconvenience. I moved forward with the CRAN submission because the 'seeking-reviewers' tag was removed from this issue on July 11 and I did not realize we could still seek reviewers once that tag was removed.

I am extremely grateful for your willingness to review `aorsf`, and I will try my best to incorporate all of your recommendations in a way that is compatible with the version of `aorsf` that is on CRAN. If you feel strongly about making a change that would not be compatible with the version of `aorsf` on CRAN, I am happy to discuss it.
