### 2022-09-02 #532 @mnwright
None.

I couldn't find contribution guidelines.

---

Thanks for the great package. Every package should be so well implemented, organized and documented! In particular, the documentation is outstanding.

I only have a few questions/comments, mostly minor:
* `?orsf` contains startup messages from randomForestSRC and riskRegression, maybe suppress them?
* The function `oobag_c_harrell()` is repeated several times in the tests (presumable they are identical). Maybe reduce redundancy here?
* In the OOB vignette, the example for the user-supplied OOB function is incorrect (as explained in the text). In my experience, some people copy&paste such things anyway, so I would try to avoid incorrect examples.
* The changelog (https://bcjaeger.github.io/aorsf/news/index.html) starts with the oldest changes. Isn't that usually the other way around?
* I think `orsf` can only be called with the `formula` parameter. Does this work for high dimensional data, e.g. p>10,000 or is it slow/failing?
* The C++ code is lacking documentation, i.e., it is unclear what the functions in the C++ code are doing. I think all exported Rcpp functions are internal (not directly called by the user), so I won't expect documentation on the same levels as for exported R functions. Still, some documentation would help contributors work with the C++ code.
* All C++ code is in one large file with many global objects. It is common practice in C++ to avoid global objects where possible. Would it make sense to use some OOP instead?

And one thing, which is beyond the scope of this review but still interesting: What about oblique RF for classification and regression? Do other packages implement the speedups used here? If not, why not have a fast oblique RF package for all outcomes instead of a specific one for survival?

In summary, this is a great submission, ready for publication. I think most of the issues above are quite easy to fix. The C++ OOP question would be a major re-design and is not necessary for acceptance of the submission.
