### 2022-09-02 #532 @mnwright
And a few notes on the runtime comparisons above:
* ranger does not reduce the number of unique time points and that's what makes it slow and increases the memory usage. The bottleneck not the sample size but that the CHF is estimated at each unique time point. rfsrc for example uses a grid (`ntime` parameter). See also here https://github.com/imbs-hl/ranger/issues/410. I re-ran that experiment with a grid in ranger and then it was fastest (and didn't show the strange behavior shown in the plots above).
* Growing only one tree might give misleading results because of computation overhead, e.g. with Rcpp.
* Formula interfaces are quite slow for many features (I guess for all packages?). That might be a problem for p=1000 (see comment above).
