### 2022-09-02 #532 @bcjaeger
@mnwright, thank you so much for your review! The suggestions you've made to improve `aorsf` are incredibly helpful. I've reviewed each of your comments and written up a short description of actions I will take to incorporate your suggestions. If you think any of the actions I am proposing should be modified, just let me know, I am happy to coordinate a better approach.

Once I have finished all of the updates, I will post and tag you.

> `?orsf` contains startup messages from `randomForestSRC` and `riskRegression`, maybe suppress them?

Whoops, sorry I missed that. I will update with `suppressPackageStartupMessages`.

> The function oobag_c_harrell() is repeated several times in the tests (presumable they are identical). Maybe reduce redundancy here?

I like this idea. I think making `oobag_c_harrell()` an unexported R function in `aorsf` will resolve this.

> In the OOB vignette, the example for the user-supplied OOB function is incorrect (as explained in the text). In my experience, some people copy&paste such things anyway, so I would try to avoid incorrect examples.

Agreed - I think using `SurvMetrics::Brier` instead of writing a simple (incorrect) Brier score function is a good fix for this.

> The changelog (https://bcjaeger.github.io/aorsf/news/index.html) starts with the oldest changes. Isn't that usually the other way around?

Whoops - yes, that needs to be reversed. Thank you! 😅

> I think `orsf` can only be called with the `formula` parameter. Does this work for high dimensional data, e.g. p>10,000 or is it slow/failing?

That is a great question - I think I need to include a high-dimensional case in the benchmark above to answer it. I'd also like to experiment with adding an `x` and `y` input for `orsf`, similar to `ranger`, to see how much using `x` and `y` rather than `formula` impacts `aorsf`'s efficiency in high-dimensional settings.

> The C++ code is lacking documentation, i.e., it is unclear what the functions in the C++ code are doing. I think all exported Rcpp functions are internal (not directly called by the user), so I won't expect documentation on the same levels as for exported R functions. Still, some documentation would help contributors work with the C++ code.

Agreed, this code needs better documentation. I will go through each function and add comments that include description of the function's purpose and a description of each input for the function, including global variables that the function depends on.

> All C++ code is in one large file with many global objects. It is common practice in C++ to avoid global objects where possible. Would it make sense to use some OOP instead?

Yes, using OOP would be a much better design and would make generalizations to oblique forests for regression and classification much cleaner. The long term plan for `aorsf` absolutely includes a more modular re-write of the C++ file. The only reason `aorsf` didn't have this design from the beginning is that I am pretty inexperienced when it comes to performant C++ code. I am slowly getting there.

> What about oblique RF for classification and regression? Do other packages implement the speedups used here? If not, why not have a fast oblique RF package for all outcomes instead of a specific one for survival?

Agreed! I focused on survival initially because I wasn't aware of any R packages for oblique survival trees (the `RLT` package fits oblique trees for regression and classification, and `obliqueRF` can fit oblique trees for classification). Developing `aorsf` into a more general package that fits all three types of trees will be a big undertaking, but should be much more feasible after updating the C++ code as described above.

> ranger does not reduce the number of unique time points and that's what makes it slow and increases the memory usage. The bottleneck not the sample size but that the CHF is estimated at each unique time point. rfsrc for example uses a grid (ntime parameter). See also here https://github.com/imbs-hl/ranger/issues/410. I re-ran that experiment with a grid in ranger and then it was fastest (and didn't show the strange behavior shown in the plots above).

Ahh, I understand. Sorry I did not control for this in my initial benchmark. I'd like to update my benchmark code so that `ranger` can use a grid of timepoints for the CHF as you described. Would you be willing to share your code for this? I can also manually restrict the number of unique event times in the benchmark data if you think that would be a better approach.

> Growing only one tree might give misleading results because of computation overhead, e.g. with Rcpp.

Agreed - I think I can revert to using 500 trees assuming the memory issue is fixed by restricting the number of unique event times.

> Formula interfaces are quite slow for many features (I guess for all packages?). That might be a problem for p=1000 (see comment above).

Good point - if `orsf()` crashes due to its `formula` input with p = 1000, I will prioritize incorporating the `x` and `y` inputs into `orsf()` and reassess.

> In summary, this is a great submission, ready for publication. I think most of the issues above are quite easy to fix.

Thank you! I sincerely appreciate the time you invested in reviewing `aorsf` and your thoughtful feedback. It will be my honor to include you as a package reviewer in the `DESCRIPTION` file.
