+++ R/orsf_pd.R
@@ -33,11 +33,15 @@
 #'   - 'chf': cumulative hazard function
 #'   - 'mort': mortality prediction
 #'
+#' @param na_action `r roxy_na_action_header()`
+#'
+#'   - `r roxy_na_action_fail()`
+#'   - `r roxy_na_action_omit()`
+#'
 #' @param expand_grid (_logical_) if `TRUE`, partial dependence will be
 #'   computed at all possible combinations of inputs in `pred_spec`. If
 #'   `FALSE`, partial dependence will be computed for each variable
 #'   in `pred_spec`, separately.
-#'
 #'
 #' @param prob_values (_numeric_) a vector of values between 0 and 1,
 #'   indicating what quantiles will be used to summarize the partial
@@ -65,7 +69,6 @@
 #' @includeRmd Rmd/orsf_pd_examples.Rmd
 #'
 #' @export
-#'
 #'
 orsf_pd_oob <- function(object,
                         pred_spec,
@@ -133,6 +136,7 @@ orsf_pd_new <- function(object,
                         new_data,
                         pred_horizon = NULL,
                         pred_type = 'risk',
+                        na_action = 'fail',
                         expand_grid = TRUE,
                         prob_values = c(0.025, 0.50, 0.975),
                         prob_labels = c('lwr', 'medn', 'upr'),
@@ -146,6 +150,7 @@ orsf_pd_new <- function(object,
                       pd_data = new_data,
                       pred_horizon = pred_horizon,
                       pred_type = pred_type,
+                      na_action = na_action,
                       expand_grid = expand_grid,
                       prob_values = prob_values,
                       prob_labels = prob_labels,
@@ -231,6 +236,7 @@ orsf_ice_new <- function(object,
                          new_data,
                          pred_horizon = NULL,
                          pred_type = 'risk',
+                         na_action = 'fail',
                          expand_grid = TRUE,
                          boundary_checks = TRUE,
                          ...){
@@ -242,6 +248,7 @@ orsf_ice_new <- function(object,
                       pd_data = new_data,
                       pred_horizon = pred_horizon,
                       pred_type = pred_type,
+                      na_action = na_action,
                       expand_grid = expand_grid,
                       boundary_checks = boundary_checks,
                       oobag = FALSE,
@@ -271,6 +278,7 @@ orsf_pred_dependence <- function(object,
                                  pred_spec,
                                  pred_horizon,
                                  pred_type,
+                                 na_action = 'fail',
                                  expand_grid,
                                  prob_values = NULL,
                                  prob_labels = NULL,
@@ -291,7 +299,8 @@ orsf_pred_dependence <- function(object,
                  boundary_checks = boundary_checks,
                  new_data        = pd_data,
                  pred_type       = pred_type,
-                 pred_horizon    = pred_horizon)
+                 pred_horizon    = pred_horizon,
+                 na_action       = na_action)
 
  if(pred_type == 'mort') stop(
   "mortality predictions aren't supported in partial dependence functions",
@@ -307,14 +316,30 @@ orsf_pred_dependence <- function(object,
 
  if(oobag) pd_data <- object$data
 
+ if(is.null(pred_horizon) && pred_type != 'mort'){
+  stop("pred_horizon must be specified for ",
+       pred_type, " predictions.", call. = FALSE)
+ }
+
  type_input <- if(expand_grid) 'grid' else 'loop'
 
- x_new <- as.matrix(
-  ref_code(x_data = pd_data,
-           fi = get_fctr_info(object),
[…]
