### 2022-09-04 b9ff036
allow pred horizon over limit
