+++ man/orsf.Rd
@@ -296,7 +296,7 @@ library(tidymodels)
 ## x dplyr::last()         masks aorsf::last()
 ## x recipes::step()       masks stats::step()
 ## x purrr::transpose()    masks aorsf::transpose()
-## * Search for functions across packages at https://www.tidymodels.org/find/
+## * Dig deeper into tidy modeling with R at https://www.tmwr.org
 }\if{html}{\out{</div>}}
 
 \if{html}{\out{<div class="sourceCode r">}}\preformatted{library(tidyverse)
@@ -553,18 +553,18 @@ analyses
 
 \if{html}{\out{<div class="sourceCode">}}\preformatted{## #  10-fold cross-validation 
 ## # A tibble: 10 x 5
-##    splits           id     recipe   train               test    
-##    <list>           <chr>  <list>   <list>              <list>  
-##  1 <split [248/28]> Fold01 <recipe> <tibble [248 x 20]> <tibble>
-##  2 <split [248/28]> Fold02 <recipe> <tibble [248 x 20]> <tibble>
-##  3 <split [248/28]> Fold03 <recipe> <tibble [248 x 20]> <tibble>
-##  4 <split [248/28]> Fold04 <recipe> <tibble [248 x 20]> <tibble>
-##  5 <split [248/28]> Fold05 <recipe> <tibble [248 x 20]> <tibble>
-##  6 <split [248/28]> Fold06 <recipe> <tibble [248 x 20]> <tibble>
-##  7 <split [249/27]> Fold07 <recipe> <tibble [249 x 20]> <tibble>
-##  8 <split [249/27]> Fold08 <recipe> <tibble [249 x 20]> <tibble>
-##  9 <split [249/27]> Fold09 <recipe> <tibble [249 x 20]> <tibble>
-## 10 <split [249/27]> Fold10 <recipe> <tibble [249 x 20]> <tibble>
+##    splits           id     recipe   train               test              
+##    <list>           <chr>  <list>   <list>              <list>            
+##  1 <split [248/28]> Fold01 <recipe> <tibble [248 x 20]> <tibble [28 x 20]>
+##  2 <split [248/28]> Fold02 <recipe> <tibble [248 x 20]> <tibble [28 x 20]>
+##  3 <split [248/28]> Fold03 <recipe> <tibble [248 x 20]> <tibble [28 x 20]>
+##  4 <split [248/28]> Fold04 <recipe> <tibble [248 x 20]> <tibble [28 x 20]>
+##  5 <split [248/28]> Fold05 <recipe> <tibble [248 x 20]> <tibble [28 x 20]>
+##  6 <split [248/28]> Fold06 <recipe> <tibble [248 x 20]> <tibble [28 x 20]>
+##  7 <split [249/27]> Fold07 <recipe> <tibble [249 x 20]> <tibble [27 x 20]>
+##  8 <split [249/27]> Fold08 <recipe> <tibble [249 x 20]> <tibble [27 x 20]>
+##  9 <split [249/27]> Fold09 <recipe> <tibble [249 x 20]> <tibble [27 x 20]>
[…]
