### 2022-09-06 da8c335
better documentation for C++ functions
