+++ src/orsf.cpp
@@ -202,6 +202,13 @@ Environment base_env("package:base");
 
 Function set_seed_r = base_env["set.seed"];
 
+// Set difference for arma vectors
+//
+// @description the same as setdiff() in R
+//
+// @param x first vector
+// @param y second vector
+//
 // [[Rcpp::export]]
 arma::uvec std_setdiff(arma::uvec& x, arma::uvec& y) {
 
@@ -221,6 +228,22 @@ arma::uvec std_setdiff(arma::uvec& x, arma::uvec& y) {
 // ---------------------------- scaling functions -----------------------------
 // ----------------------------------------------------------------------------
 
+// scale observations in predictor matrix
+//
+// @description this scales inputs in the same way as
+//   the survival::coxph() function. The main reasons we do this
+//   are to avoid exponential overflow and to prevent the scale
+//   of inputs from impacting the estimated beta coefficients.
+//   E.g., you can try multiplying numeric inputs by 100 prior
+//   to calling orsf() with orsf_control_fast(do_scale = FALSE)
+//   and you will see that you get back a different forest.
+//
+// @param x_node matrix of predictors
+// @param w_node replication weights
+// @param x_transforms matrix used to store the means and scales
+//
+// @return modified x_node and x_transform filled with values
+//
 void x_node_scale(){
 
  // set aside memory for outputs
@@ -252,6 +275,8 @@ void x_node_scale(){
 
 }
 
+// same as above function, but just the means
+// (currently not used)
 void x_node_means(){
 
  x_transforms.zeros(n_vars, 1);
@@ -265,6 +290,7 @@ void x_node_means(){
 
 }
 
+//  Same as x_node_scale, but this can be called from R
 // [[Rcpp::export]]
 List x_node_scale_exported(NumericMatrix& x_,
                            NumericVector& w_){
@@ -287,6 +313,26 @@ List x_node_scale_exported(NumericMatrix& x_,
 // ----------------------------------------------------------------------------
 // -------------------------- leaf_surv functions -----------------------------
 // ----------------------------------------------------------------------------
+
+// Create kaplan-meier survival curve in leaf node
+//
+// @description Modifies leaf_nodes by adding data from the current node,
+//   where the current node is one that is too small to be split and will
+//   be converted to a leaf.
+//
+// @param y the outcome matrix in the current leaf
+// @param w the weights vector in the current leaf
+// @param leaf_indices a matrix that indicates where leaf nodes are
+//   inside of leaf_nodes. leaf_indices has three columns:
+//   - first column: the id for the leaf
+//   - second column: starting row for the leaf
+//   - third column: ending row for the leaf
+// @param leaf_node_index_counter keeps track of where we are in leaf_node
+// @param leaf_node_counter keeps track of which leaf node we are in
+// @param leaf_nodes a matrix with three columns:
+//   - first column: time
+//   - second column: survival probability
+//   - third column: cumulative hazard
 
 void leaf_kaplan(const arma::mat& y,
                  const arma::vec& w){
@@ -320,9 +366,6 @@ void leaf_kaplan(const arma::mat& y,
 
  }
 
-
- // TODO: ADD CUMULATIVE HAZARD FUNCTION HERE
- // (https://www.randomforestsrc.org/articles/survival.html#in-bag-ib-estimator-1)
  // reset for kaplan meier loop
  n_risk = sum(w);
  person = 0;
@@ -374,6 +417,10 @@ void leaf_kaplan(const arma::mat& y,
 
 }
 
+// Same as above, but this function can be called from R and is
[…]
