### 2022-09-06 #532 @bcjaeger
I am happy to share an updated response to the excellent review from @mnwright:

1. Examples in `orsf()` now suppress package start up messages (see [here](https://bcjaeger.github.io/aorsf/reference/orsf.html#examples))
2. `oobag_c_harrell()` is now an unexported function in `aorsf` to avoid repeatedly defining it in testing scripts.
3. `SurvMetrics::Brier` is now used in place of the incorrect Brier score in the out-of-bag vignette (see [here](https://bcjaeger.github.io/aorsf/articles/oobag.html#user-supplied-out-of-bag-evaluation-functions))
4. `NEWS.md` now shows updates in the usual order (see [here](https://bcjaeger.github.io/aorsf/news/index.html)).
5. C++ functions in `src/orsf.cpp` have documentation outlining the purpose of the function and its inputs (see [here](https://github.com/bcjaeger/aorsf/blob/main/src/orsf.cpp)).

I am committed to re-writing the `orsf.cpp` functions with modular, object oriented design, using the `ranger` files as a guideline. This will, among other things, make it far more feasible to efficiently fit oblique regression and classification forests with this package. I'll be submitting a grant in the Spring and will hopefully get funding to work through this.

The benchmark of `aorsf`, `ranger`, and `randomForestSRC` has also undergone some updates:

- 500 trees are  now used instead of 1.
- the number of unique event times is now fixed at 100, 1000, or 10000 by modifying the time values in the data.
- the number of predictors is now fixed at 100, 1000, or 10000 (it was originally 10, 100, or 1000)
- To avoid high memory usage from having many leaves with many unique event times, I increase the minimal terminal node size as a function of the sample size, i.e., `node_size = n_obs / 10`.

With these updates, it is clear that `aorsf` runs slow compared to `ranger` and `randomForestSRC` in cases where the number of unique event times is 100 or 1000. It is also clear that `aorsf` does not scale as well to high dimensional settings - it does best relative to the other packages when the number of predictors is 100 and the number of unique event times is 10000 (i.e., each observation has a unique event time). In my prediction benchmark, the number of predictors is often < 100 and there are few instances of non-unique event times, so I can understand why benchmark results seem to favor `aorsf` in that comparison.

If you aggregate over all 9 panels (see second figure), `ranger` has the fastest expected time and all three methods appear to have a linear asymptotic complexity.

![temp](https://user-images.githubusercontent.com/11821781/188756046-cde16983-e6f9-4a17-80c6-82ca61124f7b.png)

![temp2](https://user-images.githubusercontent.com/11821781/188756024-88c6f5ce-cb47-48ef-a43f-e51b404ed7df.png)
