### 2022-09-16 e2f5db2
data object for pointers to arma objects
