### 2022-09-16 b7afb1c
data object for pointers to arma objects
