### 2022-09-16 #532 @jemus42
None.
- [X] As the reviewer I confirm that there are no [conflicts of interest](https://devguide.ropensci.org/policies.html#coi) for me to review this work

  - From my understanding the "gist" is "`obliqueRSF`, but _fast_", so it inherits its usefulness/audience from `obliqueRSF`.
  - It clearly has a place between existing methods and implementations.
  - The documentation is comprehensive and useful for both novice and experienced users.
  - I found contribution guidelines at `.github/CONTRIBUTING.md`.

---

The majority of my review are based on the state of the package as of [commit 2251066](https://github.com/bcjaeger/aorsf/tree/2251066cfbf920984072d0e6924d32a7dfc8ff7a).

- Overall the package features comprehensive documentation and examples.
- Function names are consistent and discoverable as well as tab-completable.
- User-supplied arguments are checked consistently via internal functions defined in `R/check.R`.
I would like to suggest the [`checkmate`](https://cran.r-project.org/package=checkmate) package for argument checks, which could greatly reduce complexity of internal functions/checks. I do concede that now that the work is already done it's likely not worth re-implementing checks again. Also, the author [wants to keep dependencies to a minimum](https://github.com/bcjaeger/aorsf/blob/2251066cfbf920984072d0e6924d32a7dfc8ff7a/R/srr-stats-standards.R#L111), so I understand the manual implementation of argument checks.
- The `check_*` functions mentioned above and internal `roxy_*` functions for documentation snippets both reduce code duplication and ensure documentation is easier to keep consistent and up to date.
    - Some of the `roxy_*` documentation helpers may be exchangeable for `roxygen2` `@importFrom` tags or templates etc., but that's merely a gut feeling about idiomatic use of `roxygen2`.
- I have not found any instances of `@srr` tags/comments I would challenge.
- I could not find any reference to parallel processing, i.e. using multiple CPU cores/threads to speed up training. That's not a direct issue I think, but maybe something to look into in the future? Especially given the application on large datasets.
- The test suite seems comprehensive and checks for correctness against well-established methods, but I'm wondering if it would be useful to add an automated regression test against `obliqueRSF` (at another location, to avoid the `Suggests:`-dependency on it).

Overall my impression of the package is very positive!

#### Performance

I have not conducted my own exhaustive experiments and have no additional notes regarding the public benchmark at [`aorsf-bench`](https://github.com/bcjaeger/aorsf-bench).

The author also kindly contributed to [`mlr3extralearners`][https://github.com/mlr-org/mlr3extralearners/pull/233], which simplifies the algorithm's inclusion in other benchmark experiments.
The initiative to connect to existing frameworks like `mlr3` and `tidymodels` is very much appreciated and allowed me to run a quick (naive) benchmark against `ranger` and `rfrsc`, showing comparable performance regarding C-index and brier score, as well as significantly shorter run-time for low-dimensional tasks (caveat: no tuning and default parameters were used for all algorithms, it was merely a quick test-run of the `mlr3` learner).

A very minor nitpicky comment regarding code efficiency:
[`fctr.R#17`](https://github.com/bcjaeger/aorsf/blob/2251066cfbf920984072d0e6924d32a7dfc8ff7a/R/fctr.R#L17) creates an empty vector (`c()`) and grows it with each `character` variable. For small datasets this is likely irrelevant, but imagining a scenario with many `character` variables supplied intentionally or accidentally by the user, it may be more appropriate to use a `vapply` construct like in this example:

``` r
# Example data
data <- as.data.frame(matrix(rnorm(100), ncol = 10))
data$V4 <- as.character(data$V4)
data$V8 <- as.character(data$V8)
data$V10 <- as.character(data$V10)

# Recreating .names argument
.names <- names(data)

# character detection without a for loop and a growing vector
which_char <- vapply(data, is.character, logical(1), USE.NAMES = FALSE)
chrs <- .names[which(which_char)]
chrs
#> [1] "V4"  "V8"  "V10"
```

Also note that the output will likely be excessive if a user were to accidentally supply a very large dataset with thousands of `character` variables, but that's a small edge case with little impact I think.

#### Methods

The inclusion of multiple easy to discover/use variable importance methods is very nice to see!

**Partial dependence plots**: Due to their [known limitations/assumptions](https://christophm.github.io/interpretable-ml-book/pdp.html#disadvantages-5), it may be appropriate to
  a) note those limitations in the documentation to ensure users unfamiliar with the technique can make an informed choice
  b) consider additionally supporting accumulated local effects, a ["faster and unbiased alternative" to PDPs](https://christophm.github.io/interpretable-ml-book/ale.html), if possible.

**Permutation feature importance**: This method has drawbacks as well, see [Hooker et. al.](https://arxiv.org/abs/1905.03151v2), so the previous comment applies with the exception that I cannot suggest a direct alternative.

I was not familiar with the ANOVA- and negation-based importance measures, and I suspect many users may feel the same way. In that regard I think it would be useful to elaborate on these methods slightly, e.g. in `vignettes/aorsf.Rmd` where they are mentioned next to the aforementioned methods, or in the help page at `?orsf_vi` in the `Details` section which already outlines the methods.
Reading the [paper/preprint in `aorsf-bench`](https://github.com/bcjaeger/aorsf-bench/blob/main/paper/arxiv/main.pdf) was helpful, but I think some rough guidance regarding pros/cons in the documentation would be beneficial for users, if one can even be provided.
Choosing interpretability methods is of course not trivial, and I don't expect the author to provide a conclusive heuristic here.
