### 2022-09-16 19796b8
faster check with vapply
