### 2022-09-16 #532 @bcjaeger
@jemus42, thank you! I will write a more complete response to your feedback and a list of proposed actions in the next few days, but for now I just wanted to say how much I appreciate your review. You found a very nice solution to speed up `fctr_check` and I've just pushed a commit to `aorsf` with your suggested use of `vapply` . As expected, it's much faster. =]

``` r
n <- 500
p <- 1000

data_check <- list()

characters <- letters[1:5]

for( i in seq(p) ) {
 data_check[[i]] <- sample(characters, size = n, replace = T)
 names(data_check)[i] <- paste('x', i, sep = '_')
}

data_check <- as.data.frame(data_check)

c_loop <- function(data, .names){

 chrs <- c()

 for( .name in .names ){

  if(is.character(data[[.name]])){
   chrs <- c(chrs, .name)
  }

 }

 chrs

}

v_apply <- function(data, .names){

 chr_index <- which(
  vapply(data[, .names], is.character, logical(1), USE.NAMES = FALSE)
 )

 .names[chr_index]

}

testthat::expect_equal(
 c_loop(data_check, names(data_check)),
 v_apply(data_check, names(data_check))
)

microbenchmark::microbenchmark(
 c_loop = c_loop(data_check, names(data_check)),
 v_apply = v_apply(data_check, names(data_check))
)
#> Unit: microseconds
#>     expr      min       lq    mean    median       uq       max neval cld
#>   c_loop 8948.500 9234.001 9747.49 9380.5010 9530.401 13457.202   100   b
#>  v_apply  328.001  339.701  484.56  346.6005  385.701  5247.201   100  a
```
