### 2022-09-17 3e2c2df
rcpp arma sample can only be defined once
