+++ R/RcppExports.R
@@ -1,70 +1,6 @@
 # Generated by using Rcpp::compileAttributes() -> do not edit by hand
 # Generator token: 10BE3573-1514-4C36-9D1C-5A225CD40393
 
-std_setdiff <- function(x, y) {
-    .Call(`_aorsf_std_setdiff`, x, y)
-}
-
-x_node_scale_exported <- function(x_, w_) {
-    .Call(`_aorsf_x_node_scale_exported`, x_, w_)
-}
-
-leaf_kaplan_testthat <- function(y, w) {
-    .Call(`_aorsf_leaf_kaplan_testthat`, y, w)
-}
-
-newtraph_cph_testthat <- function(x_in, y_in, w_in, method, cph_eps_, iter_max) {
-    .Call(`_aorsf_newtraph_cph_testthat`, x_in, y_in, w_in, method, cph_eps_, iter_max)
-}
-
-lrt_multi_testthat <- function(y_node_, w_node_, XB_, n_split_, leaf_min_events_, leaf_min_obs_) {
-    .Call(`_aorsf_lrt_multi_testthat`, y_node_, w_node_, XB_, n_split_, leaf_min_events_, leaf_min_obs_)
-}
-
-oobag_c_harrell_testthat <- function(y_mat, s_vec) {
-    .Call(`_aorsf_oobag_c_harrell_testthat`, y_mat, s_vec)
-}
-
-ostree_pred_leaf_testthat <- function(tree, x_pred_) {
-    .Call(`_aorsf_ostree_pred_leaf_testthat`, tree, x_pred_)
-}
-
-orsf_fit <- function(x, y, weights, n_tree, n_split_, mtry_, leaf_min_events_, leaf_min_obs_, split_min_events_, split_min_obs_, split_min_stat_, cph_method_, cph_eps_, cph_iter_max_, cph_do_scale_, net_alpha_, net_df_target_, oobag_pred_, oobag_pred_type_, oobag_pred_horizon_, oobag_eval_every_, oobag_importance_, oobag_importance_type_, tree_seeds, max_retry_, f_beta, type_beta_, f_oobag_eval, type_oobag_eval_) {
-    .Call(`_aorsf_orsf_fit`, x, y, weights, n_tree, n_split_, mtry_, leaf_min_events_, leaf_min_obs_, split_min_events_, split_min_obs_, split_min_stat_, cph_method_, cph_eps_, cph_iter_max_, cph_do_scale_, net_alpha_, net_df_target_, oobag_pred_, oobag_pred_type_, oobag_pred_horizon_, oobag_eval_every_, oobag_importance_, oobag_importance_type_, tree_seeds, max_retry_, f_beta, type_beta_, f_oobag_eval, type_oobag_eval_)
-}
-
-orsf_oob_negate_vi <- function(x, y, forest, last_eval_stat, time_pred_, f_oobag_eval, pred_type_, type_oobag_eval_) {
-    .Call(`_aorsf_orsf_oob_negate_vi`, x, y, forest, last_eval_stat, time_pred_, f_oobag_eval, pred_type_, type_oobag_eval_)
-}
-
-orsf_oob_permute_vi <- function(x, y, forest, last_eval_stat, time_pred_, f_oobag_eval, pred_type_, type_oobag_eval_) {
-    .Call(`_aorsf_orsf_oob_permute_vi`, x, y, forest, last_eval_stat, time_pred_, f_oobag_eval, pred_type_, type_oobag_eval_)
-}
-
-orsf_pred_uni <- function(forest, x_new, time_dbl, pred_type) {
[…]
