+++ src/RcppExports.cpp
@@ -11,269 +11,6 @@ Rcpp::Rostream<true>&  Rcpp::Rcout = Rcpp::Rcpp_cout_get();
 Rcpp::Rostream<false>& Rcpp::Rcerr = Rcpp::Rcpp_cerr_get();
 #endif
 
-// std_setdiff
-arma::uvec std_setdiff(arma::uvec& x, arma::uvec& y);
-RcppExport SEXP _aorsf_std_setdiff(SEXP xSEXP, SEXP ySEXP) {
-BEGIN_RCPP
-    Rcpp::RObject rcpp_result_gen;
-    Rcpp::RNGScope rcpp_rngScope_gen;
-    Rcpp::traits::input_parameter< arma::uvec& >::type x(xSEXP);
-    Rcpp::traits::input_parameter< arma::uvec& >::type y(ySEXP);
-    rcpp_result_gen = Rcpp::wrap(std_setdiff(x, y));
-    return rcpp_result_gen;
-END_RCPP
-}
-// x_node_scale_exported
-List x_node_scale_exported(NumericMatrix& x_, NumericVector& w_);
-RcppExport SEXP _aorsf_x_node_scale_exported(SEXP x_SEXP, SEXP w_SEXP) {
-BEGIN_RCPP
-    Rcpp::RObject rcpp_result_gen;
-    Rcpp::RNGScope rcpp_rngScope_gen;
-    Rcpp::traits::input_parameter< NumericMatrix& >::type x_(x_SEXP);
-    Rcpp::traits::input_parameter< NumericVector& >::type w_(w_SEXP);
-    rcpp_result_gen = Rcpp::wrap(x_node_scale_exported(x_, w_));
-    return rcpp_result_gen;
-END_RCPP
-}
-// leaf_kaplan_testthat
-arma::mat leaf_kaplan_testthat(const arma::mat& y, const arma::vec& w);
-RcppExport SEXP _aorsf_leaf_kaplan_testthat(SEXP ySEXP, SEXP wSEXP) {
-BEGIN_RCPP
-    Rcpp::RObject rcpp_result_gen;
-    Rcpp::RNGScope rcpp_rngScope_gen;
-    Rcpp::traits::input_parameter< const arma::mat& >::type y(ySEXP);
-    Rcpp::traits::input_parameter< const arma::vec& >::type w(wSEXP);
-    rcpp_result_gen = Rcpp::wrap(leaf_kaplan_testthat(y, w));
-    return rcpp_result_gen;
-END_RCPP
-}
-// newtraph_cph_testthat
-arma::vec newtraph_cph_testthat(NumericMatrix& x_in, NumericMatrix& y_in, NumericVector& w_in, int method, double cph_eps_, int iter_max);
-RcppExport SEXP _aorsf_newtraph_cph_testthat(SEXP x_inSEXP, SEXP y_inSEXP, SEXP w_inSEXP, SEXP methodSEXP, SEXP cph_eps_SEXP, SEXP iter_maxSEXP) {
-BEGIN_RCPP
-    Rcpp::RObject rcpp_result_gen;
-    Rcpp::RNGScope rcpp_rngScope_gen;
-    Rcpp::traits::input_parameter< NumericMatrix& >::type x_in(x_inSEXP);
-    Rcpp::traits::input_parameter< NumericMatrix& >::type y_in(y_inSEXP);
-    Rcpp::traits::input_parameter< NumericVector& >::type w_in(w_inSEXP);
-    Rcpp::traits::input_parameter< int >::type method(methodSEXP);
-    Rcpp::traits::input_parameter< double >::type cph_eps_(cph_eps_SEXP);
[…]
