### 2022-09-18 ca9e7df
bootstrap and surv classes
