+++ src/ForestSurvival.cpp
@@ -1,0 +1,16 @@
+/*-----------------------------------------------------------------------------
+ This file is part of aorsf.
+ Author: Byron C Jaeger
+ aorsf may be modified and distributed under the terms of the MIT license.
+#----------------------------------------------------------------------------*/
+
+#include <armadillo>
+
+#include "globals.h"
+#include "ForestSurvival.h"
+#include "Data.h"
+
+ namespace aorsf {
+
+
+ } // namespace aorsf
