+++ src/globals.h
@@ -62,8 +62,11 @@
  const int DEFAULT_LEAF_MIN_OBS_REGRESSION = 5;
  const int DEFAULT_LEAF_MIN_OBS_SURVIVAL = 10;
  const int DEFAULT_LEAF_MIN_OBS_PROBABILITY = 10;
+
  const int DEFAULT_MAX_RETRY = 3;
+
  const double DEFAULT_SPLIT_MIN_STAT = 3.84;
+ const int DEFAULT_SPLIT_MIN_OBS = 10;
 
  const SplitRule DEFAULT_SPLITRULE_SURVIVAL = SPLIT_LOGRANK;
 
