+++ R/RcppExports.R
@@ -1,7 +1,11 @@
 # Generated by using Rcpp::compileAttributes() -> do not edit by hand
 # Generator token: 10BE3573-1514-4C36-9D1C-5A225CD40393
 
-orsf_cpp <- function(x, y_ctns, y_intg, weights) {
-    invisible(.Call(`_aorsf_orsf_cpp`, x, y_ctns, y_intg, weights))
+bootstrap_sample_testthat <- function(x, y_ctns, y_intg, weights) {
+    .Call(`_aorsf_bootstrap_sample_testthat`, x, y_ctns, y_intg, weights)
+}
+
+orsf_cpp <- function(x, y_ctns, y_intg, weights, vi = 0L, sr = 1L, pt = 1L) {
+    invisible(.Call(`_aorsf_orsf_cpp`, x, y_ctns, y_intg, weights, vi, sr, pt))
 }
 
