### 2022-09-18 #532 @bcjaeger
@jemus42, here is the more complete response to your comments, with planned action items. If you feel like any of the proposed actions could be improved, or if I didn't address all of your comments in my response, just let me know. Otherwise, I should have updates in `aorsf` based on this response completed in the next couple of days.

> Overall the package features comprehensive documentation and examples. Function names are consistent and discoverable as well as tab-completable.

Thank you! I will continue to prioritize clear documentation and tab-friendly names in future updates.

> I would like to suggest the [checkmate](https://cran.r-project.org/package=checkmate) package for argument checks, which could greatly reduce complexity of internal functions/checks. I do concede that now that the work is already done it's likely not worth re-implementing checks again. Also, the author [wants to keep dependencies to a minimum](https://github.com/bcjaeger/aorsf/blob/2251066cfbf920984072d0e6924d32a7dfc8ff7a/R/srr-stats-standards.R#L111), so I understand the manual implementation of argument checks.

I agree regarding `checkmate`, it would have made the checks more routine (I found out `checkmate` existed shortly after I had already written a substantial amount of code in `check.R`). For now, I'd like to keep the checks as they are, with the acknowledgment that if the checks fail or need a substantial update in the future I will incorporate `checkmate`.

> The check_* functions mentioned above and internal roxy_* functions for documentation snippets both reduce code duplication and ensure documentation is easier to keep consistent and up to date. Some of the roxy_* documentation helpers may be exchangeable for roxygen2 @importFrom tags or templates etc., but that's merely a gut feeling about idiomatic use of roxygen2.

There probably is something I could do here but I also am not sure how to use `roxygen2` optimally for the things that `roxy_` functions currently do. I have learned through working with the `mlr3extralearners` package how to manage bibliographies better, so I think I could remove the `roxy_cite` code if I went that route. I am not sure if this is a high priority update though, since the `roxy_cite` functions appear to be working fine.

> I have not found any instances of @srr tags/comments I would challenge.

Excellent! Thank you for checking these.

> I could not find any reference to parallel processing, i.e. using multiple CPU cores/threads to speed up training. That's not a direct issue I think, but maybe something to look into in the future? Especially given the application on large datasets.

Good point - `aorsf` doesn't currently support parallel processing. I have initiated a branch (see https://github.com/bcjaeger/aorsf/tree/oop) where I am attempting to re-write `aorsf`'s C++ in an object-oriented framework (thanks to Marvin's suggestion above), and I see parallel processing becoming far more feasible with this framework. I've designated parallel processing as a long-term objective tied to this branch (i.e., I don't think it will be done while this review is open, but it will be done).

> The test suite seems comprehensive and checks for correctness against well-established methods, but I'm wondering if it would be useful to add an automated regression test against obliqueRSF (at another location, to avoid the Suggests:-dependency on it).

I like this idea. Would the test be something like graphing the predictions from `aorsf` in the x-axis against the predictions from `obliqueRSF` on the y-axis and expecting their correlation to be high? I'd propose to add a section to the `aorsf::orsf()` examples called "What about obliqueRSF?" and show there that `aorsf` is faster and provides very similar predictions using the `pbc_orsf` data. This shouldn't require adding `obliqueRSF` to `Suggests:` thanks to `roxygen2`'s `.Rmd` options for examples.

> The author also kindly contributed to https://github.com/mlr-org/mlr3extralearners/pull/233, which simplifies the algorithm's inclusion in other benchmark experiments. The initiative to connect to existing frameworks like mlr3 and tidymodels is very much appreciated and allowed me to run a quick (naive) benchmark against ranger and rfrsc, showing comparable performance regarding C-index and brier score, as well as significantly shorter run-time for low-dimensional tasks (caveat: no tuning and default parameters were used for all algorithms, it was merely a quick test-run of the mlr3 learner).

I'm glad to hear the `mlr3` learner ran as expected in the experiment! Would you be willing to share your code? I would like to add a section into the examples of `aorsf::orsf()` that demonstrates a simple use case of `aorsf ` in `mlr3`. A pull request would of course be welcome, but that seems like a bigger ask.

> A very minor nitpicky comment regarding code efficiency:
[fctr.R#17](https://github.com/bcjaeger/aorsf/blob/2251066cfbf920984072d0e6924d32a7dfc8ff7a/R/fctr.R#L17) creates an empty vector (c()) and grows it with each character variable. For small datasets this is likely irrelevant, but imagining a scenario with many character variables supplied intentionally or accidentally by the user, it may be more appropriate to use a vapply construct.

Thank you! I found this very helpful and have implemented your suggestion.

> Partial dependence plots: Due to their [known limitations/assumptions](https://christophm.github.io/interpretable-ml-book/pdp.html#disadvantages-5), it may be appropriate to
a) note those limitations in the documentation to ensure users unfamiliar with the technique can make an informed choice
b) consider additionally supporting accumulated local effects, a ["faster and unbiased alternative" to PDPs](https://christophm.github.io/interpretable-ml-book/ale.html), if possible.

This is a good point. I will add notes in the documentation about the limitations of PD and will recommend users engage with the `iml` package to create ALE curves using `aorsf`. I was happy to write PD functions for `aorsf` because PD is relatively straightforward to compute, but I think ALE may be a little more complex so I'd prefer to lean on the `iml` package for ALE curves.

> Permutation feature importance: This method has drawbacks as well, see [Hooker et. al.](https://arxiv.org/abs/1905.03151v2), so the previous comment applies with the exception that I cannot suggest a direct alternative.

I agree - I will add text to my documentation highlighting this and refer to the paper you linked as well as to the [disadvantages of PD section in the iml book](https://christophm.github.io/interpretable-ml-book/pdp.html#disadvantages-5)

> I was not familiar with the ANOVA- and negation-based importance measures, and I suspect many users may feel the same way. In that regard I think it would be useful to elaborate on these methods slightly, e.g. in vignettes/aorsf.Rmd where they are mentioned next to the aforementioned methods, or in the help page at ?orsf_vi in the Details section which already outlines the methods. Reading the [paper/preprint in aorsf-bench](https://github.com/bcjaeger/aorsf-bench/blob/main/paper/arxiv/main.pdf) was helpful, but I think some rough guidance regarding pros/cons in the documentation would be beneficial for users, if one can even be provided. Choosing interpretability methods is of course not trivial, and I don't expect the author to provide a conclusive heuristic here.

That is a good point. ANOVA importance was developed and evaluated by [Menze et al](https://link.springer.com/chapter/10.1007/978-3-642-23783-6_29) for oblique classification random forests. Negation importance is developed and evaluated in the ArXiv paper you mentioned. According to the simulation study in the ArXiv paper, all three methods (negation, ANOVA, and permutation) appear to be valid. It seems as if negation has an edge in the general comparison based on our simulation study, but that isn't quite enough evidence for me to recommend negation as a default approach. I will add some notes to the documentation about the origin of these three methods and will add pros/cons for each one.

> Overall my impression of the package is very positive!

I am thrilled to hear that. =] Thank you for reviewing closely and providing so many useful ideas to improve `aorsf`!
