### 2022-09-19 9d48cc7
informal memory trace check
