### 2022-09-19 #532 @jemus42
- `roxy_*`-functions: Sure, they serve their purpose and I don't see any reason to forcibly substitute all of them :) Might as well be consistent.
- Parallel processing: Neat! Always nice to have some future enhancements already in the works.
- Regression test: Yes, a scatterplot against `obliqueRSF` predictions would be a good way to check consistency. Wouldn't `obliqueRSF` then still be a build-time dependency for the docs though? I haven't had a chance to experiment with `roxygen2`'s .Rmd option you're referring to, but I mean it has to be installed at some point to render it? Wouldn't CRAN require it as a dependency then?
- Interpretability methods: All good, some documentation notes and leaning on existing implementations is fine :)

Regarding the `mlr3extralearners` test-run I did, my code is fairly bare-bones, and unfortunately due to `mlr3proba` being in a somewhat ill-defined state after its CRAN removal I'm not sure how to communicate that to users. *Technically* it works fine, just with an asterisk or two.

``` r
library(mlr3verse)
#> Loading required package: mlr3
library(mlr3proba)
library(mlr3extralearners)
library(mlr3viz)
library(aorsf)
library(ggplot2)

# Less logging during training
lgr::get_logger("mlr3")$set_threshold("warn")

# Some example tasks without missing values
tasks <- tsks(c("actg", "rats"))

# Learners with default parameters
learners <- lrns(
  c("surv.ranger", "surv.rfsrc", "surv.aorsf"),
  predict_sets = c("train", "test")
)

# Brier (Graf) score, c-index and training time as measures
measures <- msrs(c("surv.graf", "surv.cindex", "time_train"))

# Benchmark with 10-fold CV
design <- benchmark_grid(
  tasks = tasks,
  learners = learners,
  resamplings = rsmps("cv", folds = 10)
)

# Parallelization (optional)
future::plan("multisession", workers = 5)
benchmark_result <- benchmark(design)

bm_scores <- benchmark_result$score(measures, predict_sets = "test")

bm_scores |>
  dplyr::select(task_id, learner_id, surv.graf, surv.cindex, time_train) |>
  tidyr::pivot_longer(cols = surv.graf:time_train, names_to = "measure") |>
  ggplot(aes(x = learner_id, y = value)) +
  facet_wrap(vars(measure), nrow = 1, scales = "free_y") +
  geom_violin(draw_quantiles = c(.25, .5, .75)) +
  theme_minimal()
```

![](https://i.imgur.com/nTrX5yg.png)

``` r

bm_summary <- benchmark_result$aggregate(measures)
bm_summary[, c("task_id", "learner_id", "surv.graf", "surv.cindex")]
#>    task_id  learner_id  surv.graf surv.cindex
#> 1:    actg surv.ranger 0.05933293   0.7265903
#> 2:    actg  surv.rfsrc 0.05776890   0.7373370
#> 3:    actg  surv.aorsf 0.05803357   0.7282324
#> 4:    rats surv.ranger 0.07233919   0.7992290
#> 5:    rats  surv.rfsrc 0.08087658   0.7636673
#> 6:    rats  surv.aorsf 0.08338740   0.7874255
```
