### 2022-09-20 e845af1
arma objects causing trouble unless in header
