+++ src/Tree.h
@@ -7,48 +7,34 @@
 #ifndef TREE_H_
 #define TREE_H_
 
-#include "Data.h"
-#include "globals.h"
-
+#include <armadillo>
+#include <RcppArmadillo.h>
 // [[Rcpp::depends(RcppArmadillo)]]
+
+#include "Data.h"
+#include <Rcpp.h>
+#include "globals.h"
 
  namespace aorsf {
 
  class Tree {
+
  public:
 
-  // Default constructor with no arguments
-  Tree();
+  Tree() = default;
 
-  // Create from loaded forest
-  // Tree(arma::mat& coef,
-  //      arma::umat& coef_indices,
-  //      arma::vec& cutpoint,
-  //      arma::uvec& next_left_node,
-  //      arma::mat& pred,
-  //      arma::umat& pred_indices); // TODO: pred_indices to survival tree
-
-  // Expect to redefine constructors for survival/classif/regression
-  virtual ~Tree() = default;
-
-  // Don't allow unitialized trees
-  Tree(const Tree&) = delete;
-  Tree& operator=(const Tree&) = delete;
-
-  void init(
-    const Data* data,
-    const int mtry,
-    const int max_retry,
-    SplitRule split_rule,
-    const int n_split,
-    const int leaf_min_obs,
-    const int split_min_obs,
-    const int split_min_stat,
-    PredType pred_type,
-    const int oobag_eval_every,
-    VariableImportance variable_importance,
-    const int seed
-  ) {
+  Tree(const Data* data,
+       int mtry,
+       int max_retry,
+       SplitRule split_rule,
+       int n_split,
+       int leaf_min_obs,
+       int split_min_obs,
+       int split_min_stat,
+       PredType pred_type,
+       int oobag_eval_every,
+       VariableImportance variable_importance,
+       int seed){
 
    this->data = data;
    this->mtry = mtry;
@@ -62,7 +48,72 @@
    this->oobag_eval_every = oobag_eval_every;
    this->variable_importance = variable_importance;
 
-  };
+   int a = 2;
+   int b = 4;
+
+   this->coef = Rcpp::NumericMatrix(a,b);
+   this->coef_indices = arma::umat(a, b);
+
+  }
+
+  // Pointer to original data
+  const Data* data;
+
+  // number of predictors used to split a node
+  int mtry;
+
+  // maximum number of retry attempts to split a node
+  int max_retry;
+
+  // how to measure quality of a node split
+  SplitRule split_rule;
+
+  // number of cutpoints to assess during node split
+  int n_split;
+
+  // minimum number of observations needed in a leaf node
+  int leaf_min_obs;
+
+  // minimum number of observations needed to split a node
+  int split_min_obs;
+
+  // minimum value of split statistic needed to split a node
+  int split_min_stat;
+
+  // what type of oobag prediction to compute
+  PredType pred_type;
+
+  // evaluate oobag error every X trees
+  int oobag_eval_every;
+
+  // what type of variable importance to compute
+  VariableImportance variable_importance;
+
+  // random seed to be set before growing
+  int seed;
+
+  // coefficients for linear combinations;
+  // one row per variable (mtry rows), one column per node
+  // leaf nodes have all coefficients=0
+  Rcpp::NumericMatrix coef;
+
+  // indices of the predictors used by
+  arma::umat coef_indices;
+
+  // cutpoints used to split the node
+  arma::vec cutpoint;
+
+  // directions to the next node (right node = left node + 1)
+  arma::uvec left_node;
+
+  // predicted values (only in leaf nodes)
+  arma::mat pred;
+
+  // indices of predicted values for each leaf node
+  // TODO move to survivaltree
+  arma::umat pred_indices;
+
+  void guess_max_nodes();
 
   int get_mtry() const {
[…]
