+++ src/RcppExports.cpp
@@ -26,7 +26,7 @@ BEGIN_RCPP
 END_RCPP
 }
 // orsf_cpp
-void orsf_cpp(arma::mat& x, arma::vec& y_ctns, arma::ivec& y_intg, arma::vec& weights, const int vi, const int sr, const int pt);
+void orsf_cpp(arma::mat& x, arma::vec& y_ctns, arma::ivec& y_intg, arma::vec& weights, int vi, int sr, int pt);
 RcppExport SEXP _aorsf_orsf_cpp(SEXP xSEXP, SEXP y_ctnsSEXP, SEXP y_intgSEXP, SEXP weightsSEXP, SEXP viSEXP, SEXP srSEXP, SEXP ptSEXP) {
 BEGIN_RCPP
     Rcpp::RNGScope rcpp_rngScope_gen;
@@ -34,9 +34,9 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< arma::vec& >::type y_ctns(y_ctnsSEXP);
     Rcpp::traits::input_parameter< arma::ivec& >::type y_intg(y_intgSEXP);
     Rcpp::traits::input_parameter< arma::vec& >::type weights(weightsSEXP);
-    Rcpp::traits::input_parameter< const int >::type vi(viSEXP);
-    Rcpp::traits::input_parameter< const int >::type sr(srSEXP);
-    Rcpp::traits::input_parameter< const int >::type pt(ptSEXP);
+    Rcpp::traits::input_parameter< int >::type vi(viSEXP);
+    Rcpp::traits::input_parameter< int >::type sr(srSEXP);
+    Rcpp::traits::input_parameter< int >::type pt(ptSEXP);
     orsf_cpp(x, y_ctns, y_intg, weights, vi, sr, pt);
     return R_NilValue;
 END_RCPP
