+++ man/orsf_vi.Rd
@@ -133,14 +133,12 @@ the ‘raw’ variable importance values are stored in the fit object
 \if{html}{\out{<div class="sourceCode r">}}\preformatted{fit$importance
 }\if{html}{\out{</div>}}
 
-\if{html}{\out{<div class="sourceCode">}}\preformatted{##     edema_1   ascites_1        bili      copper         age 
-##  0.39603233  0.35528942  0.27837977  0.19605331  0.18822292 
-##     albumin     protime        chol   edema_0.5       stage 
-##  0.17033964  0.15320911  0.14883599  0.14599194  0.13627743 
-##   spiders_1         ast    hepato_1       sex_f        trig 
-##  0.13495783  0.12754159  0.12057626  0.10669014  0.09723320 
-##    alk.phos    platelet trt_placebo 
-##  0.09183673  0.07599581  0.06846999
+\if{html}{\out{<div class="sourceCode">}}\preformatted{##     edema_1   ascites_1        bili      copper         age     albumin 
+##  0.39603233  0.35528942  0.27837977  0.19605331  0.18822292  0.17033964 
+##     protime        chol   edema_0.5       stage   spiders_1         ast 
+##  0.15320911  0.14883599  0.14599194  0.13627743  0.13495783  0.12754159 
+##    hepato_1       sex_f        trig    alk.phos    platelet trt_placebo 
+##  0.12057626  0.10669014  0.09723320  0.09183673  0.07599581  0.06846999
 }\if{html}{\out{</div>}}
 
 these are ‘raw’ because values for factors have not been aggregated into
@@ -158,12 +156,12 @@ To get aggregated values across all levels of each factor, use
 \if{html}{\out{<div class="sourceCode r">}}\preformatted{orsf_vi(fit)
 }\if{html}{\out{</div>}}
 
-\if{html}{\out{<div class="sourceCode">}}\preformatted{##    ascites       bili      edema     copper        age    albumin 
-## 0.35528942 0.27837977 0.24719876 0.19605331 0.18822292 0.17033964 
-##    protime       chol      stage    spiders        ast     hepato 
-## 0.15320911 0.14883599 0.13627743 0.13495783 0.12754159 0.12057626 
-##        sex       trig   alk.phos   platelet        trt 
-## 0.10669014 0.09723320 0.09183673 0.07599581 0.06846999
+\if{html}{\out{<div class="sourceCode">}}\preformatted{##    ascites       bili      edema     copper        age    albumin    protime 
+## 0.35528942 0.27837977 0.24719876 0.19605331 0.18822292 0.17033964 0.15320911 
+##       chol      stage    spiders        ast     hepato        sex       trig 
+## 0.14883599 0.13627743 0.13495783 0.12754159 0.12057626 0.10669014 0.09723320 
+##   alk.phos   platelet        trt 
+## 0.09183673 0.07599581 0.06846999
 }\if{html}{\out{</div>}}
 }
 
@@ -181,29 +179,27 @@ You can fit an ORSF without VI, then add VI later
 orsf_vi_negate(fit_no_vi)
 }\if{html}{\out{</div>}}
 
-\if{html}{\out{<div class="sourceCode">}}\preformatted{##          bili           age        copper       protime 
-##  0.0164096687  0.0119295687  0.0105230256  0.0095853303 
-##       ascites         edema           sex       spiders 
-##  0.0053136070  0.0036428720  0.0029693686  0.0027609919 
-##      alk.phos        hepato         stage       albumin 
-##  0.0025005209  0.0022400500  0.0018753907  0.0009897895 
-##           ast          trig           trt      platelet 
-##  0.0005730360 -0.0013544488 -0.0017191081 -0.0031256512 
-##          chol 
-## -0.0040633465
[…]

[…]
