### 2022-09-21 0701f17
informal test comparing to obliqueRSF
