+++ man/orsf.Rd
@@ -267,560 +267,6 @@ A user should impute missing values using an R package with that purpose,
 such as \code{recipes} or \code{mlr3pipelines}.
 }
 
-\section{Examples}{
-First we load some relevant packages
-
-\if{html}{\out{<div class="sourceCode r">}}\preformatted{set.seed(329730)
-suppressPackageStartupMessages(\{
- library(aorsf)
- library(survival)
- library(tidymodels)
- library(tidyverse)
- library(randomForestSRC)
- library(ranger)
- library(riskRegression) 
- library(obliqueRSF)
-\})
-}\if{html}{\out{</div>}}
-
-The entry-point into \code{aorsf} is the standard call to \code{orsf()}:
-
-\if{html}{\out{<div class="sourceCode r">}}\preformatted{fit <- orsf(pbc_orsf, Surv(time, status) ~ . - id)
-}\if{html}{\out{</div>}}
-
-printing \code{fit} provides quick descriptive summaries:
-
-\if{html}{\out{<div class="sourceCode r">}}\preformatted{fit
-}\if{html}{\out{</div>}}
-
-\if{html}{\out{<div class="sourceCode">}}\preformatted{## ---------- Oblique random survival forest
-## 
-##      Linear combinations: Accelerated
-##           N observations: 276
-##                 N events: 111
-##                  N trees: 500
-##       N predictors total: 17
-##    N predictors per node: 5
-##  Average leaves per tree: 24
-## Min observations in leaf: 5
-##       Min events in leaf: 1
-##           OOB stat value: 0.84
-##            OOB stat type: Harrell's C-statistic
-##      Variable importance: anova
-## 
-## -----------------------------------------
-}\if{html}{\out{</div>}}
-\subsection{Model control}{
-
-For these examples we will make use of the \code{orsf_control_} functions to
-build and compare models based on their out-of-bag predictions. We will
-also standardize the out-of-bag samples using the input argument
-\code{tree_seeds}
-\subsection{Accelerated linear combinations}{
-
-The accelerated ORSF ensemble is the default because it has a nice
-balance of computational speed and prediction accuracy. It runs a single
-iteration of Newton Raphson scoring on the Cox partial likelihood
-function to find linear combinations of predictors.
-
-\if{html}{\out{<div class="sourceCode r">}}\preformatted{fit_accel <- orsf(pbc_orsf, 
-                  control = orsf_control_fast(),
-                  formula = Surv(time, status) ~ . - id,
-                  tree_seeds = 1:500)
-}\if{html}{\out{</div>}}
-}
-
-\subsection{Linear combinations with Cox regression}{
-
-\code{orsf_control_cph} runs Cox regression in each non-terminal node of each
-survival tree, using the regression coefficients to create linear
-combinations of predictors:
-
-\if{html}{\out{<div class="sourceCode r">}}\preformatted{fit_cph <- orsf(pbc_orsf, 
-                control = orsf_control_cph(),
-                formula = Surv(time, status) ~ . - id,
-                tree_seeds = 1:500)
-}\if{html}{\out{</div>}}
-}
-
-\subsection{Linear combinations with penalized cox regression}{
-
[…]
