### 2022-09-21 #532 @bcjaeger
> Regression test: Yes, a scatterplot against obliqueRSF predictions would be a good way to check consistency. Wouldn't obliqueRSF then still be a build-time dependency for the docs though? I haven't had a chance to experiment with roxygen2's .Rmd option you're referring to, but I mean it has to be installed at some point to render it? Wouldn't CRAN require it as a dependency then?

The `@includeRmd` option in `roxygen2` seems to be a valid backdoor to include packages in your examples without making them dependencies or suggested packages. I think this is because the `.Rmd` files are knitted prior to sending the package out, so the code is only run locally.

I am happy to share some updates!

- I've added the `mlr3` example to my examples in `orsf.R` (see [here](https://bcjaeger.github.io/aorsf/reference/orsf.html#mlr-pipelines)).

- I've added documentation on limitations of partial dependence (see [here](https://bcjaeger.github.io/aorsf/reference/orsf_pd_oob.html#details) and [here](https://bcjaeger.github.io/aorsf/articles/pd.html#limitations-of-pd)).

- I have placed the regression test for `obliqueRSF` into a testing file, and plan to run the test locally to prevent adding `obliqueRSF` as a suggested package. I am a little uncertain regarding the minimal correlation that would be acceptable, but I feel comfortable with 0.97 (i.e., the current correlation from the reprex below).

``` r
# Similar to obliqueRSF?
suppressPackageStartupMessages({
 library(aorsf)
 library(obliqueRSF)
})

set.seed(50)

fit_aorsf <- orsf(pbc_orsf,
                  formula = Surv(time, status) ~ . - id,
                  n_tree = 100)
fit_obliqueRSF <- ORSF(pbc_orsf, ntree = 100, verbose = FALSE)

risk_aorsf <- predict(fit_aorsf, new_data = pbc_orsf, pred_horizon = 3500)
risk_obliqueRSF <- 1-predict(fit_obliqueRSF, newdata = pbc_orsf, times = 3500)

cor(risk_obliqueRSF, risk_aorsf)
#>           [,1]
#> [1,] 0.9747177
plot(risk_obliqueRSF, risk_aorsf)
```

![](https://i.imgur.com/QCTe5F9.png)

@jemus42, do you feel these updates (in addition to the earlier update using `vapply` instead of `c()`) have successfully incorporated the feedback from your review into `aorsf`?
