### 2022-09-21 cbd49de
discuss pros and cons of each vi technique
