### 2022-09-21 #532 @jemus42
Yes, I have nothing more to add I think :)

Regarding the regression test, I don't think there's a meaningful cutoff one could define, but if for example you were to introduce a bug on the `aorsf` C++ code that resulted in nonsense predictions you would most likely immediately recognize them in that plot, and that's the main appeal :)
