### 2022-09-22 #532 @ropensci-review-bot
Approved! Thanks @bcjaeger for submitting and @chjackson, @mnwright, @jemus42 for your reviews! :grin:
